Plot options for plotEvents

This function provides the plotting options for the plotEvents() function.

plotEvents.control(
  xlab = NULL,
  ylab = NULL,
  main = NULL,
  terminal.name = NULL,
  recurrent.name = NULL,
  recurrent.type = NULL,
  legend.position = NULL,
  base_size = 12,
  cex = NULL,
  alpha = 0.7,
  width = NULL,
  bar.color = NULL,
  recurrent.color = NULL,
  recurrent.shape = NULL,
  recurrent.stroke = NULL,
  terminal.color = NULL,
  terminal.shape = NULL,
  terminal.stroke = NULL,
  not.terminal.color = NULL,
  not.terminal.shape = NULL
)

Arguments

xlab

a character string indicating the label for the x axis. The default value is "Time".

ylab

a character string indicating the label for the y axis. The default value is "Subject".

main

a character string indicating the title of the plot.

terminal.name

a character string indicating the label for the terminal event displayed in the legend. The default value is "Terminal event".

recurrent.name

a character string indicating the label for the recurrent event displayed in the legend. The default value is "Recurrent events".

recurrent.type

a factor indicating the labels for the different recurrent event types. This option is only available when there are more than one types of recurrent events. The default value is "Recurrent events 1", "Recurrent events 2", ....

legend.position

a character string specifies the position of the legend. The available options are "none", "left", "right", "bottom", "top", "bottomright", "bottomleft","topleft", "topright", or a two-element numeric vector specifies the coordinate of the legend. The legend is placed inside of the plotting region when legend.position is specified as one of "bottomright", "bottomleft","topleft", "topright". Otherwise, the argument is passed to the ggplot theme environment.

base_size

a numerical value to specify the base font size, given in pts. This argument is passed to the ggplot theme environment. The default value is 12.

cex

a numerical value specifies the size of the points.

alpha

a numerical value specifies the transparency of the points.

width

a numerical value specifies the width of the event plot. By ggplot default, set to 90% of the resolution of the data.

bar.color

a numerical value or a character string specifies color for lines. Default to gray.

recurrent.color

a numerical value or a character string specifies color for recurrent events. Default to green.

recurrent.shape

a numerical value or a character string specifies shape for recurrent events. Default to circle.

recurrent.stroke

a numerical value or a character string specifies stroke for recurrent events. Default to circle.

terminal.color

a numerical value or a character string specifies color for terminal events. Default to red.

terminal.shape

a numerical value or a character string specifies shape for terminal events. Default to triangle.

terminal.stroke

a numerical value or a character string specifies stroke for terminal events. Default to triangle.

not.terminal.color

a numerical value or a character string specifies color for non-terminal events. Non-terminal events are not plotted at default.

not.terminal.shape

a numerical value or a character string specifies shape for terminal events. Non-terminal events are not plotted at default.

See also