Package options for reReg

This function provides the fitting options for the reReg() function.

reReg.control(
  eqType = c("logrank", "gehan", "gehan_s"),
  solver = c("BB::dfsane", "BB::BBsolve", "BB::BBoptim", "optimx::optimr",
    "dfoptim::hjk", "dfoptim::mads", "optim", "nleqslv::nleqslv"),
  tol = 1e-07,
  cppl = NULL,
  cppl.wfun = list(NULL, NULL),
  init = list(alpha = 0, beta = 0, eta = 0, theta = 0),
  boot.parallel = FALSE,
  boot.parCl = NULL,
  maxit1 = 100,
  maxit2 = 10,
  trace = FALSE,
  numAdj = 1e-07
)

Arguments

eqType

a character string indicating whether the log-rank type estimating equation or the Gehan-type estimating equation (when available) will be used.

solver

a character string specifying the equation solver to be used for root search.

tol

a positive numerical value specifying the absolute error tolerance in root search.

cppl

a character string indicating either to improve the proportional rate model via the generalized method of moments (cppl = "GMM") or empirical likelihood estimation (cppl = "EL"). This option is only used when model = "cox.HH".

cppl.wfun

a list of (up to two) weight functions to be combined with the weighted pseudo-partial likelihood scores. Available options are "Gehan" and "cumbase", which correspond to the Gehan's weight and the cumulative baseline hazard function, respectively. Alternatively, the weight functions can be specified with function formulas. This option is only used when model = "cox.HH".

init

a list contains the initial guesses used for root search.

boot.parallel

an logical value indicating whether parallel computation will be applied when se = "boot" is specified in reReg().

boot.parCl

an integer value specifying the number of CPU cores to be used when parallel = TRUE. The default value is half the CPU cores on the current host.

maxit1, maxit2

max number of iteration used when model = "cox.HH".

trace

a logical variable denoting whether some of the intermediate results of iterations should be displayed to the user. Default is FALSE.

numAdj

a positive numerical value specifying the small constant used in heuristic adjustment of the borrow strength method.

See also