Create an reSurv Object

Create a recurrent event survival object, used as a response variable in reReg. This function is being deprecated in Version 1.1.7. A recurrent event object is now being created with Recur(). See '?Recur()' for details.

reSurv(time1, time2, id, event, status, origin = 0)

Arguments

time1
when "time2" is provided, this vector is treated as the starting time for the gap time between two successive recurrent events. In the absence of "time2", this is the observation time of recurrence on calendar time scale, in which, the time corresponds to the time since entry/inclusion in the study.

time2
an optional vector for ending time for the gap time between two successive recurrent events.

id
subject's id.

event
a binary vector used as the recurrent event indicator. event = 1 for recurrent times.

status
a binary vector used as the status indicator for the terminal event. status = 0 for censored times.

origin
a numerical vector indicating the time origin of subjects. When origin is a scalar, reSurv assumes all subjects have the same origin. Otherwise, origin needs to be a numerical vector, with length equals to the number of subjects. In this case, each element corresponds to different origins for different subjects. This argument is only needed when "time2" is missing.

Examples

set.seed(1)
dat <- simSC(200, c(-1, 1), c(-1, 1))

## being deprecated in Verson 1.1.7
with(dat, reSurv(Time, id, event, status))

#> Warning: 'reSurv()' is being deprecated in Version 1.1.7. Output is prepared by 'Recur()'.
#>  See '?Recur()' for details.

#>   [1] 1: (0.0000, 4.1939+]                                               
#>   [2] 2: (0.0000, 0.4220], (0.4220, 0.7060], ..., (1.2931, 1.7378*]      
#>   [3] 3: (0.0000, 27.8297], (27.8297, 49.3334+]                          
#>   [4] 4: (0.0000, 3.5663+]                                               
#>   [5] 5: (0.0000, 0.0276], (0.0276, 0.3024], ..., (35.5598, 59.8088+]    
#>   [6] 6: (0.0000, 0.9859], (0.9859, 2.3660], ..., (3.4217, 8.1096+]      
#>   [7] 7: (0.0000, 1.3793], (1.3793, 10.4272], ..., (25.3046, 45.9819+]   
#>   [8] 8: (0.0000, 0.0564], (0.0564, 0.2060], ..., (1.1791, 1.1937*]      
#>   [9] 9: (0.0000, 3.1155], (3.1155, 3.7286], ..., (21.1107, 53.5430+]    
#>  [10] 10: (0.0000, 0.0224], (0.0224, 1.3998], ..., (14.3499, 30.2477+]   
#>  [11] 11: (0.0000, 0.0755], (0.0755, 0.2167], ..., (4.0920, 7.8248+]     
#>  [12] 12: (0.0000, 5.2976], (5.2976, 60.0000+]                           
#>  [13] 13: (0.0000, 0.7602], (0.7602, 0.9261], ..., (12.3359, 60.0000+]   
#>  [14] 14: (0.0000, 1.3152+]                                              
#>  [15] 15: (0.0000, 0.0469], (0.0469, 0.3025], ..., (21.9283, 26.7154*]   
#>  [16] 16: (0.0000, 60.0000+]                                             
#>  [17] 17: (0.0000, 32.4884+]                                             
#>  [18] 18: (0.0000, 60.0000+]                                             
#>  [19] 19: (0.0000, 21.7731+]                                             
#>  [20] 20: (0.0000, 2.2005], (2.2005, 28.5968+]                           
#>  [21] 21: (0.0000, 19.8022], (19.8022, 40.9622+]                         
#>  [22] 22: (0.0000, 0.9727], (0.9727, 2.1431], ..., (28.4250, 60.0000+]   
#>  [23] 23: (0.0000, 2.7144], (2.7144, 7.3384+]                            
#>  [24] 24: (0.0000, 0.0762], (0.0762, 0.1865], ..., (9.6314, 13.5243+]    
#>  [25] 25: (0.0000, 0.6611], (0.6611, 0.7965], ..., (16.1909, 24.3334+]   
#>  [26] 26: (0.0000, 0.0385], (0.0385, 0.1608], ..., (6.6432, 6.8203*]     
#>  [27] 27: (0.0000, 0.0382], (0.0382, 0.0985], ..., (18.3202, 60.0000+]   
#>  [28] 28: (0.0000, 0.0091*]                                              
#>  [29] 29: (0.0000, 2.9126+]                                              
#>  [30] 30: (0.0000, 0.9483], (0.9483, 1.1582], (1.1582, 2.7012+]          
#>  [31] 31: (0.0000, 0.1602], (0.1602, 0.8470], ..., (20.7914, 60.0000+]   
#>  [32] 32: (0.0000, 4.0675], (4.0675, 60.0000+]                           
#>  [33] 33: (0.0000, 0.0018], (0.0018, 0.1633], ..., (0.3409, 0.4080*]     
#>  [34] 34: (0.0000, 0.2069], (0.2069, 0.9686], ..., (7.2497, 8.2745+]     
#>  [35] 35: (0.0000, 38.6777], (38.6777, 60.0000+]                         
#>  [36] 36: (0.0000, 0.7862], (0.7862, 0.8419], ..., (49.2388, 60.0000+]   
#>  [37] 37: (0.0000, 0.8320], (0.8320, 6.3206], ..., (53.3568, 60.0000+]   
#>  [38] 38: (0.0000, 0.0254], (0.0254, 0.1117], ..., (3.0250, 3.2085+]     
#>  [39] 39: (0.0000, 1.9214], (1.9214, 13.0673+]                           
#>  [40] 40: (0.0000, 0.5774], (0.5774, 0.7782], (0.7782, 1.0653*]          
#>  [41] 41: (0.0000, 0.3697], (0.3697, 2.3635], ..., (4.3296, 60.0000+]    
#>  [42] 42: (0.0000, 0.0709], (0.0709, 0.0770], ..., (0.6271, 1.0153*]     
#>  [43] 43: (0.0000, 0.3110*]                                              
#>  [44] 44: (0.0000, 2.1150*]                                              
#>  [45] 45: (0.0000, 0.2198], (0.2198, 0.6528], ..., (17.1894, 17.3457+]   
#>  [46] 46: (0.0000, 1.3165], (1.3165, 3.8825], ..., (5.5116, 16.9859+]    
#>  [47] 47: (0.0000, 0.3050], (0.3050, 0.3518], ..., (0.6241, 2.8628*]     
#>  [48] 48: (0.0000, 13.4123], (13.4123, 19.7926+]                         
#>  [49] 49: (0.0000, 0.3423], (0.3423, 0.5334], ..., (43.5570, 60.0000+]   
#>  [50] 50: (0.0000, 4.1856], (4.1856, 5.5388], ..., (29.1396, 60.0000+]   
#>  [51] 51: (0.0000, 0.4839], (0.4839, 0.4968], ..., (0.7886, 0.9674+]     
#>  [52] 52: (0.0000, 1.2673], (1.2673, 4.6764], ..., (23.6977, 31.4293+]   
#>  [53] 53: (0.0000, 0.9122], (0.9122, 1.4916+]                            
#>  [54] 54: (0.0000, 0.0561], (0.0561, 1.3932], ..., (14.5763, 33.4254+]   
#>  [55] 55: (0.0000, 3.4611], (3.4611, 7.8629+]                            
#>  [56] 56: (0.0000, 0.0483], (0.0483, 0.1243], ..., (0.7432, 3.8557*]     
#>  [57] 57: (0.0000, 31.0839+]                                             
#>  [58] 58: (0.0000, 0.3138], (0.3138, 0.7740], ..., (22.9936, 23.5281+]   
#>  [59] 59: (0.0000, 0.2635], (0.2635, 0.6228], ..., (2.0408, 2.7623*]     
#>  [60] 60: (0.0000, 0.5610], (0.5610, 58.8840+]                           
#>  [61] 61: (0.0000, 0.3918*]                                              
#>  [62] 62: (0.0000, 12.0660], (12.0660, 14.9237+]                         
#>  [63] 63: (0.0000, 0.0514], (0.0514, 0.1477], ..., (29.4005, 60.0000+]   
#>  [64] 64: (0.0000, 0.1402], (0.1402, 16.2452], (16.2452, 60.0000+]       
#>  [65] 65: (0.0000, 0.0285], (0.0285, 0.1528], ..., (8.1595, 9.4588*]     
#>  [66] 66: (0.0000, 0.5731*]                                              
#>  [67] 67: (0.0000, 0.9741], (0.9741, 2.0285], ..., (2.2902, 3.4172+]     
#>  [68] 68: (0.0000, 0.7058+]                                              
#>  [69] 69: (0.0000, 2.3642], (2.3642, 2.4615], (2.4615, 60.0000+]         
#>  [70] 70: (0.0000, 0.1671], (0.1671, 3.2458], ..., (3.4636, 5.4292*]     
#>  [71] 71: (0.0000, 0.9312], (0.9312, 8.3824], (8.3824, 60.0000+]         
#>  [72] 72: (0.0000, 0.2243], (0.2243, 0.3831], ..., (3.8633, 4.3904*]     
#>  [73] 73: (0.0000, 19.1460+]                                             
#>  [74] 74: (0.0000, 0.1218], (0.1218, 10.3010], (10.3010, 60.0000+]       
#>  [75] 75: (0.0000, 1.2854], (1.2854, 60.0000+]                           
#>  [76] 76: (0.0000, 0.7864], (0.7864, 60.0000+]                           
#>  [77] 77: (0.0000, 14.7017], (14.7017, 47.1695+]                         
#>  [78] 78: (0.0000, 0.1847], (0.1847, 0.6563], ..., (5.9610, 10.8273+]    
#>  [79] 79: (0.0000, 1.1958], (1.1958, 1.6888], ..., (22.3231, 60.0000+]   
#>  [80] 80: (0.0000, 0.2668], (0.2668, 0.4614], ..., (4.7256, 9.6897*]     
#>  [81] 81: (0.0000, 0.2466], (0.2466, 0.3025], ..., (28.2620, 43.7449+]   
#>  [82] 82: (0.0000, 0.0199], (0.0199, 0.1099], ..., (1.4919, 3.7963+]     
#>  [83] 83: (0.0000, 0.9455+]                                              
#>  [84] 84: (0.0000, 0.4360], (0.4360, 0.4424], ..., (41.9064, 60.0000+]   
#>  [85] 85: (0.0000, 1.4408+]                                              
#>  [86] 86: (0.0000, 0.0022], (0.0022, 0.0164], ..., (11.8213, 12.2415*]   
#>  [87] 87: (0.0000, 0.2622], (0.2622, 2.3970+]                            
#>  [88] 88: (0.0000, 4.0281+]                                              
#>  [89] 89: (0.0000, 0.1192], (0.1192, 0.6737], ..., (19.1438, 40.7039+]   
#>  [90] 90: (0.0000, 2.6533], (2.6533, 7.5542], ..., (16.7730, 60.0000+]   
#>  [91] 91: (0.0000, 0.1813], (0.1813, 3.7734], ..., (38.9600, 39.2281+]   
#>  [92] 92: (0.0000, 13.0989], (13.0989, 20.6036], ..., (45.2517, 60.0000+]
#>  [93] 93: (0.0000, 2.3861], (2.3861, 3.3904], (3.3904, 10.8095+]         
#>  [94] 94: (0.0000, 0.8110], (0.8110, 1.5011], ..., (14.5904, 15.0829*]   
#>  [95] 95: (0.0000, 9.3932], (9.3932, 19.8858+]                           
#>  [96] 96: (0.0000, 6.0415], (6.0415, 16.3964*]                           
#>  [97] 97: (0.0000, 0.1147], (0.1147, 0.4367], ..., (27.4787, 60.0000+]   
#>  [98] 98: (0.0000, 14.8624+]                                             
#>  [99] 99: (0.0000, 16.1137+]                                             
#> [100] 100: (0.0000, 2.0292], (2.0292, 17.5638], ..., (24.7633, 30.4540+] 
#> [101] 101: (0.0000, 0.2228], (0.2228, 0.2259], ..., (0.4402, 0.5615*]    
#> [102] 102: (0.0000, 0.0494], (0.0494, 0.3593], ..., (1.2053, 1.2369*]    
#> [103] 103: (0.0000, 27.1265+]                                            
#> [104] 104: (0.0000, 0.8268], (0.8268, 2.7076], ..., (19.2702, 19.3323+]  
#> [105] 105: (0.0000, 2.5198*]                                             
#> [106] 106: (0.0000, 0.1236], (0.1236, 0.1819], ..., (3.1792, 3.8315*]    
#> [107] 107: (0.0000, 11.2995+]                                            
#> [108] 108: (0.0000, 0.1591], (0.1591, 0.2060], (0.2060, 3.8735+]         
#> [109] 109: (0.0000, 0.0429], (0.0429, 0.0576], ..., (0.1464, 0.1905*]    
#> [110] 110: (0.0000, 0.0641*]                                             
#> [111] 111: (0.0000, 0.1070], (0.1070, 1.0162], ..., (10.1607, 19.4814+]  
#> [112] 112: (0.0000, 15.9173], (15.9173, 17.9019], (17.9019, 60.0000+]    
#> [113] 113: (0.0000, 0.0657], (0.0657, 0.2235], ..., (0.3495, 0.7056*]    
#> [114] 114: (0.0000, 11.8459], (11.8459, 23.2224*]                        
#> [115] 115: (0.0000, 0.1231], (0.1231, 2.6182], ..., (3.4555, 4.4275+]    
#> [116] 116: (0.0000, 0.5347], (0.5347, 1.2448], (1.2448, 1.3201+]         
#> [117] 117: (0.0000, 9.6803], (9.6803, 12.2911], (12.2911, 60.0000+]      
#> [118] 118: (0.0000, 0.9918], (0.9918, 60.0000+]                          
#> [119] 119: (0.0000, 0.4323], (0.4323, 0.9787], ..., (57.0146, 60.0000+]  
#> [120] 120: (0.0000, 0.0667], (0.0667, 0.1311*]                           
#> [121] 121: (0.0000, 7.2813+]                                             
#> [122] 122: (0.0000, 0.0490], (0.0490, 0.2071], ..., (13.7969, 30.5598+]  
#> [123] 123: (0.0000, 60.0000+]                                            
#> [124] 124: (0.0000, 0.3529], (0.3529, 9.7130], (9.7130, 11.7963*]        
#> [125] 125: (0.0000, 0.8298], (0.8298, 2.6456], ..., (3.3778, 4.3753+]    
#> [126] 126: (0.0000, 3.3505], (3.3505, 47.9797*]                          
#> [127] 127: (0.0000, 0.1545], (0.1545, 0.4070], ..., (0.5757, 3.0877+]    
#> [128] 128: (0.0000, 0.0334], (0.0334, 60.0000+]                          
#> [129] 129: (0.0000, 1.4197+]                                             
#> [130] 130: (0.0000, 2.2834], (2.2834, 21.0913+]                          
#> [131] 131: (0.0000, 0.0567+]                                             
#> [132] 132: (0.0000, 0.6541], (0.6541, 2.3745*]                           
#> [133] 133: (0.0000, 0.5389], (0.5389, 0.8951], ..., (6.7334, 22.9528+]   
#> [134] 134: (0.0000, 0.1956], (0.1956, 0.2891], ..., (7.2722, 7.5890*]    
#> [135] 135: (0.0000, 0.1912], (0.1912, 0.1964], (0.1964, 0.2359*]         
#> [136] 136: (0.0000, 0.4751], (0.4751, 1.0699], ..., (24.5560, 38.2816+]  
#> [137] 137: (0.0000, 16.4606+]                                            
#> [138] 138: (0.0000, 0.1685], (0.1685, 2.3897], (2.3897, 44.2901+]        
#> [139] 139: (0.0000, 0.3702], (0.3702, 60.0000+]                          
#> [140] 140: (0.0000, 0.0995], (0.0995, 0.3184], ..., (56.4982, 60.0000+]  
#> [141] 141: (0.0000, 22.2126+]                                            
#> [142] 142: (0.0000, 0.6679], (0.6679, 1.2204], (1.2204, 1.3808*]         
#> [143] 143: (0.0000, 0.0215], (0.0215, 0.0998], ..., (0.3990, 0.6926*]    
#> [144] 144: (0.0000, 0.3498], (0.3498, 0.4453], ..., (3.6977, 4.5617+]    
#> [145] 145: (0.0000, 1.3728], (1.3728, 2.7651], ..., (16.4156, 18.2321+]  
#> [146] 146: (0.0000, 0.2692], (0.2692, 0.8881], ..., (4.4387, 9.9300*]    
#> [147] 147: (0.0000, 7.4812+]                                             
#> [148] 148: (0.0000, 0.9865], (0.9865, 2.2654], ..., (7.0282, 7.9999+]    
#> [149] 149: (0.0000, 0.3762], (0.3762, 0.4328], ..., (1.8875, 2.4293*]    
#> [150] 150: (0.0000, 0.0910], (0.0910, 0.4390], ..., (2.1062, 25.1158+]   
#> [151] 151: (0.0000, 3.4940], (3.4940, 25.3259], (25.3259, 45.1222+]      
#> [152] 152: (0.0000, 1.7864], (1.7864, 7.2195], (7.2195, 9.5792+]         
#> [153] 153: (0.0000, 0.7874], (0.7874, 2.6515*]                           
#> [154] 154: (0.0000, 5.5211], (5.5211, 41.6133+]                          
#> [155] 155: (0.0000, 0.1137], (0.1137, 9.6278], (9.6278, 20.6738*]        
#> [156] 156: (0.0000, 2.7274], (2.7274, 4.3713*]                           
#> [157] 157: (0.0000, 0.1500], (0.1500, 0.1602], ..., (13.6229, 17.2046+]  
#> [158] 158: (0.0000, 0.9707], (0.9707, 2.6660*]                           
#> [159] 159: (0.0000, 12.8927], (12.8927, 23.6012+]                        
#> [160] 160: (0.0000, 0.0813], (0.0813, 0.1225], ..., (11.5289, 15.4727+]  
#> [161] 161: (0.0000, 0.6257], (0.6257, 0.7828], ..., (37.4610, 60.0000+]  
#> [162] 162: (0.0000, 0.2446], (0.2446, 17.6765+]                          
#> [163] 163: (0.0000, 1.5187], (1.5187, 3.9270], ..., (17.8582, 22.7722*]  
#> [164] 164: (0.0000, 0.0294], (0.0294, 0.0855], ..., (1.1457, 1.2137*]    
#> [165] 165: (0.0000, 0.0554], (0.0554, 0.1506], ..., (2.9271, 3.1379*]    
#> [166] 166: (0.0000, 0.6354+]                                             
#> [167] 167: (0.0000, 2.2898*]                                             
#> [168] 168: (0.0000, 0.6285], (0.6285, 0.8874], ..., (10.8140, 25.8737*]  
#> [169] 169: (0.0000, 0.6257], (0.6257, 7.3773], (7.3773, 31.4921+]        
#> [170] 170: (0.0000, 0.1489], (0.1489, 0.5603], ..., (30.3105, 40.7827+]  
#> [171] 171: (0.0000, 0.3585], (0.3585, 0.5097], ..., (12.9871, 20.8588+]  
#> [172] 172: (0.0000, 1.5274], (1.5274, 13.2414], (13.2414, 60.0000+]      
#> [173] 173: (0.0000, 60.0000+]                                            
#> [174] 174: (0.0000, 0.4313], (0.4313, 1.5870], ..., (52.8374, 60.0000+]  
#> [175] 175: (0.0000, 4.3355+]                                             
#> [176] 176: (0.0000, 29.4135], (29.4135, 60.0000+]                        
#> [177] 177: (0.0000, 15.4249+]                                            
#> [178] 178: (0.0000, 2.8364], (2.8364, 13.0248+]                          
#> [179] 179: (0.0000, 0.1827], (0.1827, 0.2604], (0.2604, 36.1902+]        
#> [180] 180: (0.0000, 0.8836], (0.8836, 0.9499*]                           
#> [181] 181: (0.0000, 0.5596], (0.5596, 2.3804+]                           
#> [182] 182: (0.0000, 0.3451], (0.3451, 1.9079], ..., (50.8194, 60.0000+]  
#> [183] 183: (0.0000, 0.9137], (0.9137, 17.8781], ..., (28.2033, 60.0000+] 
#> [184] 184: (0.0000, 1.2003], (1.2003, 3.4004], (3.4004, 60.0000+]        
#> [185] 185: (0.0000, 1.4699], (1.4699, 7.1552*]                           
#> [186] 186: (0.0000, 0.1166], (0.1166, 0.2692], ..., (3.1273, 35.7107+]   
#> [187] 187: (0.0000, 0.0804], (0.0804, 60.0000+]                          
#> [188] 188: (0.0000, 1.4822*]                                             
#> [189] 189: (0.0000, 1.0906], (1.0906, 60.0000+]                          
#> [190] 190: (0.0000, 2.2382], (2.2382, 3.0685], ..., (18.2445, 28.8604+]  
#> [191] 191: (0.0000, 2.3099], (2.3099, 4.1342], ..., (29.8602, 37.1067+]  
#> [192] 192: (0.0000, 5.0416], (5.0416, 6.8241], ..., (8.1832, 15.4027+]   
#> [193] 193: (0.0000, 19.0580+]                                            
#> [194] 194: (0.0000, 0.0559], (0.0559, 1.3729], ..., (3.8318, 4.3652*]    
#> [195] 195: (0.0000, 1.3455], (1.3455, 12.2261], ..., (28.8277, 51.8301+] 
#> [196] 196: (0.0000, 0.0520], (0.0520, 0.1116], ..., (3.7786, 3.9391*]    
#> [197] 197: (0.0000, 1.8058], (1.8058, 31.8297], (31.8297, 34.0967+]      
#> [198] 198: (0.0000, 0.3472], (0.3472, 0.4190], ..., (24.0310, 32.8532*]  
#> [199] 199: (0.0000, 1.2344], (1.2344, 3.9052], (3.9052, 5.8988*]         
#> [200] 200: (0.0000, 4.3315], (4.3315, 15.2804], ..., (30.2639, 60.0000+] 

## Use Recur() instead
with(dat, Recur(Time, id, event, status))

#>   [1] 1: (0.0000, 4.1939+]                                               
#>   [2] 2: (0.0000, 0.4220], (0.4220, 0.7060], ..., (1.2931, 1.7378*]      
#>   [3] 3: (0.0000, 27.8297], (27.8297, 49.3334+]                          
#>   [4] 4: (0.0000, 3.5663+]                                               
#>   [5] 5: (0.0000, 0.0276], (0.0276, 0.3024], ..., (35.5598, 59.8088+]    
#>   [6] 6: (0.0000, 0.9859], (0.9859, 2.3660], ..., (3.4217, 8.1096+]      
#>   [7] 7: (0.0000, 1.3793], (1.3793, 10.4272], ..., (25.3046, 45.9819+]   
#>   [8] 8: (0.0000, 0.0564], (0.0564, 0.2060], ..., (1.1791, 1.1937*]      
#>   [9] 9: (0.0000, 3.1155], (3.1155, 3.7286], ..., (21.1107, 53.5430+]    
#>  [10] 10: (0.0000, 0.0224], (0.0224, 1.3998], ..., (14.3499, 30.2477+]   
#>  [11] 11: (0.0000, 0.0755], (0.0755, 0.2167], ..., (4.0920, 7.8248+]     
#>  [12] 12: (0.0000, 5.2976], (5.2976, 60.0000+]                           
#>  [13] 13: (0.0000, 0.7602], (0.7602, 0.9261], ..., (12.3359, 60.0000+]   
#>  [14] 14: (0.0000, 1.3152+]                                              
#>  [15] 15: (0.0000, 0.0469], (0.0469, 0.3025], ..., (21.9283, 26.7154*]   
#>  [16] 16: (0.0000, 60.0000+]                                             
#>  [17] 17: (0.0000, 32.4884+]                                             
#>  [18] 18: (0.0000, 60.0000+]                                             
#>  [19] 19: (0.0000, 21.7731+]                                             
#>  [20] 20: (0.0000, 2.2005], (2.2005, 28.5968+]                           
#>  [21] 21: (0.0000, 19.8022], (19.8022, 40.9622+]                         
#>  [22] 22: (0.0000, 0.9727], (0.9727, 2.1431], ..., (28.4250, 60.0000+]   
#>  [23] 23: (0.0000, 2.7144], (2.7144, 7.3384+]                            
#>  [24] 24: (0.0000, 0.0762], (0.0762, 0.1865], ..., (9.6314, 13.5243+]    
#>  [25] 25: (0.0000, 0.6611], (0.6611, 0.7965], ..., (16.1909, 24.3334+]   
#>  [26] 26: (0.0000, 0.0385], (0.0385, 0.1608], ..., (6.6432, 6.8203*]     
#>  [27] 27: (0.0000, 0.0382], (0.0382, 0.0985], ..., (18.3202, 60.0000+]   
#>  [28] 28: (0.0000, 0.0091*]                                              
#>  [29] 29: (0.0000, 2.9126+]                                              
#>  [30] 30: (0.0000, 0.9483], (0.9483, 1.1582], (1.1582, 2.7012+]          
#>  [31] 31: (0.0000, 0.1602], (0.1602, 0.8470], ..., (20.7914, 60.0000+]   
#>  [32] 32: (0.0000, 4.0675], (4.0675, 60.0000+]                           
#>  [33] 33: (0.0000, 0.0018], (0.0018, 0.1633], ..., (0.3409, 0.4080*]     
#>  [34] 34: (0.0000, 0.2069], (0.2069, 0.9686], ..., (7.2497, 8.2745+]     
#>  [35] 35: (0.0000, 38.6777], (38.6777, 60.0000+]                         
#>  [36] 36: (0.0000, 0.7862], (0.7862, 0.8419], ..., (49.2388, 60.0000+]   
#>  [37] 37: (0.0000, 0.8320], (0.8320, 6.3206], ..., (53.3568, 60.0000+]   
#>  [38] 38: (0.0000, 0.0254], (0.0254, 0.1117], ..., (3.0250, 3.2085+]     
#>  [39] 39: (0.0000, 1.9214], (1.9214, 13.0673+]                           
#>  [40] 40: (0.0000, 0.5774], (0.5774, 0.7782], (0.7782, 1.0653*]          
#>  [41] 41: (0.0000, 0.3697], (0.3697, 2.3635], ..., (4.3296, 60.0000+]    
#>  [42] 42: (0.0000, 0.0709], (0.0709, 0.0770], ..., (0.6271, 1.0153*]     
#>  [43] 43: (0.0000, 0.3110*]                                              
#>  [44] 44: (0.0000, 2.1150*]                                              
#>  [45] 45: (0.0000, 0.2198], (0.2198, 0.6528], ..., (17.1894, 17.3457+]   
#>  [46] 46: (0.0000, 1.3165], (1.3165, 3.8825], ..., (5.5116, 16.9859+]    
#>  [47] 47: (0.0000, 0.3050], (0.3050, 0.3518], ..., (0.6241, 2.8628*]     
#>  [48] 48: (0.0000, 13.4123], (13.4123, 19.7926+]                         
#>  [49] 49: (0.0000, 0.3423], (0.3423, 0.5334], ..., (43.5570, 60.0000+]   
#>  [50] 50: (0.0000, 4.1856], (4.1856, 5.5388], ..., (29.1396, 60.0000+]   
#>  [51] 51: (0.0000, 0.4839], (0.4839, 0.4968], ..., (0.7886, 0.9674+]     
#>  [52] 52: (0.0000, 1.2673], (1.2673, 4.6764], ..., (23.6977, 31.4293+]   
#>  [53] 53: (0.0000, 0.9122], (0.9122, 1.4916+]                            
#>  [54] 54: (0.0000, 0.0561], (0.0561, 1.3932], ..., (14.5763, 33.4254+]   
#>  [55] 55: (0.0000, 3.4611], (3.4611, 7.8629+]                            
#>  [56] 56: (0.0000, 0.0483], (0.0483, 0.1243], ..., (0.7432, 3.8557*]     
#>  [57] 57: (0.0000, 31.0839+]                                             
#>  [58] 58: (0.0000, 0.3138], (0.3138, 0.7740], ..., (22.9936, 23.5281+]   
#>  [59] 59: (0.0000, 0.2635], (0.2635, 0.6228], ..., (2.0408, 2.7623*]     
#>  [60] 60: (0.0000, 0.5610], (0.5610, 58.8840+]                           
#>  [61] 61: (0.0000, 0.3918*]                                              
#>  [62] 62: (0.0000, 12.0660], (12.0660, 14.9237+]                         
#>  [63] 63: (0.0000, 0.0514], (0.0514, 0.1477], ..., (29.4005, 60.0000+]   
#>  [64] 64: (0.0000, 0.1402], (0.1402, 16.2452], (16.2452, 60.0000+]       
#>  [65] 65: (0.0000, 0.0285], (0.0285, 0.1528], ..., (8.1595, 9.4588*]     
#>  [66] 66: (0.0000, 0.5731*]                                              
#>  [67] 67: (0.0000, 0.9741], (0.9741, 2.0285], ..., (2.2902, 3.4172+]     
#>  [68] 68: (0.0000, 0.7058+]                                              
#>  [69] 69: (0.0000, 2.3642], (2.3642, 2.4615], (2.4615, 60.0000+]         
#>  [70] 70: (0.0000, 0.1671], (0.1671, 3.2458], ..., (3.4636, 5.4292*]     
#>  [71] 71: (0.0000, 0.9312], (0.9312, 8.3824], (8.3824, 60.0000+]         
#>  [72] 72: (0.0000, 0.2243], (0.2243, 0.3831], ..., (3.8633, 4.3904*]     
#>  [73] 73: (0.0000, 19.1460+]                                             
#>  [74] 74: (0.0000, 0.1218], (0.1218, 10.3010], (10.3010, 60.0000+]       
#>  [75] 75: (0.0000, 1.2854], (1.2854, 60.0000+]                           
#>  [76] 76: (0.0000, 0.7864], (0.7864, 60.0000+]                           
#>  [77] 77: (0.0000, 14.7017], (14.7017, 47.1695+]                         
#>  [78] 78: (0.0000, 0.1847], (0.1847, 0.6563], ..., (5.9610, 10.8273+]    
#>  [79] 79: (0.0000, 1.1958], (1.1958, 1.6888], ..., (22.3231, 60.0000+]   
#>  [80] 80: (0.0000, 0.2668], (0.2668, 0.4614], ..., (4.7256, 9.6897*]     
#>  [81] 81: (0.0000, 0.2466], (0.2466, 0.3025], ..., (28.2620, 43.7449+]   
#>  [82] 82: (0.0000, 0.0199], (0.0199, 0.1099], ..., (1.4919, 3.7963+]     
#>  [83] 83: (0.0000, 0.9455+]                                              
#>  [84] 84: (0.0000, 0.4360], (0.4360, 0.4424], ..., (41.9064, 60.0000+]   
#>  [85] 85: (0.0000, 1.4408+]                                              
#>  [86] 86: (0.0000, 0.0022], (0.0022, 0.0164], ..., (11.8213, 12.2415*]   
#>  [87] 87: (0.0000, 0.2622], (0.2622, 2.3970+]                            
#>  [88] 88: (0.0000, 4.0281+]                                              
#>  [89] 89: (0.0000, 0.1192], (0.1192, 0.6737], ..., (19.1438, 40.7039+]   
#>  [90] 90: (0.0000, 2.6533], (2.6533, 7.5542], ..., (16.7730, 60.0000+]   
#>  [91] 91: (0.0000, 0.1813], (0.1813, 3.7734], ..., (38.9600, 39.2281+]   
#>  [92] 92: (0.0000, 13.0989], (13.0989, 20.6036], ..., (45.2517, 60.0000+]
#>  [93] 93: (0.0000, 2.3861], (2.3861, 3.3904], (3.3904, 10.8095+]         
#>  [94] 94: (0.0000, 0.8110], (0.8110, 1.5011], ..., (14.5904, 15.0829*]   
#>  [95] 95: (0.0000, 9.3932], (9.3932, 19.8858+]                           
#>  [96] 96: (0.0000, 6.0415], (6.0415, 16.3964*]                           
#>  [97] 97: (0.0000, 0.1147], (0.1147, 0.4367], ..., (27.4787, 60.0000+]   
#>  [98] 98: (0.0000, 14.8624+]                                             
#>  [99] 99: (0.0000, 16.1137+]                                             
#> [100] 100: (0.0000, 2.0292], (2.0292, 17.5638], ..., (24.7633, 30.4540+] 
#> [101] 101: (0.0000, 0.2228], (0.2228, 0.2259], ..., (0.4402, 0.5615*]    
#> [102] 102: (0.0000, 0.0494], (0.0494, 0.3593], ..., (1.2053, 1.2369*]    
#> [103] 103: (0.0000, 27.1265+]                                            
#> [104] 104: (0.0000, 0.8268], (0.8268, 2.7076], ..., (19.2702, 19.3323+]  
#> [105] 105: (0.0000, 2.5198*]                                             
#> [106] 106: (0.0000, 0.1236], (0.1236, 0.1819], ..., (3.1792, 3.8315*]    
#> [107] 107: (0.0000, 11.2995+]                                            
#> [108] 108: (0.0000, 0.1591], (0.1591, 0.2060], (0.2060, 3.8735+]         
#> [109] 109: (0.0000, 0.0429], (0.0429, 0.0576], ..., (0.1464, 0.1905*]    
#> [110] 110: (0.0000, 0.0641*]                                             
#> [111] 111: (0.0000, 0.1070], (0.1070, 1.0162], ..., (10.1607, 19.4814+]  
#> [112] 112: (0.0000, 15.9173], (15.9173, 17.9019], (17.9019, 60.0000+]    
#> [113] 113: (0.0000, 0.0657], (0.0657, 0.2235], ..., (0.3495, 0.7056*]    
#> [114] 114: (0.0000, 11.8459], (11.8459, 23.2224*]                        
#> [115] 115: (0.0000, 0.1231], (0.1231, 2.6182], ..., (3.4555, 4.4275+]    
#> [116] 116: (0.0000, 0.5347], (0.5347, 1.2448], (1.2448, 1.3201+]         
#> [117] 117: (0.0000, 9.6803], (9.6803, 12.2911], (12.2911, 60.0000+]      
#> [118] 118: (0.0000, 0.9918], (0.9918, 60.0000+]                          
#> [119] 119: (0.0000, 0.4323], (0.4323, 0.9787], ..., (57.0146, 60.0000+]  
#> [120] 120: (0.0000, 0.0667], (0.0667, 0.1311*]                           
#> [121] 121: (0.0000, 7.2813+]                                             
#> [122] 122: (0.0000, 0.0490], (0.0490, 0.2071], ..., (13.7969, 30.5598+]  
#> [123] 123: (0.0000, 60.0000+]                                            
#> [124] 124: (0.0000, 0.3529], (0.3529, 9.7130], (9.7130, 11.7963*]        
#> [125] 125: (0.0000, 0.8298], (0.8298, 2.6456], ..., (3.3778, 4.3753+]    
#> [126] 126: (0.0000, 3.3505], (3.3505, 47.9797*]                          
#> [127] 127: (0.0000, 0.1545], (0.1545, 0.4070], ..., (0.5757, 3.0877+]    
#> [128] 128: (0.0000, 0.0334], (0.0334, 60.0000+]                          
#> [129] 129: (0.0000, 1.4197+]                                             
#> [130] 130: (0.0000, 2.2834], (2.2834, 21.0913+]                          
#> [131] 131: (0.0000, 0.0567+]                                             
#> [132] 132: (0.0000, 0.6541], (0.6541, 2.3745*]                           
#> [133] 133: (0.0000, 0.5389], (0.5389, 0.8951], ..., (6.7334, 22.9528+]   
#> [134] 134: (0.0000, 0.1956], (0.1956, 0.2891], ..., (7.2722, 7.5890*]    
#> [135] 135: (0.0000, 0.1912], (0.1912, 0.1964], (0.1964, 0.2359*]         
#> [136] 136: (0.0000, 0.4751], (0.4751, 1.0699], ..., (24.5560, 38.2816+]  
#> [137] 137: (0.0000, 16.4606+]                                            
#> [138] 138: (0.0000, 0.1685], (0.1685, 2.3897], (2.3897, 44.2901+]        
#> [139] 139: (0.0000, 0.3702], (0.3702, 60.0000+]                          
#> [140] 140: (0.0000, 0.0995], (0.0995, 0.3184], ..., (56.4982, 60.0000+]  
#> [141] 141: (0.0000, 22.2126+]                                            
#> [142] 142: (0.0000, 0.6679], (0.6679, 1.2204], (1.2204, 1.3808*]         
#> [143] 143: (0.0000, 0.0215], (0.0215, 0.0998], ..., (0.3990, 0.6926*]    
#> [144] 144: (0.0000, 0.3498], (0.3498, 0.4453], ..., (3.6977, 4.5617+]    
#> [145] 145: (0.0000, 1.3728], (1.3728, 2.7651], ..., (16.4156, 18.2321+]  
#> [146] 146: (0.0000, 0.2692], (0.2692, 0.8881], ..., (4.4387, 9.9300*]    
#> [147] 147: (0.0000, 7.4812+]                                             
#> [148] 148: (0.0000, 0.9865], (0.9865, 2.2654], ..., (7.0282, 7.9999+]    
#> [149] 149: (0.0000, 0.3762], (0.3762, 0.4328], ..., (1.8875, 2.4293*]    
#> [150] 150: (0.0000, 0.0910], (0.0910, 0.4390], ..., (2.1062, 25.1158+]   
#> [151] 151: (0.0000, 3.4940], (3.4940, 25.3259], (25.3259, 45.1222+]      
#> [152] 152: (0.0000, 1.7864], (1.7864, 7.2195], (7.2195, 9.5792+]         
#> [153] 153: (0.0000, 0.7874], (0.7874, 2.6515*]                           
#> [154] 154: (0.0000, 5.5211], (5.5211, 41.6133+]                          
#> [155] 155: (0.0000, 0.1137], (0.1137, 9.6278], (9.6278, 20.6738*]        
#> [156] 156: (0.0000, 2.7274], (2.7274, 4.3713*]                           
#> [157] 157: (0.0000, 0.1500], (0.1500, 0.1602], ..., (13.6229, 17.2046+]  
#> [158] 158: (0.0000, 0.9707], (0.9707, 2.6660*]                           
#> [159] 159: (0.0000, 12.8927], (12.8927, 23.6012+]                        
#> [160] 160: (0.0000, 0.0813], (0.0813, 0.1225], ..., (11.5289, 15.4727+]  
#> [161] 161: (0.0000, 0.6257], (0.6257, 0.7828], ..., (37.4610, 60.0000+]  
#> [162] 162: (0.0000, 0.2446], (0.2446, 17.6765+]                          
#> [163] 163: (0.0000, 1.5187], (1.5187, 3.9270], ..., (17.8582, 22.7722*]  
#> [164] 164: (0.0000, 0.0294], (0.0294, 0.0855], ..., (1.1457, 1.2137*]    
#> [165] 165: (0.0000, 0.0554], (0.0554, 0.1506], ..., (2.9271, 3.1379*]    
#> [166] 166: (0.0000, 0.6354+]                                             
#> [167] 167: (0.0000, 2.2898*]                                             
#> [168] 168: (0.0000, 0.6285], (0.6285, 0.8874], ..., (10.8140, 25.8737*]  
#> [169] 169: (0.0000, 0.6257], (0.6257, 7.3773], (7.3773, 31.4921+]        
#> [170] 170: (0.0000, 0.1489], (0.1489, 0.5603], ..., (30.3105, 40.7827+]  
#> [171] 171: (0.0000, 0.3585], (0.3585, 0.5097], ..., (12.9871, 20.8588+]  
#> [172] 172: (0.0000, 1.5274], (1.5274, 13.2414], (13.2414, 60.0000+]      
#> [173] 173: (0.0000, 60.0000+]                                            
#> [174] 174: (0.0000, 0.4313], (0.4313, 1.5870], ..., (52.8374, 60.0000+]  
#> [175] 175: (0.0000, 4.3355+]                                             
#> [176] 176: (0.0000, 29.4135], (29.4135, 60.0000+]                        
#> [177] 177: (0.0000, 15.4249+]                                            
#> [178] 178: (0.0000, 2.8364], (2.8364, 13.0248+]                          
#> [179] 179: (0.0000, 0.1827], (0.1827, 0.2604], (0.2604, 36.1902+]        
#> [180] 180: (0.0000, 0.8836], (0.8836, 0.9499*]                           
#> [181] 181: (0.0000, 0.5596], (0.5596, 2.3804+]                           
#> [182] 182: (0.0000, 0.3451], (0.3451, 1.9079], ..., (50.8194, 60.0000+]  
#> [183] 183: (0.0000, 0.9137], (0.9137, 17.8781], ..., (28.2033, 60.0000+] 
#> [184] 184: (0.0000, 1.2003], (1.2003, 3.4004], (3.4004, 60.0000+]        
#> [185] 185: (0.0000, 1.4699], (1.4699, 7.1552*]                           
#> [186] 186: (0.0000, 0.1166], (0.1166, 0.2692], ..., (3.1273, 35.7107+]   
#> [187] 187: (0.0000, 0.0804], (0.0804, 60.0000+]                          
#> [188] 188: (0.0000, 1.4822*]                                             
#> [189] 189: (0.0000, 1.0906], (1.0906, 60.0000+]                          
#> [190] 190: (0.0000, 2.2382], (2.2382, 3.0685], ..., (18.2445, 28.8604+]  
#> [191] 191: (0.0000, 2.3099], (2.3099, 4.1342], ..., (29.8602, 37.1067+]  
#> [192] 192: (0.0000, 5.0416], (5.0416, 6.8241], ..., (8.1832, 15.4027+]   
#> [193] 193: (0.0000, 19.0580+]                                            
#> [194] 194: (0.0000, 0.0559], (0.0559, 1.3729], ..., (3.8318, 4.3652*]    
#> [195] 195: (0.0000, 1.3455], (1.3455, 12.2261], ..., (28.8277, 51.8301+] 
#> [196] 196: (0.0000, 0.0520], (0.0520, 0.1116], ..., (3.7786, 3.9391*]    
#> [197] 197: (0.0000, 1.8058], (1.8058, 31.8297], (31.8297, 34.0967+]      
#> [198] 198: (0.0000, 0.3472], (0.3472, 0.4190], ..., (24.0310, 32.8532*]  
#> [199] 199: (0.0000, 1.2344], (1.2344, 3.9052], (3.9052, 5.8988*]         
#> [200] 200: (0.0000, 4.3315], (4.3315, 15.2804], ..., (30.2639, 60.0000+]