Function to generate simulated recurrent event data

The function simGSC() generates simulated recurrent event data from either a Cox-type model, an accelerated mean model, an accelerated rate model, or a generalized scale-change model.

simGSC(
  n,
  summary = FALSE,
  para,
  xmat,
  censoring,
  frailty,
  tau,
  origin,
  Lam0,
  Haz0
)

Arguments

n

number of observation.

summary

a logical value indicating whether a brief data summary will be printed.

para

a list of numerical vectors for the regression coefficients in the joint scale-change model. The names of the list elements are alpha, beta, eta, and theta, correspond to \(\alpha\), \(\beta\), \(\eta\), and \(\theta\) in the joint scale-change model, respectively. See Details for reReg.

xmat

an optional matrix specifying the design matrix.

censoring

a numeric variable specifying the censoring times for each of the \(n\) observation.

frailty

a numeric variable specifying the frailty variable.

tau

a numeric value specifying the maximum observation time.

origin

a numeric value specifying the time origin.

Lam0

is an optional function that specifies the baseline cumulative rate function. When left-unspecified, the recurrent events are generated using the baseline rate function of $$\lambda_0(t) = \frac{2}{1 + t},$$ or equivalently, the cumulative rate function of $$\Lambda_0(t) = 2\log(1 + t).$$

Haz0

is an optional function that specifies the baseline hazard function. When left-unspecified, the recurrent events are generated using the baseline hazard function $$h_0(t) = \frac{1}{5(1 + t)},$$ or equivalently, the cumulative hazard function of $$H_0(t) = \log(1 + t) / 5.$$

Details

The function simGSC() generates simulated recurrent event data over the interval \((0, \tau)\) based on the specification of the recurrent process and the terminal events. Specifically, the rate function, \(\lambda(t)\), of the recurrent process can be specified as one of the following model: $$\lambda(t) = Z \lambda_0(te^{X^\top\alpha}) e^{X^\top\beta}, h(t) = Z h_0(te^{X^\top\eta})e^{X^\top\theta}, $$ where \(\lambda_0(t)\) is the baseline rate function, \(h_0(t)\) is the baseline hazard function, \(X\) is a \(n\) by \(p\) covariate matrix and \(\alpha\), \(Z\) is an unobserved shared frailty variable, and \((\alpha, \eta)\) and \((\beta, \theta)\) correspond to the shape and size parameters of the rate function and the hazard function, respectively.

Under the default settings, the simGSC() function assumes \(p = 2\) and the regression parameters to be \(\alpha = \eta = (0, 0)^\top\), and \(\beta = \theta = (1, 1)^\top\). When the xmat argument is not specified, the simGSC() function assumes \(X_i\) is a two-dimensional vector \(X_i = (X_{i1}, X_{i2}), i = 1, \ldots, n\), where \(X_{i1}\) is a Bernoulli variable with rate 0.5 and \(X_{i2}\) is a standard normal variable. With the default xmat, the censoring time $C$ is generated from an exponential distribution with mean \(\tau X_{i1} + Z^2\tau(1 - X_{i1})\). Thus, the censoring distribution is covariate dependent and is informative when \(Z\) is not a constant. When the frailty argument is not specified, the frailty variable \(Z\) is generated from a gamma distribution with a unit mean and a variance of 0.25. The default values for tau and origin are 60 and 0, respectively. When arguments Lam0 and Haz0 are left unspecified, the simGSC() function uses \(\Lambda_0(t) = 2\log(1 + t)\) and \(H_0(t) = \log(1 + t) / 5\), respectively. This is equivalent to setting Lam0 = function(x) 2 * log(1 + x) and Haz0 = function(x) log(1 + x) / 5. Overall, the default specifications generate the recurrent events and the terminal events from the model: $$\lambda(t) = \displaystyle \frac{2Z}{1 + te^{-X_{i1} - X_{i2}}}, h(t) = \displaystyle \frac{Z}{5(1 + te^{X_{i1} + X_{i2}})}, t\in[0, 60].$$ See online vignette for more examples.

See also

Examples

set.seed(123)
simGSC(100, summary = TRUE)
#> Call: 
#> simGSC(n = 100, summary = TRUE)
#> 
#> Summary:
#> Sample size:                                    100 
#> Number of recurrent event observed:             343 
#> Average number of recurrent event per subject:  3.43 
#> Proportion of subjects with a terminal event:   0.53 
#> Median time-to-terminal event:                  11.202 
#> 
#> 
#>      id     t.start      t.stop event status x1           x2
#> 1     1  0.00000000  0.79162625     1      0  1  0.157076450
#> 2     1  0.79162625  1.26276390     1      0  1  0.157076450
#> 3     1  1.26276390 45.11318934     1      0  1  0.157076450
#> 4     1 45.11318934 60.00000000     0      0  1  0.157076450
#> 5     2  0.00000000  0.30227581     1      0  0  0.922419336
#> 6     2  0.30227581  0.90335773     1      0  0  0.922419336
#> 7     2  0.90335773  2.38441607     1      0  0  0.922419336
#> 8     2  2.38441607  2.60041466     1      0  0  0.922419336
#> 9     2  2.60041466  2.96855402     0      1  0  0.922419336
#> 10    3  0.00000000 60.00000000     0      0  1 -0.490958575
#> 11    4  0.00000000  1.43868491     0      1  1  1.098001882
#> 12    5  0.00000000  0.26107532     1      0  0 -0.102333833
#> 13    5  0.26107532  0.46163165     1      0  0 -0.102333833
#> 14    5  0.46163165  0.46192449     1      0  0 -0.102333833
#> 15    5  0.46192449  1.14105041     1      0  0 -0.102333833
#> 16    5  1.14105041  3.82760327     1      0  0 -0.102333833
#> 17    5  3.82760327  4.06105187     1      0  0 -0.102333833
#> 18    5  4.06105187  5.41661163     1      0  0 -0.102333833
#> 19    5  5.41661163  5.50413410     1      0  0 -0.102333833
#> 20    5  5.50413410  6.50628946     1      0  0 -0.102333833
#> 21    5  6.50628946 11.74388386     1      0  0 -0.102333833
#> 22    5 11.74388386 13.00832888     1      0  0 -0.102333833
#> 23    5 13.00832888 13.60817463     1      0  0 -0.102333833
#> 24    5 13.60817463 20.85935224     1      0  0 -0.102333833
#> 25    5 20.85935224 36.50247425     1      0  0 -0.102333833
#> 26    5 36.50247425 37.56205361     1      0  0 -0.102333833
#> 27    5 37.56205361 44.90413413     1      0  0 -0.102333833
#> 28    5 44.90413413 60.00000000     0      0  0 -0.102333833
#> 29    6  0.00000000  0.07968560     0      1  0  0.487571468
#> 30    7  0.00000000  0.52713723     0      0  0 -0.433783058
#> 31    8  0.00000000  0.07596889     0      0  0 -0.250593796
#> 32    9  0.00000000  0.15887552     1      0  0  0.392795581
#> 33    9  0.15887552  0.45145920     1      0  0  0.392795581
#> 34    9  0.45145920  0.95050460     1      0  0  0.392795581
#> 35    9  0.95050460  1.22943797     0      1  0  0.392795581
#> 36   10  0.00000000  0.58502638     1      0  1 -1.051123662
#> 37   10  0.58502638  1.88079670     1      0  1 -1.051123662
#> 38   10  1.88079670  2.25568405     1      0  1 -1.051123662
#> 39   10  2.25568405  3.30296307     1      0  1 -1.051123662
#> 40   10  3.30296307  4.97999693     1      0  1 -1.051123662
#> 41   10  4.97999693  6.98154161     1      0  1 -1.051123662
#> 42   10  6.98154161 31.61794063     0      1  1 -1.051123662
#> 43   11  0.00000000  4.33261659     1      0  0 -0.021102465
#> 44   11  4.33261659  6.77366561     1      0  0 -0.021102465
#> 45   11  6.77366561  8.22945935     1      0  0 -0.021102465
#> 46   11  8.22945935  9.61893919     0      0  0 -0.021102465
#> 47   12  0.00000000 18.03138005     0      0  1 -0.202404707
#> 48   13  0.00000000  9.04088247     1      0  1 -0.056382983
#> 49   13  9.04088247 11.20185483     0      1  1 -0.056382983
#> 50   14  0.00000000  0.44361488     0      1  1  0.898570220
#> 51   15  0.00000000  0.08839869     1      0  0 -0.405163129
#> 52   15  0.08839869  0.18584832     1      0  0 -0.405163129
#> 53   15  0.18584832  1.79615217     1      0  0 -0.405163129
#> 54   15  1.79615217  2.08960718     1      0  0 -0.405163129
#> 55   15  2.08960718  3.14205836     1      0  0 -0.405163129
#> 56   15  3.14205836  3.24554692     1      0  0 -0.405163129
#> 57   15  3.24554692  4.30789368     1      0  0 -0.405163129
#> 58   15  4.30789368  6.58344304     1      0  0 -0.405163129
#> 59   15  6.58344304  6.94031987     1      0  0 -0.405163129
#> 60   15  6.94031987  7.19481553     1      0  0 -0.405163129
#> 61   15  7.19481553  8.02533680     1      0  0 -0.405163129
#> 62   15  8.02533680 10.47299913     1      0  0 -0.405163129
#> 63   15 10.47299913 11.10968894     1      0  0 -0.405163129
#> 64   15 11.10968894 11.27363789     1      0  0 -0.405163129
#> 65   15 11.27363789 25.48988929     1      0  0 -0.405163129
#> 66   15 25.48988929 60.00000000     0      0  0 -0.405163129
#> 67   16  0.00000000  0.57676712     0      1  1  0.950450037
#> 68   17  0.00000000  0.54498747     1      0  0  0.354477220
#> 69   17  0.54498747  1.13209144     0      1  0  0.354477220
#> 70   18  0.00000000  3.02912424     0      1  1  0.368097387
#> 71   19  0.00000000  9.25047301     0      1  0  0.682888277
#> 72   20  0.00000000  0.05210728     0      1  0 -0.288131946
#> 73   21  0.00000000  0.77471756     1      0  0 -0.402366160
#> 74   21  0.77471756  1.60270233     1      0  0 -0.402366160
#> 75   21  1.60270233  1.78024184     1      0  0 -0.402366160
#> 76   21  1.78024184  1.88561006     1      0  0 -0.402366160
#> 77   21  1.88561006  3.19259507     1      0  0 -0.402366160
#> 78   21  3.19259507 11.79773426     1      0  0 -0.402366160
#> 79   21 11.79773426 15.21785161     0      0  0 -0.402366160
#> 80   22  0.00000000  1.37767984     0      0  0 -0.267532906
#> 81   23  0.00000000  1.50594395     1      0  0  0.395773955
#> 82   23  1.50594395  1.75861105     1      0  0  0.395773955
#> 83   23  1.75861105  1.80267914     1      0  0  0.395773955
#> 84   23  1.80267914  4.51812448     1      0  0  0.395773955
#> 85   23  4.51812448  5.28666642     1      0  0  0.395773955
#> 86   23  5.28666642  6.53089379     1      0  0  0.395773955
#> 87   23  6.53089379 13.53141563     0      0  0  0.395773955
#> 88   24  0.00000000  0.41991347     1      0  0 -0.353803575
#> 89   24  0.41991347  2.33241050     1      0  0 -0.353803575
#> 90   24  2.33241050  2.92127177     1      0  0 -0.353803575
#> 91   24  2.92127177  3.10026121     1      0  0 -0.353803575
#> 92   24  3.10026121  4.52826189     1      0  0 -0.353803575
#> 93   24  4.52826189  6.08884111     1      0  0 -0.353803575
#> 94   24  6.08884111  6.45393470     0      1  0 -0.353803575
#> 95   25  0.00000000  6.29397761     0      0  1 -0.637771686
#> 96   26  0.00000000  1.13027008     1      0  0  1.187732231
#> 97   26  1.13027008  1.98111944     1      0  0  1.187732231
#> 98   26  1.98111944  2.04596710     0      1  0  1.187732231
#> 99   27  0.00000000  0.05681231     0      1  1 -0.546794874
#> 100  28  0.00000000  0.20117010     0      1  0  0.096220791
#> 101  29  0.00000000  0.64909412     1      0  0 -0.063080191
#> 102  29  0.64909412  1.33593293     0      0  0 -0.063080191
#> 103  30  0.00000000  1.50659690     1      0  0 -0.694216070
#> 104  30  1.50659690  2.51742284     1      0  0 -0.694216070
#> 105  30  2.51742284  3.04602864     1      0  0 -0.694216070
#> 106  30  3.04602864  4.57322485     1      0  0 -0.694216070
#> 107  30  4.57322485  4.75602077     1      0  0 -0.694216070
#> 108  30  4.75602077  7.91947850     1      0  0 -0.694216070
#> 109  30  7.91947850  9.55289875     1      0  0 -0.694216070
#> 110  30  9.55289875 10.00277008     1      0  0 -0.694216070
#> 111  30 10.00277008 10.64665836     1      0  0 -0.694216070
#> 112  30 10.64665836 11.03089566     1      0  0 -0.694216070
#> 113  30 11.03089566 11.08933110     1      0  0 -0.694216070
#> 114  30 11.08933110 12.09100807     1      0  0 -0.694216070
#> 115  30 12.09100807 26.91632324     1      0  0 -0.694216070
#> 116  30 26.91632324 44.82494415     0      0  0 -0.694216070
#> 117  31  0.00000000  0.93724384     1      0  0  0.234951096
#> 118  31  0.93724384  1.36756560     0      0  0  0.234951096
#> 119  32  0.00000000  0.94782802     1      0  1  0.480191179
#> 120  32  0.94782802 12.52976869     0      1  1  0.480191179
#> 121  33  0.00000000  0.07466582     0      1  1 -0.002570644
#> 122  34  0.00000000  0.50137025     1      0  0 -0.435232221
#> 123  34  0.50137025  0.65878014     1      0  0 -0.435232221
#> 124  34  0.65878014  2.05024074     1      0  0 -0.435232221
#> 125  34  2.05024074  2.55702490     1      0  0 -0.435232221
#> 126  34  2.55702490  8.43139702     1      0  0 -0.435232221
#> 127  34  8.43139702 10.84759217     1      0  0 -0.435232221
#> 128  34 10.84759217 23.15024896     0      0  0 -0.435232221
#> 129  35  0.00000000  0.22102875     1      0  0 -0.231439520
#> 130  35  0.22102875  0.51182518     1      0  0 -0.231439520
#> 131  35  0.51182518  0.62828817     1      0  0 -0.231439520
#> 132  35  0.62828817  0.63033587     1      0  0 -0.231439520
#> 133  35  0.63033587  1.32460875     1      0  0 -0.231439520
#> 134  35  1.32460875  2.16115727     1      0  0 -0.231439520
#> 135  35  2.16115727  2.37637214     1      0  0 -0.231439520
#> 136  35  2.37637214  2.46100310     1      0  0 -0.231439520
#> 137  35  2.46100310  4.25553284     1      0  0 -0.231439520
#> 138  35  4.25553284  7.63551503     1      0  0 -0.231439520
#> 139  35  7.63551503  9.49738975     1      0  0 -0.231439520
#> 140  35  9.49738975  9.89725753     1      0  0 -0.231439520
#> 141  35  9.89725753 12.91355879     0      1  0 -0.231439520
#> 142  36  0.00000000  0.34421769     1      0  0 -0.955575628
#> 143  36  0.34421769  0.52706635     1      0  0 -0.955575628
#> 144  36  0.52706635  1.68677164     1      0  0 -0.955575628
#> 145  36  1.68677164  2.24275156     1      0  0 -0.955575628
#> 146  36  2.24275156  7.11223899     0      0  0 -0.955575628
#> 147  37  0.00000000  0.65590477     1      0  1  0.184908022
#> 148  37  0.65590477  2.81058410     1      0  1  0.184908022
#> 149  37  2.81058410  9.74247339     0      1  1  0.184908022
#> 150  38  0.00000000  6.15402891     1      0  1 -0.231144029
#> 151  38  6.15402891 60.00000000     0      0  1 -0.231144029
#> 152  39  0.00000000  0.02654707     0      1  1  0.160977946
#> 153  40  0.00000000  0.07855291     1      0  0 -0.142000695
#> 154  40  0.07855291  0.29034769     1      0  0 -0.142000695
#> 155  40  0.29034769  1.26327474     1      0  0 -0.142000695
#> 156  40  1.26327474  2.27996544     1      0  0 -0.142000695
#> 157  40  2.27996544  2.65567126     1      0  0 -0.142000695
#> 158  40  2.65567126  7.18636429     1      0  0 -0.142000695
#> 159  40  7.18636429 11.14522228     1      0  0 -0.142000695
#> 160  40 11.14522228 21.42065920     1      0  0 -0.142000695
#> 161  40 21.42065920 23.85875892     1      0  0 -0.142000695
#> 162  40 23.85875892 42.40323869     1      0  0 -0.142000695
#> 163  40 42.40323869 60.00000000     0      0  0 -0.142000695
#> 164  41  0.00000000  0.95032158     1      0  1  0.545928222
#> 165  41  0.95032158  1.53798878     1      0  1  0.545928222
#> 166  41  1.53798878  4.93164904     0      1  1  0.545928222
#> 167  42  0.00000000  0.71398107     1      0  0  0.080108181
#> 168  42  0.71398107  0.72424859     0      1  0  0.080108181
#> 169  43  0.00000000  6.28047434     0      0  0  1.354084142
#> 170  44  0.00000000  0.16585079     0      0  0  0.143144965
#> 171  45  0.00000000  1.05565463     1      0  0 -0.016255125
#> 172  45  1.05565463  2.10469672     1      0  0 -0.016255125
#> 173  45  2.10469672  6.49277294     1      0  0 -0.016255125
#> 174  45  6.49277294  6.57651593     0      0  0 -0.016255125
#> 175  46  0.00000000  0.73564547     1      0  0 -0.494839010
#> 176  46  0.73564547  1.31027832     1      0  0 -0.494839010
#> 177  46  1.31027832  2.32866788     1      0  0 -0.494839010
#> 178  46  2.32866788  5.86168764     1      0  0 -0.494839010
#> 179  46  5.86168764  6.83580771     1      0  0 -0.494839010
#> 180  46  6.83580771  7.08830167     1      0  0 -0.494839010
#> 181  46  7.08830167 11.79661639     1      0  0 -0.494839010
#> 182  46 11.79661639 13.44688553     1      0  0 -0.494839010
#> 183  46 13.44688553 15.58755959     1      0  0 -0.494839010
#> 184  46 15.58755959 16.97441871     1      0  0 -0.494839010
#> 185  46 16.97441871 21.90748581     1      0  0 -0.494839010
#> 186  46 21.90748581 60.00000000     0      0  0 -0.494839010
#> 187  47  0.00000000  6.15409598     0      1  1  0.105904964
#> 188  48  0.00000000  4.57039521     1      0  1 -0.486008767
#> 189  48  4.57039521  9.78986723     1      0  1 -0.486008767
#> 190  48  9.78986723 28.18814054     1      0  1 -0.486008767
#> 191  48 28.18814054 60.00000000     0      0  1 -0.486008767
#> 192  49  0.00000000  0.61907998     0      1  1  0.274612007
#> 193  50  0.00000000  0.11987149     1      0  0 -0.524510765
#> 194  50  0.11987149  0.83112324     1      0  0 -0.524510765
#> 195  50  0.83112324  1.65675604     1      0  0 -0.524510765
#> 196  50  1.65675604  1.66117712     0      0  0 -0.524510765
#> 197  51  0.00000000  0.12832105     1      0  0  0.224566967
#> 198  51  0.12832105  0.25821327     0      1  0  0.224566967
#> 199  52  0.00000000  0.94766476     1      0  1 -0.192927826
#> 200  52  0.94766476  3.88605892     1      0  1 -0.192927826
#> 201  52  3.88605892 23.27870623     1      0  1 -0.192927826
#> 202  52 23.27870623 26.92918303     1      0  1 -0.192927826
#> 203  52 26.92918303 41.38015492     0      0  1 -0.192927826
#> 204  53  0.00000000  0.35141989     1      0  1 -0.785067761
#> 205  53  0.35141989  1.59501429     1      0  1 -0.785067761
#> 206  53  1.59501429  6.25901255     1      0  1 -0.785067761
#> 207  53  6.25901255 45.91733962     0      0  1 -0.785067761
#> 208  54  0.00000000  0.54384052     1      0  0  0.613471301
#> 209  54  0.54384052  1.03401351     1      0  0  0.613471301
#> 210  54  1.03401351  1.23009416     1      0  0  0.613471301
#> 211  54  1.23009416  2.01223306     1      0  0  0.613471301
#> 212  54  2.01223306  3.24446113     1      0  0  0.613471301
#> 213  54  3.24446113  8.71315010     1      0  0  0.613471301
#> 214  54  8.71315010 16.24555194     1      0  0  0.613471301
#> 215  54 16.24555194 17.00400999     1      0  0  0.613471301
#> 216  54 17.00400999 28.59499664     1      0  0  0.613471301
#> 217  54 28.59499664 45.44767103     1      0  0  0.613471301
#> 218  54 45.44767103 60.00000000     0      0  0  0.613471301
#> 219  55  0.00000000  0.14591559     1      0  1  0.333347198
#> 220  55  0.14591559  0.43850607     1      0  1  0.333347198
#> 221  55  0.43850607  0.81476877     1      0  1  0.333347198
#> 222  55  0.81476877  3.09244733     0      1  1  0.333347198
#> 223  56  0.00000000  0.19336449     1      0  0 -0.575801582
#> 224  56  0.19336449  0.33180288     1      0  0 -0.575801582
#> 225  56  0.33180288  2.30109340     1      0  0 -0.575801582
#> 226  56  2.30109340  2.51210750     1      0  0 -0.575801582
#> 227  56  2.51210750  4.84482959     1      0  0 -0.575801582
#> 228  56  4.84482959 12.95972589     1      0  0 -0.575801582
#> 229  56 12.95972589 18.41915516     1      0  0 -0.575801582
#> 230  56 18.41915516 35.70057106     0      0  0 -0.575801582
#> 231  57  0.00000000  1.15811178     1      0  0 -0.080394242
#> 232  57  1.15811178  2.86500299     0      0  0 -0.080394242
#> 233  58  0.00000000  0.24401303     1      0  1 -0.486456888
#> 234  58  0.24401303  1.35890638     1      0  1 -0.486456888
#> 235  58  1.35890638  3.26246296     0      1  1 -0.486456888
#> 236  59  0.00000000  0.08723446     1      0  0 -0.305230830
#> 237  59  0.08723446  5.30702960     0      0  0 -0.305230830
#> 238  60  0.00000000  0.30145862     1      0  0  0.249638438
#> 239  60  0.30145862  0.39423451     1      0  0  0.249638438
#> 240  60  0.39423451  2.48048620     1      0  0  0.249638438
#> 241  60  2.48048620  4.32144450     1      0  0  0.249638438
#> 242  60  4.32144450  6.06868874     1      0  0  0.249638438
#> 243  60  6.06868874  7.12077742     1      0  0  0.249638438
#> 244  60  7.12077742 11.47633594     0      0  0  0.249638438
#> 245  61  0.00000000  4.65032702     0      0  0  0.439682070
#> 246  62  0.00000000  0.17498146     0      1  1  0.461564620
#> 247  63  0.00000000  0.42106093     1      0  0  0.346566986
#> 248  63  0.42106093  1.57534657     1      0  0  0.346566986
#> 249  63  1.57534657  2.91813256     0      1  0  0.346566986
#> 250  64  0.00000000  0.16921676     0      1  1 -0.069750162
#> 251  65  0.00000000  0.17859176     1      0  0 -1.655339363
#> 252  65  0.17859176  0.27838475     1      0  0 -1.655339363
#> 253  65  0.27838475  0.66610227     1      0  0 -1.655339363
#> 254  65  0.66610227  0.72066250     1      0  0 -1.655339363
#> 255  65  0.72066250  1.47180157     1      0  0 -1.655339363
#> 256  65  1.47180157  2.73261150     1      0  0 -1.655339363
#> 257  65  2.73261150  3.18454415     1      0  0 -1.655339363
#> 258  65  3.18454415  6.51805133     0      1  0 -1.655339363
#> 259  66  0.00000000  6.71670423     0      1  1  0.269232900
#> 260  67  0.00000000  0.22311224     0      1  1  0.446242933
#> 261  68  0.00000000  1.51687570     0      1  1  0.244731018
#> 262  69  0.00000000  0.18673390     1      0  1 -0.598175873
#> 263  69  0.18673390  0.34899649     0      0  1 -0.598175873
#> 264  70  0.00000000  1.15854727     0      1  0  0.231368927
#> 265  71  0.00000000  0.46071326     1      0  0  0.485851788
#> 266  71  0.46071326  1.51947842     1      0  0  0.485851788
#> 267  71  1.51947842  1.56495641     1      0  0  0.485851788
#> 268  71  1.56495641  1.87183690     1      0  0  0.485851788
#> 269  71  1.87183690 39.35334799     0      0  0  0.485851788
#> 270  72  0.00000000  0.02094241     1      0  0 -0.736079482
#> 271  72  0.02094241  0.08672262     1      0  0 -0.736079482
#> 272  72  0.08672262  1.98382511     1      0  0 -0.736079482
#> 273  72  1.98382511  4.24543687     1      0  0 -0.736079482
#> 274  72  4.24543687  5.07665333     1      0  0 -0.736079482
#> 275  72  5.07665333 12.13539233     1      0  0 -0.736079482
#> 276  72 12.13539233 13.34700252     1      0  0 -0.736079482
#> 277  72 13.34700252 13.66708844     1      0  0 -0.736079482
#> 278  72 13.66708844 16.07956783     0      0  0 -0.736079482
#> 279  73  0.00000000  0.90083523     0      1  1  0.264482470
#> 280  74  0.00000000  0.30203451     0      1  0  1.227200176
#> 281  75  0.00000000  0.13850024     1      0  0 -0.824672740
#> 282  75  0.13850024  0.35575232     1      0  0 -0.824672740
#> 283  75  0.35575232  0.78882687     1      0  0 -0.824672740
#> 284  75  0.78882687  0.89604837     1      0  0 -0.824672740
#> 285  75  0.89604837  0.92914305     1      0  0 -0.824672740
#> 286  75  0.92914305  1.15403070     1      0  0 -0.824672740
#> 287  75  1.15403070  1.29915312     1      0  0 -0.824672740
#> 288  75  1.29915312  1.33072868     1      0  0 -0.824672740
#> 289  75  1.33072868  1.70559707     1      0  0 -0.824672740
#> 290  75  1.70559707  1.71682778     1      0  0 -0.824672740
#> 291  75  1.71682778  1.96496754     1      0  0 -0.824672740
#> 292  75  1.96496754  1.98660726     1      0  0 -0.824672740
#> 293  75  1.98660726  2.94479200     1      0  0 -0.824672740
#> 294  75  2.94479200  3.06126113     1      0  0 -0.824672740
#> 295  75  3.06126113  3.37712395     1      0  0 -0.824672740
#> 296  75  3.37712395  3.66929411     1      0  0 -0.824672740
#> 297  75  3.66929411  4.12850098     1      0  0 -0.824672740
#> 298  75  4.12850098  4.22745938     1      0  0 -0.824672740
#> 299  75  4.22745938  4.89096688     0      1  0 -0.824672740
#> 300  76  0.00000000  0.39385840     1      0  1  0.397902578
#> 301  76  0.39385840  6.15053166     0      1  1  0.397902578
#> 302  77  0.00000000  0.09365299     1      0  0 -0.170646583
#> 303  77  0.09365299  1.09119578     1      0  0 -0.170646583
#> 304  77  1.09119578  2.40993384     1      0  0 -0.170646583
#> 305  77  2.40993384  3.51427673     1      0  0 -0.170646583
#> 306  77  3.51427673  4.33947793     1      0  0 -0.170646583
#> 307  77  4.33947793  6.60910482     1      0  0 -0.170646583
#> 308  77  6.60910482  7.71178029     1      0  0 -0.170646583
#> 309  77  7.71178029  9.35732662     1      0  0 -0.170646583
#> 310  77  9.35732662  9.60146405     1      0  0 -0.170646583
#> 311  77  9.60146405 20.19636516     1      0  0 -0.170646583
#> 312  77 20.19636516 23.58836390     1      0  0 -0.170646583
#> 313  77 23.58836390 30.67249041     1      0  0 -0.170646583
#> 314  77 30.67249041 32.63159612     1      0  0 -0.170646583
#> 315  77 32.63159612 44.32515460     1      0  0 -0.170646583
#> 316  77 44.32515460 52.73087987     1      0  0 -0.170646583
#> 317  77 52.73087987 53.05953510     1      0  0 -0.170646583
#> 318  77 53.05953510 60.00000000     0      0  0 -0.170646583
#> 319  78  0.00000000  0.07858750     1      0  0 -0.702626900
#> 320  78  0.07858750  0.40095783     1      0  0 -0.702626900
#> 321  78  0.40095783  0.64362627     1      0  0 -0.702626900
#> 322  78  0.64362627  1.05161616     1      0  0 -0.702626900
#> 323  78  1.05161616  1.09811038     1      0  0 -0.702626900
#> 324  78  1.09811038  1.50948731     0      0  0 -0.702626900
#> 325  79  0.00000000  0.31669278     1      0  1 -0.461480408
#> 326  79  0.31669278  0.57557313     1      0  1 -0.461480408
#> 327  79  0.57557313  0.68087304     1      0  1 -0.461480408
#> 328  79  0.68087304  0.99911935     1      0  1 -0.461480408
#> 329  79  0.99911935 12.78250323     0      1  1 -0.461480408
#> 330  80  0.00000000  0.46109438     1      0  0  0.004949065
#> 331  80  0.46109438  3.25756385     1      0  0  0.004949065
#> 332  80  3.25756385  7.35074220     1      0  0  0.004949065
#> 333  80  7.35074220  9.39342785     1      0  0  0.004949065
#> 334  80  9.39342785 14.42394606     1      0  0  0.004949065
#> 335  80 14.42394606 17.42522135     1      0  0  0.004949065
#> 336  80 17.42522135 22.08724074     1      0  0  0.004949065
#> 337  80 22.08724074 34.78937953     1      0  0  0.004949065
#> 338  80 34.78937953 36.44221633     1      0  0  0.004949065
#> 339  80 36.44221633 55.75291106     1      0  0  0.004949065
#> 340  80 55.75291106 60.00000000     0      0  0  0.004949065
#> 341  81  0.00000000  9.02577005     1      0  1 -0.204429887
#> 342  81  9.02577005 13.21497289     0      1  1 -0.204429887
#> 343  82  0.00000000  0.18149717     1      0  1 -0.694122644
#> 344  82  0.18149717  0.57995248     1      0  1 -0.694122644
#> 345  82  0.57995248  0.73895334     0      1  1 -0.694122644
#> 346  83  0.00000000  0.04422852     0      1  1 -0.132297619
#> 347  84  0.00000000  0.04871240     1      0  0 -0.473649139
#> 348  84  0.04871240  0.06495945     1      0  0 -0.473649139
#> 349  84  0.06495945  0.17149636     1      0  0 -0.473649139
#> 350  84  0.17149636  0.78337725     1      0  0 -0.473649139
#> 351  84  0.78337725  1.32719705     1      0  0 -0.473649139
#> 352  84  1.32719705  2.91829623     1      0  0 -0.473649139
#> 353  84  2.91829623  4.36579096     1      0  0 -0.473649139
#> 354  84  4.36579096  4.48391501     1      0  0 -0.473649139
#> 355  84  4.48391501  5.44933433     1      0  0 -0.473649139
#> 356  84  5.44933433  5.71707436     1      0  0 -0.473649139
#> 357  84  5.71707436 12.18437166     1      0  0 -0.473649139
#> 358  84 12.18437166 14.59368263     1      0  0 -0.473649139
#> 359  84 14.59368263 15.64552977     1      0  0 -0.473649139
#> 360  84 15.64552977 22.47543056     1      0  0 -0.473649139
#> 361  84 22.47543056 27.64150557     1      0  0 -0.473649139
#> 362  84 27.64150557 40.82102725     1      0  0 -0.473649139
#> 363  84 40.82102725 45.45347103     1      0  0 -0.473649139
#> 364  84 45.45347103 46.63442545     1      0  0 -0.473649139
#> 365  84 46.63442545 60.00000000     0      0  0 -0.473649139
#> 366  85  0.00000000  0.48047661     1      0  0  0.369760664
#> 367  85  0.48047661  1.19914847     1      0  0  0.369760664
#> 368  85  1.19914847  3.97802258     1      0  0  0.369760664
#> 369  85  3.97802258 10.11107753     1      0  0  0.369760664
#> 370  85 10.11107753 12.45774093     1      0  0  0.369760664
#> 371  85 12.45774093 15.99120428     0      1  0  0.369760664
#> 372  86  0.00000000  0.16771877     1      0  0  0.448389354
#> 373  86  0.16771877  8.94393114     1      0  0  0.448389354
#> 374  86  8.94393114 15.50336528     1      0  0  0.448389354
#> 375  86 15.50336528 15.51861450     1      0  0  0.448389354
#> 376  86 15.51861450 27.12591882     1      0  0  0.448389354
#> 377  86 27.12591882 31.21858855     1      0  0  0.448389354
#> 378  86 31.21858855 33.41878360     0      0  0  0.448389354
#> 379  87  0.00000000  2.86679606     1      0  0 -0.173000444
#> 380  87  2.86679606  9.83961336     1      0  0 -0.173000444
#> 381  87  9.83961336 23.60365448     0      0  0 -0.173000444
#> 382  88  0.00000000  4.45045302     1      0  1 -0.891028534
#> 383  88  4.45045302 22.89950570     0      0  1 -0.891028534
#> 384  89  0.00000000  0.40282541     0      1  0  0.232470877
#> 385  90  0.00000000  0.12338417     1      0  0 -0.975515305
#> 386  90  0.12338417  0.54509689     1      0  0 -0.975515305
#> 387  90  0.54509689  0.99259467     1      0  0 -0.975515305
#> 388  90  0.99259467  1.03162337     1      0  0 -0.975515305
#> 389  90  1.03162337  1.37625495     1      0  0 -0.975515305
#> 390  90  1.37625495  1.39620311     1      0  0 -0.975515305
#> 391  90  1.39620311  1.49651550     1      0  0 -0.975515305
#> 392  90  1.49651550  1.94472561     1      0  0 -0.975515305
#> 393  90  1.94472561  7.33259398     1      0  0 -0.975515305
#> 394  90  7.33259398  8.51453793     1      0  0 -0.975515305
#> 395  90  8.51453793 11.06627566     1      0  0 -0.975515305
#> 396  90 11.06627566 13.10017134     1      0  0 -0.975515305
#> 397  90 13.10017134 13.59857468     1      0  0 -0.975515305
#> 398  90 13.59857468 14.79794355     1      0  0 -0.975515305
#> 399  90 14.79794355 23.58773471     1      0  0 -0.975515305
#> 400  90 23.58773471 24.64989354     1      0  0 -0.975515305
#> 401  90 24.64989354 31.82882907     1      0  0 -0.975515305
#> 402  90 31.82882907 44.17050896     1      0  0 -0.975515305
#> 403  90 44.17050896 59.93686882     0      0  0 -0.975515305
#> 404  91  0.00000000  0.51218749     1      0  0 -0.258056670
#> 405  91  0.51218749  1.42096228     1      0  0 -0.258056670
#> 406  91  1.42096228  3.65033169     1      0  0 -0.258056670
#> 407  91  3.65033169  4.34800730     1      0  0 -0.258056670
#> 408  91  4.34800730 23.39744948     1      0  0 -0.258056670
#> 409  91 23.39744948 27.55704919     1      0  0 -0.258056670
#> 410  91 27.55704919 55.68663066     0      0  0 -0.258056670
#> 411  92  0.00000000  2.01208192     0      1  0  0.244540159
#> 412  93  0.00000000  0.43448769     0      1  1  0.451012363
#> 413  94  0.00000000 13.01693201     0      1  1  0.320181497
#> 414  95  0.00000000  0.37479471     1      0  0  0.475604439
#> 415  95  0.37479471  1.48531088     1      0  0  0.475604439
#> 416  95  1.48531088  1.64303900     1      0  0  0.475604439
#> 417  95  1.64303900  2.24502438     1      0  0  0.475604439
#> 418  95  2.24502438  4.17605499     1      0  0  0.475604439
#> 419  95  4.17605499 37.84312660     1      0  0  0.475604439
#> 420  95 37.84312660 60.00000000     0      0  0  0.475604439
#> 421  96  0.00000000  0.12223435     0      1  1 -0.299561605
#> 422  97  0.00000000  0.09063569     1      0  0 -0.665347496
#> 423  97  0.09063569  0.32890187     1      0  0 -0.665347496
#> 424  97  0.32890187  0.78996199     1      0  0 -0.665347496
#> 425  97  0.78996199  0.88446804     1      0  0 -0.665347496
#> 426  97  0.88446804  1.69266784     1      0  0 -0.665347496
#> 427  97  1.69266784  3.48190466     1      0  0 -0.665347496
#> 428  97  3.48190466  3.50208921     1      0  0 -0.665347496
#> 429  97  3.50208921  4.16088755     1      0  0 -0.665347496
#> 430  97  4.16088755  5.74684092     1      0  0 -0.665347496
#> 431  97  5.74684092 16.75506562     1      0  0 -0.665347496
#> 432  97 16.75506562 18.95771882     1      0  0 -0.665347496
#> 433  97 18.95771882 46.43449909     1      0  0 -0.665347496
#> 434  97 46.43449909 55.70267324     1      0  0 -0.665347496
#> 435  97 55.70267324 60.00000000     0      0  0 -0.665347496
#> 436  98  0.00000000  0.09775081     1      0  0 -0.296104873
#> 437  98  0.09775081  0.71956989     1      0  0 -0.296104873
#> 438  98  0.71956989  1.40108895     1      0  0 -0.296104873
#> 439  98  1.40108895  1.78064510     1      0  0 -0.296104873
#> 440  98  1.78064510  2.98046872     1      0  0 -0.296104873
#> 441  98  2.98046872  4.18897652     0      1  0 -0.296104873
#> 442  99  0.00000000  0.37126084     0      1  0  0.900525428
#> 443 100  0.00000000  0.30475566     0      1  1  0.527669296