Function to generate simulated recurrent event data

The function simSC() generates simulated recurrent event data from either a Cox-type model, an accelerated mean model, an accelerated rate model, or a scale-change model.

simSC(
  n,
  a1 = a2,
  b1 = b2,
  a2 = a1,
  b2 = b1,
  type = "cox",
  zVar = 0.25,
  tau = 60,
  summary = FALSE
)

Arguments

n
number of observation.

a1, a2, b1, b2
are numeric vectors of length 2. These correspond to the \(\alpha\), \(\beta\), \(\eta\), and \(\theta\) in the joint model. See Details

type
is a character string specifying the underlying model. The rate function type and the hazard function type are separated by a vertical bar "|", with the rate function on the left. For example, type = "cox|am" generates the recurrent process from a Cox model and the terminal event from an accelerated mean model. Setting type = "cox" gives type = "cox|cox".

zVar
a numeric variable specifying the variance of the fraility variable,\(Z\), when zVar > 0. When zVar = 0, \(Z\) is set to a fixed constant 1. The default value is 0.25.

tau
a numeric value specifying the maximum observation time.

summary
a logical value indicating whether a brief data summary will be printed.

Details

The function simSC() generates simulated recurrent event data over the interval \((0, \tau)\) based on the specification of the recurrent process and the terminal events. Specifically, the rate function, \(\lambda(t)\), of the recurrent process can be specified as one of the following model: $$\lambda(t) = Z \lambda_0(te^{X^\top\alpha}) e^{X^\top\beta}, h(t) = Z h_0(te^{X^\top\eta})e^{X^\top\theta}, $$ where \(\lambda_0(t)\) is the baseline rate function, \(h_0(t)\) is the baseline hazard function, \(X\) is a \(n\) by \(p\) covariate matrix and \(\alpha\), \(Z\) is an unobserved shared frailty variable, and \((\alpha, \eta)\) and \((\beta, \theta)\) correspond to the shape and size parameters of the rate function and the hazard function, respectively.

For all scenarios, two covariates are considered; \(X = (X_1, X_2)\), where \(X_1\) follows a Bernoulli distribution with probability 0.5 and \(X_2\) follows a standard normal distribution. The censoring time could be either independent (given covariates) or informative. The simulated data is used for illustration. An informative censoring time, \(C\), is generated separately from an exponential distribution with a rate parameter of 1 / 60 if \(X_1\) is 1, or \(Z^2 / 30\) if \(X_1\) is 0. The observed recurrent events is then observed up to the minimum of \(C\), terminal event, and \(\tau\). Lastly, we assume the baseline functions $$\lambda_0(t) = \frac{2}{1 + t}, h_0(t) = \frac{1}{8(1 + t)}.$$

See also

Examples

set.seed(123)
simSC(100, c(-1, 1), c(-1, 1), summary = TRUE)

#> 
#> Summary results for number of recurrent event per subject:
#>    Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
#>    0.00    1.00    2.00    4.29    6.00   34.00 
#> 
#> Number of failures: 24 (24%); Number of censored events: 76 (76%)
#> 
#> Number of x1 == 1: 41 (41%); Number of x1 == 0: 59 (59%)
#> Summary results for x2:
#>     Min.  1st Qu.   Median     Mean  3rd Qu.     Max. 
#> -3.31068 -0.82464 -0.03736 -0.01454  0.73703  2.70817 
#> 
#> 

#>      id         Time event status x1           x2
#> 1     1 44.426167910     1      0  1  0.314152900
#> 2     1 60.000000000     0      0  1  0.314152900
#> 3     2  0.094921034     1      0  0  1.844838671
#> 4     2  0.125154110     1      0  0  1.844838671
#> 5     2  0.128363300     1      0  0  1.844838671
#> 6     2  0.128893639     1      0  0  1.844838671
#> 7     2  0.151798159     1      0  0  1.844838671
#> 8     2  0.156259693     1      0  0  1.844838671
#> 9     2  0.162467082     1      0  0  1.844838671
#> 10    2  0.192499996     1      0  0  1.844838671
#> 11    2  0.205361520     1      0  0  1.844838671
#> 12    2  0.212755086     0      1  0  1.844838671
#> 13    3 60.000000000     0      0  1 -0.981917149
#> 14    4  0.056292351     1      0  1  2.196003763
#> 15    4  0.072279983     1      0  1  2.196003763
#> 16    4  0.185730262     1      0  1  2.196003763
#> 17    4  0.289228289     1      0  1  2.196003763
#> 18    4  1.361980416     1      0  1  2.196003763
#> 19    4  1.874695287     1      0  1  2.196003763
#> 20    4  3.977850859     1      0  1  2.196003763
#> 21    4  4.480471395     1      0  1  2.196003763
#> 22    4  4.823564663     1      0  1  2.196003763
#> 23    4  4.982837852     1      0  1  2.196003763
#> 24    4  5.317014620     1      0  1  2.196003763
#> 25    4  5.340135942     1      0  1  2.196003763
#> 26    4  5.901524918     1      0  1  2.196003763
#> 27    4  5.956915706     1      0  1  2.196003763
#> 28    4  8.325919518     0      0  1  2.196003763
#> 29    5  0.129123255     1      0  0 -0.204667665
#> 30    5  0.693168398     1      0  0 -0.204667665
#> 31    5  1.332298107     1      0  0 -0.204667665
#> 32    5  1.858225226     1      0  0 -0.204667665
#> 33    5  3.180539355     0      0  0 -0.204667665
#> 34    6  0.216686503     1      0  0  0.975142936
#> 35    6  0.319404774     1      0  0  0.975142936
#> 36    6  0.588142311     1      0  0  0.975142936
#> 37    6  0.618091546     1      0  0  0.975142936
#> 38    6  1.085061481     1      0  0  0.975142936
#> 39    6  1.298248152     1      0  0  0.975142936
#> 40    6  1.354466114     1      0  0  0.975142936
#> 41    6  1.433272512     1      0  0  0.975142936
#> 42    6  1.479426384     1      0  0  0.975142936
#> 43    6  1.486390825     1      0  0  0.975142936
#> 44    6  1.603648039     1      0  0  0.975142936
#> 45    6  3.036196533     1      0  0  0.975142936
#> 46    6  4.378587285     1      0  0  0.975142936
#> 47    6  5.733083522     1      0  0  0.975142936
#> 48    6  7.070218220     1      0  0  0.975142936
#> 49    6  8.161021327     0      0  0  0.975142936
#> 50    7 12.204839610     0      0  0 -0.867566116
#> 51    8  0.480456759     1      0  0 -0.501187592
#> 52    8  9.182117175     0      0  0 -0.501187592
#> 53    9  0.004517227     1      0  0  0.785591163
#> 54    9  0.010845527     1      0  0  0.785591163
#> 55    9  0.231110693     1      0  0  0.785591163
#> 56    9  0.428345024     1      0  0  0.785591163
#> 57    9  0.481381589     1      0  0  0.785591163
#> 58    9  0.850984615     1      0  0  0.785591163
#> 59    9  0.957997791     1      0  0  0.785591163
#> 60    9  1.796847636     1      0  0  0.785591163
#> 61    9  2.040113204     1      0  0  0.785591163
#> 62    9  3.912920907     1      0  0  0.785591163
#> 63    9  4.187313729     0      1  0  0.785591163
#> 64   10 60.000000000     0      0  1 -2.102247324
#> 65   11  0.493291422     1      0  0 -0.042204930
#> 66   11  0.716490115     1      0  0 -0.042204930
#> 67   11  0.720544178     1      0  0 -0.042204930
#> 68   11  2.348637230     1      0  0 -0.042204930
#> 69   11  3.635269917     0      0  0 -0.042204930
#> 70   12  0.747892583     1      0  1 -0.404809413
#> 71   12  1.156108466     1      0  1 -0.404809413
#> 72   12 18.031380046     0      0  1 -0.404809413
#> 73   13 18.452001885     0      0  1 -0.112765966
#> 74   14  0.113047048     1      0  1  1.797140441
#> 75   14  0.136108410     1      0  1  1.797140441
#> 76   14  0.365452359     1      0  1  1.797140441
#> 77   14  0.391098229     1      0  1  1.797140441
#> 78   14  0.527733393     1      0  1  1.797140441
#> 79   14  0.590708249     1      0  1  1.797140441
#> 80   14  0.902421315     1      0  1  1.797140441
#> 81   14  1.019193212     1      0  1  1.797140441
#> 82   14  1.787599159     1      0  1  1.797140441
#> 83   14  2.533196580     1      0  1  1.797140441
#> 84   14  2.703503861     1      0  1  1.797140441
#> 85   14  3.003116132     1      0  1  1.797140441
#> 86   14  3.533033566     1      0  1  1.797140441
#> 87   14  5.135138577     1      0  1  1.797140441
#> 88   14  6.001520455     1      0  1  1.797140441
#> 89   14  8.129608877     1      0  1  1.797140441
#> 90   14  8.147785928     1      0  1  1.797140441
#> 91   14 14.463682858     1      0  1  1.797140441
#> 92   14 15.714660929     1      0  1  1.797140441
#> 93   14 16.000543655     1      0  1  1.797140441
#> 94   14 16.698048789     1      0  1  1.797140441
#> 95   14 19.074894239     1      0  1  1.797140441
#> 96   14 20.816736023     1      0  1  1.797140441
#> 97   14 21.072708316     0      1  1  1.797140441
#> 98   15  1.995573195     1      0  0 -0.810326258
#> 99   15  4.124126960     1      0  0 -0.810326258
#> 100  15 10.801766722     1      0  0 -0.810326258
#> 101  15 11.024803256     0      0  0 -0.810326258
#> 102  16  0.234114367     1      0  1  1.900900074
#> 103  16  0.851754751     1      0  1  1.900900074
#> 104  16  1.606035952     1      0  1  1.900900074
#> 105  16  1.762289385     1      0  1  1.900900074
#> 106  16  1.826441917     1      0  1  1.900900074
#> 107  16  3.124673952     1      0  1  1.900900074
#> 108  16  3.243853424     1      0  1  1.900900074
#> 109  16  4.395090357     1      0  1  1.900900074
#> 110  16  4.992642076     1      0  1  1.900900074
#> 111  16  5.642810119     0      0  1  1.900900074
#> 112  17  1.327927087     1      0  0  0.708954440
#> 113  17  2.701109679     1      0  0  0.708954440
#> 114  17 20.995077410     1      0  0  0.708954440
#> 115  17 25.360391444     1      0  0  0.708954440
#> 116  17 25.797039543     0      0  0  0.708954440
#> 117  18  0.367863189     1      0  1  0.736194775
#> 118  18  1.671345267     1      0  1  0.736194775
#> 119  18  2.488031575     1      0  1  0.736194775
#> 120  18 10.495364042     0      0  1  0.736194775
#> 121  19  0.003664993     0      1  0  1.365776554
#> 122  20  1.659010263     1      0  0 -0.576263893
#> 123  20  8.270060779     1      0  0 -0.576263893
#> 124  20 57.861444747     0      0  0 -0.576263893
#> 125  21  5.085302762     1      0  0 -0.804732320
#> 126  21  8.368098177     1      0  0 -0.804732320
#> 127  21 60.000000000     0      0  0 -0.804732320
#> 128  22  1.229565771     1      0  0 -0.535065812
#> 129  22  4.557610284     1      0  0 -0.535065812
#> 130  22  8.449352122     1      0  0 -0.535065812
#> 131  22 23.274458489     0      0  0 -0.535065812
#> 132  23  6.281136752     1      0  0  0.791547909
#> 133  23 13.303086271     1      0  0  0.791547909
#> 134  23 35.697385624     1      0  0  0.791547909
#> 135  23 60.000000000     0      0  0  0.791547909
#> 136  24  2.452900450     0      1  0 -0.707607151
#> 137  25  6.293977611     0      0  1 -1.275543372
#> 138  26  0.045506337     1      0  0  2.375464461
#> 139  26  0.178439471     1      0  0  2.375464461
#> 140  26  0.229764689     1      0  0  2.375464461
#> 141  26  0.391617573     1      0  0  2.375464461
#> 142  26  0.433009312     1      0  0  2.375464461
#> 143  26  0.444372840     1      0  0  2.375464461
#> 144  26  0.494782288     1      0  0  2.375464461
#> 145  26  0.534103344     1      0  0  2.375464461
#> 146  26  0.709791714     1      0  0  2.375464461
#> 147  26  0.715533114     1      0  0  2.375464461
#> 148  26  0.734933214     1      0  0  2.375464461
#> 149  26  0.780807122     1      0  0  2.375464461
#> 150  26  1.022881411     1      0  0  2.375464461
#> 151  26  1.099555227     1      0  0  2.375464461
#> 152  26  1.210136244     1      0  0  2.375464461
#> 153  26  1.416993677     1      0  0  2.375464461
#> 154  26  1.435961365     1      0  0  2.375464461
#> 155  26  1.730702372     1      0  0  2.375464461
#> 156  26  2.048041744     1      0  0  2.375464461
#> 157  26  2.124960370     1      0  0  2.375464461
#> 158  26  2.298323196     1      0  0  2.375464461
#> 159  26  2.591394978     1      0  0  2.375464461
#> 160  26  3.301813924     1      0  0  2.375464461
#> 161  26  3.947651924     1      0  0  2.375464461
#> 162  26  4.131293060     1      0  0  2.375464461
#> 163  26  4.251416452     1      0  0  2.375464461
#> 164  26  4.292132687     1      0  0  2.375464461
#> 165  26  4.427350682     1      0  0  2.375464461
#> 166  26  4.585815077     1      0  0  2.375464461
#> 167  26  4.987689301     1      0  0  2.375464461
#> 168  26  5.283218406     1      0  0  2.375464461
#> 169  26  5.305954276     1      0  0  2.375464461
#> 170  26  5.574535236     1      0  0  2.375464461
#> 171  26  5.827979402     1      0  0  2.375464461
#> 172  26  6.318148700     0      1  0  2.375464461
#> 173  27  2.527861668     0      0  1 -1.093589748
#> 174  28  0.051920460     1      0  0  0.192441582
#> 175  28  0.144681734     1      0  0  0.192441582
#> 176  28  0.247917295     1      0  0  0.192441582
#> 177  28  0.750263763     1      0  0  0.192441582
#> 178  28  1.221031857     0      0  0  0.192441582
#> 179  29  0.226611928     1      0  0 -0.126160381
#> 180  29  0.377627929     0      0  0 -0.126160381
#> 181  30 43.694549172     0      0  0 -1.388432140
#> 182  31  0.045878423     1      0  0  0.469902192
#> 183  31  0.512235913     1      0  0  0.469902192
#> 184  31  0.646405086     0      0  0  0.469902192
#> 185  32  1.760455139     1      0  1  0.960382357
#> 186  32 22.378842132     0      0  1  0.960382357
#> 187  33  0.951569404     1      0  1 -0.005141288
#> 188  33  7.505540052     1      0  1 -0.005141288
#> 189  33 60.000000000     0      0  1 -0.005141288
#> 190  34  0.515407932     1      0  0 -0.870464442
#> 191  34  0.896151084     1      0  0 -0.870464442
#> 192  34  0.931286070     0      1  0 -0.870464442
#> 193  35  0.944737736     1      0  0 -0.462879040
#> 194  35  6.428637348     1      0  0 -0.462879040
#> 195  35  7.544621338     1      0  0 -0.462879040
#> 196  35 19.917613942     0      0  0 -0.462879040
#> 197  36 60.000000000     0      0  0 -1.911151255
#> 198  37  0.367580507     1      0  1  0.369816044
#> 199  37  0.484168162     1      0  1  0.369816044
#> 200  37  3.400619282     1      0  1  0.369816044
#> 201  37  6.287594223     1      0  1  0.369816044
#> 202  37  9.211417503     1      0  1  0.369816044
#> 203  37 11.041441969     1      0  1  0.369816044
#> 204  37 60.000000000     0      0  1  0.369816044
#> 205  38 60.000000000     0      0  1 -0.462288057
#> 206  39  8.910640641     1      0  1  0.321955892
#> 207  39  9.839924491     1      0  1  0.321955892
#> 208  39 15.569450855     0      0  1  0.321955892
#> 209  40  0.389491429     1      0  0 -0.284001390
#> 210  40  1.121575864     1      0  0 -0.284001390
#> 211  40  3.341656791     1      0  0 -0.284001390
#> 212  40  6.553368190     1      0  0 -0.284001390
#> 213  40  7.197774398     1      0  0 -0.284001390
#> 214  40 29.655227938     1      0  0 -0.284001390
#> 215  40 31.327400120     1      0  0 -0.284001390
#> 216  40 60.000000000     0      0  0 -0.284001390
#> 217  41  0.380419607     1      0  1  1.091856443
#> 218  41  2.947484754     1      0  1  1.091856443
#> 219  41  3.485678620     1      0  1  1.091856443
#> 220  41 12.721763638     0      1  1  1.091856443
#> 221  42  2.647586860     1      0  0  0.160216363
#> 222  42  4.222113713     1      0  0  0.160216363
#> 223  42 11.930325739     0      0  0  0.160216363
#> 224  43  0.060869045     1      0  0  2.708168284
#> 225  43  0.141328881     1      0  0  2.708168284
#> 226  43  0.156319137     1      0  0  2.708168284
#> 227  43  0.283256048     1      0  0  2.708168284
#> 228  43  0.374894071     1      0  0  2.708168284
#> 229  43  0.632541474     1      0  0  2.708168284
#> 230  43  0.754572416     1      0  0  2.708168284
#> 231  43  0.787107817     1      0  0  2.708168284
#> 232  43  0.907544674     1      0  0  2.708168284
#> 233  43  0.927490095     1      0  0  2.708168284
#> 234  43  1.062092047     0      1  0  2.708168284
#> 235  44  9.902336551     1      0  0  0.286289929
#> 236  44 10.162711700     0      0  0  0.286289929
#> 237  45  1.130904205     1      0  0 -0.032510251
#> 238  45  1.184472075     1      0  0 -0.032510251
#> 239  45  1.555413216     1      0  0 -0.032510251
#> 240  45 60.000000000     0      0  0 -0.032510251
#> 241  46  7.260021570     1      0  0 -0.989678019
#> 242  46  7.935806037     1      0  0 -0.989678019
#> 243  46 10.324601177     1      0  0 -0.989678019
#> 244  46 37.580687785     0      0  0 -0.989678019
#> 245  47  4.021075863     1      0  1  0.211809928
#> 246  47  6.647520104     1      0  1  0.211809928
#> 247  47 60.000000000     0      0  1  0.211809928
#> 248  48  2.798082020     1      0  1 -0.972017534
#> 249  48  4.302783453     1      0  1 -0.972017534
#> 250  48 60.000000000     0      0  1 -0.972017534
#> 251  49  0.176975346     1      0  1  0.549224014
#> 252  49  0.743075179     1      0  1  0.549224014
#> 253  49  0.813259869     1      0  1  0.549224014
#> 254  49  1.153431976     1      0  1  0.549224014
#> 255  49  1.354066606     1      0  1  0.549224014
#> 256  49  2.304182033     1      0  1  0.549224014
#> 257  49  4.215737676     1      0  1  0.549224014
#> 258  49  9.764389491     1      0  1  0.549224014
#> 259  49 11.533429121     1      0  1  0.549224014
#> 260  49 12.113327177     1      0  1  0.549224014
#> 261  49 16.493736204     1      0  1  0.549224014
#> 262  49 19.744157497     1      0  1  0.549224014
#> 263  49 20.497090013     1      0  1  0.549224014
#> 264  49 30.745422361     1      0  1  0.549224014
#> 265  49 31.091008194     1      0  1  0.549224014
#> 266  49 34.601702514     0      1  1  0.549224014
#> 267  50  0.230084990     1      0  0 -1.049021530
#> 268  50 48.263490207     0      0  0 -1.049021530
#> 269  51  0.067797362     1      0  0  0.449133935
#> 270  51  0.264214934     1      0  0  0.449133935
#> 271  51  0.462465771     1      0  0  0.449133935
#> 272  51  0.806881110     1      0  0  0.449133935
#> 273  51  0.886435166     1      0  0  0.449133935
#> 274  51  5.881111896     1      0  0  0.449133935
#> 275  51  8.606163806     1      0  0  0.449133935
#> 276  51  8.942277607     1      0  0  0.449133935
#> 277  51 14.864964116     0      0  0  0.449133935
#> 278  52 13.784086765     1      0  1 -0.385855653
#> 279  52 27.970609279     1      0  1 -0.385855653
#> 280  52 41.380154919     0      0  1 -0.385855653
#> 281  53 45.917339621     0      0  1 -1.570135523
#> 282  54  0.132881709     1      0  0  1.226942601
#> 283  54  0.220699343     1      0  0  1.226942601
#> 284  54  0.428962098     1      0  0  1.226942601
#> 285  54  0.517601886     1      0  0  1.226942601
#> 286  54  0.638980442     1      0  0  1.226942601
#> 287  54  0.656049954     1      0  0  1.226942601
#> 288  54  1.253642173     1      0  0  1.226942601
#> 289  54  1.407435662     1      0  0  1.226942601
#> 290  54  1.694336288     1      0  0  1.226942601
#> 291  54  1.767217579     1      0  0  1.226942601
#> 292  54  2.159906265     1      0  0  1.226942601
#> 293  54  2.408250542     1      0  0  1.226942601
#> 294  54  2.417505832     1      0  0  1.226942601
#> 295  54  2.796181193     1      0  0  1.226942601
#> 296  54  2.863436992     1      0  0  1.226942601
#> 297  54  2.962453010     1      0  0  1.226942601
#> 298  54  3.324677789     1      0  0  1.226942601
#> 299  54  3.562060613     1      0  0  1.226942601
#> 300  54  3.913280487     1      0  0  1.226942601
#> 301  54  3.950248888     1      0  0  1.226942601
#> 302  54  4.455959171     1      0  0  1.226942601
#> 303  54  4.989831279     1      0  0  1.226942601
#> 304  54  5.192937332     1      0  0  1.226942601
#> 305  54  5.329145827     1      0  0  1.226942601
#> 306  54  5.378969353     1      0  0  1.226942601
#> 307  54  5.514449219     1      0  0  1.226942601
#> 308  54  9.091785237     1      0  0  1.226942601
#> 309  54 16.474294053     1      0  0  1.226942601
#> 310  54 19.403666424     1      0  0  1.226942601
#> 311  54 30.590409252     1      0  0  1.226942601
#> 312  54 30.962417132     0      0  0  1.226942601
#> 313  55  0.317940276     1      0  1  0.666694396
#> 314  55  1.168614638     1      0  1  0.666694396
#> 315  55  1.713942123     1      0  1  0.666694396
#> 316  55  2.607341604     1      0  1  0.666694396
#> 317  55  5.271562205     1      0  1  0.666694396
#> 318  55  5.639152530     1      0  1  0.666694396
#> 319  55 10.220013936     1      0  1  0.666694396
#> 320  55 19.553747046     1      0  1  0.666694396
#> 321  55 28.152110925     1      0  1  0.666694396
#> 322  55 38.274645992     0      0  1  0.666694396
#> 323  56 22.337666746     1      0  0 -1.151603163
#> 324  56 39.205379870     1      0  0 -1.151603163
#> 325  56 60.000000000     0      0  0 -1.151603163
#> 326  57  1.123956508     1      0  0 -0.160788483
#> 327  57  7.478829245     1      0  0 -0.160788483
#> 328  57  7.608661820     1      0  0 -0.160788483
#> 329  57 22.129391233     0      0  0 -0.160788483
#> 330  58 35.858826935     0      0  1 -0.972913776
#> 331  59  0.106334237     1      0  0 -0.610461659
#> 332  59  0.143066590     1      0  0 -0.610461659
#> 333  59  0.399736278     1      0  0 -0.610461659
#> 334  59  1.329412686     0      0  0 -0.610461659
#> 335  60  0.289244137     1      0  0  0.499276876
#> 336  60  1.119889409     1      0  0  0.499276876
#> 337  60  1.861822395     1      0  0  0.499276876
#> 338  60  1.921929449     1      0  0  0.499276876
#> 339  60  2.411083779     1      0  0  0.499276876
#> 340  60  2.546129412     1      0  0  0.499276876
#> 341  60  5.750243466     1      0  0  0.499276876
#> 342  60  6.923099549     1      0  0  0.499276876
#> 343  60  7.432479386     1      0  0  0.499276876
#> 344  60  9.919762625     0      0  0  0.499276876
#> 345  61  0.329834180     1      0  0  0.879364139
#> 346  61  1.287722228     1      0  0  0.879364139
#> 347  61  1.659425957     1      0  0  0.879364139
#> 348  61  1.756849354     1      0  0  0.879364139
#> 349  61  4.536289698     1      0  0  0.879364139
#> 350  61  7.866204392     1      0  0  0.879364139
#> 351  61 11.104906100     1      0  0  0.879364139
#> 352  61 15.570550002     1      0  0  0.879364139
#> 353  61 30.687902796     1      0  0  0.879364139
#> 354  61 58.924853849     1      0  0  0.879364139
#> 355  61 60.000000000     0      0  0  0.879364139
#> 356  62  0.423540384     1      0  1  0.923129239
#> 357  62  5.744994749     1      0  1  0.923129239
#> 358  62  7.426180952     1      0  1  0.923129239
#> 359  62 60.000000000     0      0  1  0.923129239
#> 360  63  0.233545879     1      0  0  0.693133972
#> 361  63  0.234017989     1      0  0  0.693133972
#> 362  63  0.538450199     1      0  0  0.693133972
#> 363  63  0.627526061     1      0  0  0.693133972
#> 364  63  0.861556521     1      0  0  0.693133972
#> 365  63  1.102409116     1      0  0  0.693133972
#> 366  63  3.258602170     1      0  0  0.693133972
#> 367  63  6.293075161     1      0  0  0.693133972
#> 368  63 14.284947113     1      0  0  0.693133972
#> 369  63 17.246547226     1      0  0  0.693133972
#> 370  63 30.399024532     1      0  0  0.693133972
#> 371  63 33.538318472     0      1  0  0.693133972
#> 372  64  0.150317201     1      0  1 -0.139500325
#> 373  64  4.096807626     1      0  1 -0.139500325
#> 374  64 13.747458877     0      0  1 -0.139500325
#> 375  65 30.085791836     0      0  0 -3.310678727
#> 376  66  5.254560647     1      0  1  0.538465799
#> 377  66 19.785907300     0      0  1  0.538465799
#> 378  67  0.042599725     1      0  1  0.892485867
#> 379  67  0.460391154     1      0  1  0.892485867
#> 380  67  0.486884850     1      0  1  0.892485867
#> 381  67  0.623984366     1      0  1  0.892485867
#> 382  67  1.316361786     1      0  1  0.892485867
#> 383  67 16.813802408     0      1  1  0.892485867
#> 384  68  0.353305779     1      0  1  0.489462036
#> 385  68  1.326752035     1      0  1  0.489462036
#> 386  68  2.319360850     1      0  1  0.489462036
#> 387  68  2.344870833     0      0  1  0.489462036
#> 388  69  0.348996492     0      0  1 -1.196351747
#> 389  70  1.148963316     1      0  0  0.462737853
#> 390  70  1.312017315     1      0  0  0.462737853
#> 391  70  2.542578920     1      0  0  0.462737853
#> 392  70  5.151759989     1      0  0  0.462737853
#> 393  70 11.690276404     1      0  0  0.462737853
#> 394  70 60.000000000     0      0  0  0.462737853
#> 395  71  0.266150966     1      0  0  0.971703576
#> 396  71  0.349670849     0      1  0  0.971703576
#> 397  72 42.047029902     0      0  0 -1.472158964
#> 398  73  0.393912335     1      0  1  0.528964941
#> 399  73 50.322384926     1      0  1  0.528964941
#> 400  73 60.000000000     0      0  1  0.528964941
#> 401  74  0.226227290     0      1  0  2.454400353
#> 402  75  0.604137764     1      0  0 -1.649345480
#> 403  75  1.302897180     0      0  0 -1.649345480
#> 404  76  0.297104808     1      0  1  0.795805155
#> 405  76  1.442722956     1      0  1  0.795805155
#> 406  76  6.228937470     1      0  1  0.795805155
#> 407  76 33.654580313     1      0  1  0.795805155
#> 408  76 60.000000000     0      0  1  0.795805155
#> 409  77  0.247724744     1      0  0 -0.341293167
#> 410  77  0.883275948     1      0  0 -0.341293167
#> 411  77  0.965598205     1      0  0 -0.341293167
#> 412  77  1.267030058     1      0  0 -0.341293167
#> 413  77  2.136561426     1      0  0 -0.341293167
#> 414  77 11.737178714     1      0  0 -0.341293167
#> 415  77 19.006126729     1      0  0 -0.341293167
#> 416  77 19.508749921     1      0  0 -0.341293167
#> 417  77 33.068292801     1      0  0 -0.341293167
#> 418  77 37.221890340     0      0  0 -0.341293167
#> 419  78  1.634828342     1      0  0 -1.405253800
#> 420  78  2.166839239     0      0  0 -1.405253800
#> 421  79 18.849900991     0      1  1 -0.922960815
#> 422  80  0.057899323     1      0  0  0.009898130
#> 423  80  0.563149307     1      0  0  0.009898130
#> 424  80  1.729418601     1      0  0  0.009898130
#> 425  80  1.742868436     1      0  0  0.009898130
#> 426  80  2.184840182     1      0  0  0.009898130
#> 427  80  3.269702071     1      0  0  0.009898130
#> 428  80  6.823633670     0      0  0  0.009898130
#> 429  81  2.088983908     1      0  1 -0.408859774
#> 430  81 60.000000000     0      0  1 -0.408859774
#> 431  82 10.473374901     0      0  1 -1.388245289
#> 432  83  4.738025556     1      0  1 -0.264595238
#> 433  83 10.707105999     1      0  1 -0.264595238
#> 434  83 23.299277838     1      0  1 -0.264595238
#> 435  83 36.803331552     0      0  1 -0.264595238
#> 436  84  3.101080222     1      0  0 -0.947298279
#> 437  84  6.626583844     1      0  0 -0.947298279
#> 438  84 31.520187178     0      1  0 -0.947298279
#> 439  85  0.644340177     1      0  0  0.739521328
#> 440  85  0.710910608     1      0  0  0.739521328
#> 441  85  1.536251400     1      0  0  0.739521328
#> 442  85  1.816507927     1      0  0  0.739521328
#> 443  85  2.401143401     1      0  0  0.739521328
#> 444  85  2.613061991     1      0  0  0.739521328
#> 445  85  3.121227252     0      1  0  0.739521328
#> 446  86  0.016017404     1      0  0  0.896778709
#> 447  86  0.232698496     1      0  0  0.896778709
#> 448  86  0.577077685     1      0  0  0.896778709
#> 449  86  0.978526520     1      0  0  0.896778709
#> 450  86  1.057186346     1      0  0  0.896778709
#> 451  86  1.462097773     0      1  0  0.896778709
#> 452  87  0.895563396     1      0  0 -0.346000887
#> 453  87 33.465054977     0      0  0 -0.346000887
#> 454  88 22.899505699     0      0  1 -1.782057068
#> 455  89  0.037101618     1      0  0  0.464941753
#> 456  89  0.102710022     1      0  0  0.464941753
#> 457  89  0.216567111     1      0  0  0.464941753
#> 458  89  0.796068242     1      0  0  0.464941753
#> 459  89  0.958993995     1      0  0  0.464941753
#> 460  89  1.028313254     1      0  0  0.464941753
#> 461  89  1.074377107     0      0  0  0.464941753
#> 462  90 60.000000000     0      0  0 -1.951030610
#> 463  91  0.078056200     0      1  0 -0.516113340
#> 464  92  0.176685484     1      0  0  0.489080318
#> 465  92  0.296731331     1      0  0  0.489080318
#> 466  92  0.538848097     1      0  0  0.489080318
#> 467  92  0.571099060     1      0  0  0.489080318
#> 468  92  0.998071193     1      0  0  0.489080318
#> 469  92  1.283195611     1      0  0  0.489080318
#> 470  92  2.193886966     1      0  0  0.489080318
#> 471  92  3.611732953     1      0  0  0.489080318
#> 472  92  7.750845917     0      0  0  0.489080318
#> 473  93  0.025896790     1      0  1  0.902024726
#> 474  93  2.654534720     1      0  1  0.902024726
#> 475  93  2.711107298     1      0  1  0.902024726
#> 476  93  3.396050813     1      0  1  0.902024726
#> 477  93  4.747309685     0      0  1  0.902024726
#> 478  94  3.672871226     1      0  1  0.640362995
#> 479  94  5.452322593     1      0  1  0.640362995
#> 480  94 16.537795976     1      0  1  0.640362995
#> 481  94 21.891690916     1      0  1  0.640362995
#> 482  94 25.337334766     1      0  1  0.640362995
#> 483  94 60.000000000     0      0  1  0.640362995
#> 484  95  0.070926623     1      0  0  0.951208879
#> 485  95  0.423656815     1      0  0  0.951208879
#> 486  95  0.487280034     1      0  0  0.951208879
#> 487  95  0.496622564     1      0  0  0.951208879
#> 488  95  0.553466751     1      0  0  0.951208879
#> 489  95  0.605335217     1      0  0  0.951208879
#> 490  95  0.803361018     1      0  0  0.951208879
#> 491  95  0.965024233     0      1  0  0.951208879
#> 492  96  3.769555449     0      1  1 -0.599123210
#> 493  97  1.274006209     0      1  0 -1.330694993
#> 494  98 11.432728756     1      0  0 -0.592209746
#> 495  98 11.998121025     1      0  0 -0.592209746
#> 496  98 14.377420871     1      0  0 -0.592209746
#> 497  98 22.412248553     1      0  0 -0.592209746
#> 498  98 23.615610800     1      0  0 -0.592209746
#> 499  98 45.509140172     1      0  0 -0.592209746
#> 500  98 59.117118140     0      1  0 -0.592209746
#> 501  99  0.024035847     1      0  0  1.801050856
#> 502  99  0.169182346     1      0  0  1.801050856
#> 503  99  0.204678212     1      0  0  1.801050856
#> 504  99  0.429791121     1      0  0  1.801050856
#> 505  99  0.564102773     1      0  0  1.801050856
#> 506  99  0.612301930     1      0  0  1.801050856
#> 507  99  0.920473254     1      0  0  1.801050856
#> 508  99  1.029332820     1      0  0  1.801050856
#> 509  99  1.351443755     1      0  0  1.801050856
#> 510  99  1.689399361     0      1  0  1.801050856
#> 511 100  0.062282158     1      0  1  1.055338591
#> 512 100  0.092470102     1      0  1  1.055338591
#> 513 100  0.099136473     1      0  1  1.055338591
#> 514 100  0.119418775     1      0  1  1.055338591
#> 515 100  0.494375485     1      0  1  1.055338591
#> 516 100  0.535604140     1      0  1  1.055338591
#> 517 100  0.928665456     1      0  1  1.055338591
#> 518 100  2.430407324     1      0  1  1.055338591
#> 519 100  2.924293729     1      0  1  1.055338591
#> 520 100  5.242506460     1      0  1  1.055338591
#> 521 100  5.727066391     1      0  1  1.055338591
#> 522 100  7.211164372     1      0  1  1.055338591
#> 523 100  8.749865224     1      0  1  1.055338591
#> 524 100  9.247636836     1      0  1  1.055338591
#> 525 100 10.889850475     1      0  1  1.055338591
#> 526 100 11.144609855     1      0  1  1.055338591
#> 527 100 11.155463685     1      0  1  1.055338591
#> 528 100 42.939965866     1      0  1  1.055338591
#> 529 100 44.020624512     0      0  1  1.055338591