Plot options for plotEvents

This function provides the plotting options for the plotEvents() function.

plotEvents.control(
  xlab = NULL,
  ylab = NULL,
  main = NULL,
  terminal.name = NULL,
  recurrent.name = NULL,
  recurrent.type = NULL,
  legend.position = "top",
  base_size = 12,
  cex = NULL,
  alpha = 0.7
)

Arguments

xlab
a character string indicating the label for the x axis. The default value is "Time".

ylab
a character string indicating the label for the y axis. The default value is "Subject".

main
a character string indicating the title of the plot.

terminal.name
a character string indicating the label for the terminal event displayed in the legend. The default value is "Terminal event".

recurrent.name
a character string indicating the label for the recurrent event displayed in the legend. The default value is "Recurrent events".

recurrent.type
a factor indicating the labels for the different recurrent event types. This option is only available when there are more than one types of recurrent events. The default value is "Recurrent events 1", "Recurrent events 2", ....

legend.position
a character string specifies the position of the legend. The available options are "none", "left", "right", "bottom", "top", or a two-element numeric vector specifies the coordinate of the legend. This argument is passed to the ggplot theme environment. The default value is "top".

base_size
a numerical value to specify the base font size, given in pts. This argument is passed to the ggplot theme environment. The default value is 12.

cex
a numerical value specifies the size of the points.

alpha
a numerical value specifies the transparency of the points.

See also