Fits Semiparametric Regression Models for Recurrent Event Data

Fits a semiparametric regression model for the recurrent event data. The rate function of the underlying process for the recurrent event process can be specified as a Cox-type model, an accelerated mean model, or a generalized scale-change model. See details for model specifications.

reReg(formula, data, B = 200, method = c("cox.LWYY", "cox.GL",
  "cox.HW", "am.GL", "am.XCHWY", "sc.XCYH"), se = c("NULL", "bootstrap",
  "resampling"), contrasts = NULL, control = list())

Arguments

formula
a formula object, with the response on the left of a "~" operator, and the predictors on the right. The response must be a recurrent event survival object as returned by function reSurv.

data
an optional data frame in which to interpret the variables occurring in the "formula".

B
a numeric value specifies the number of resampling for variance estimation. When B = 0, variance estimation will not be performed.

method
a character string specifying the underlying model. See Details.

se
a character string specifying the method for standard error estimation. See Details.

contrasts
an optional list.

control
a list of control parameters.

Details

Suppose the recurrent event process and the failure events are observed in the time interval \(t\in[0,\tau]\), for some constant \(\tau\). We formulate the rate function, \(\lambda(t)\), for the recurrent event process and the hazard function, \(h(t)\), for the censoring time under the following model specifications:

Cox-type model:

$$\lambda(t) = Z \lambda_0(t) e^{X^\top\alpha}, h(t) = Z h_0(t)e^{X^\top\beta},$$

Accelerated mean model:

$$\lambda(t) = Z \lambda_0(te^{X^\top\alpha})e^{X^\top\alpha}, h(t) = Z h_0(te^{X^\top\beta})e^{X^\top\beta},$$

Scale-change model:

$$\lambda(t) = Z \lambda_0(te^{X^\top\alpha})e^{X^\top\beta},$$

where \(\lambda_0(t)\) is the baseline rate function, \(h_0(t)\) is the baseline hazard function, \(X\) is a \(n\) by \(p\) covariate matrix and \(\alpha\), \(Z\) is an unobserved shared frailty variable, and \(\beta\) are unknown \(p\)-dimensional regression parameters.

The reReg function fits models with the following available methods:

method = "cox.LWYY"

assumes the Cox-type model with Z = 1 and requires independent censoring. The returned result is equivalent to that from coxph. See reference Lin et al. (2000).

method = "cox.HW"

assumes the Cox-type model with unspecified Z, thus accommodate informative censoring. See the references See reference Wang, Qin and Chiang (2001) and Huang and Wang (2004).

method = "am.GL"

assumes the accelerated mean model with Z = 1 and requires independent censoring. See the reference Ghosh and Lin (2003).

method = "am.XCHWY"

assumes the accelerated mean model with unspecified Z, thus accommodate informative censoring. See the reference Xu et al. (2017).

method = "sc.XCYH"

assumes the generalized scale-change model, and includes the methods "cox.HW" and "am.XCHWY" as special cases. Informative censoring is accounted for through the unspecified frailty variable Z. The methods also provide a hypothesis test of these submodels.

The available methods for variance estimation are:

NULL

variance estimation will not be performed. This is equivalent to setting B = 0.

"resampling"

performs the efficient resampling-based sandwich estimator that works with methods "cox.HW", "am.XCHWY" and "sc.XCYH".

"bootstrap"

works with all fitting methods.

The control list consists of the following parameters:

tol

absolute error tolerance.

a0, b0

initial guesses used for root search.

solver

the equation solver used for root search. The available options are BB::BBsolve, BB::dfsane, BB:BBoptim, and optim.

parallel

an logical value indicating whether parallel computation will be applied when se = "bootstrap" is called.

parCl

an integer value specifying the number of CPU cores to be used when parallel = TRUE. The default value is half the CPU cores on the current host.

References

Xu, G., Chiou, S.H., Huang, C.-Y., Wang, M.-C. and Yan, J. (2017). Joint Scale-change Models for Recurrent Events and Failure Time. Journal of the American Statistical Association, 112(518): 796--805.

Lin, D., Wei, L., Yang, I. and Ying, Z. (2000). Semiparametric Regression for the Mean and Rate Functions of Recurrent Events. Journal of the Royal Statistical Society: Series B (Methodological), 62: 711--730.

Wang, M.-C., Qin, J., and Chiang, C.-T. (2001). Analyzing Recurrent Event Data with Informative Censoring. Journal of the American Statistical Association, 96(455): 1057--1065.

Ghosh, D. and Lin, D.Y. (2003). Semiparametric Analysis of Recurrent Events Data in the Presence of Dependent Censoring. Biometrics, 59: 877--885.

Huang, C.-Y. and Wang, M.-C. (2004). Joint Modeling and Estimation for Recurrent Event Processes and Failure Time Data. Journal of the American Statistical Association, 99(468): 1153--1165.

See also

Examples

## readmission data
data(readmission, package = "frailtypack")
set.seed(123)
## Accelerated Mean Model
(fit <- reReg(reSurv(t.stop, id, event, death) ~ sex + chemo,
              data = subset(readmission, id < 50),
              method = "am.XCHWY", se = "resampling", B = 20))

#> Call: reReg(formula = reSurv(t.stop, id, event, death) ~ sex + chemo, 
#>     data = subset(readmission, id < 50), B = 20, method = "am.XCHWY", 
#>     se = "resampling")
#> 
#> Coefficients:
#>        sexFemale chemoTreated
#> alpha -0.4467567    0.3957031
#> beta  -0.5670908    0.5613490

summary(fit)

#> Call: reReg(formula = reSurv(t.stop, id, event, death) ~ sex + chemo, 
#>     data = subset(readmission, id < 50), B = 20, method = "am.XCHWY", 
#>     se = "resampling")
#> 
#> Method: Xu et al. (2016) Model 
#> 
#> Coefficients (rate):
#>              Estimate StdErr z.value p.value
#> sexFemale      -0.447  0.648  -0.690   0.490
#> chemoTreated    0.396  0.459   0.862   0.389
#> 
#> Coefficients (hazard):
#>              Estimate StdErr z.value p.value
#> sexFemale      -0.567  1.014  -0.559   0.576
#> chemoTreated    0.561  0.933   0.602   0.547
#> 

## Generalized Scale-Change Model
set.seed(123)
(fit <- reReg(reSurv(t.stop, id, event, death) ~ sex + chemo,
              data = subset(readmission, id < 50),
              method = "sc.XCYH", se = "resampling", B = 20))

#> Call: reReg(formula = reSurv(t.stop, id, event, death) ~ sex + chemo, 
#>     data = subset(readmission, id < 50), B = 20, method = "sc.XCYH", 
#>     se = "resampling")
#> 
#> Coefficients:
#>        sexFemale chemoTreated
#> alpha -0.3432329     2.026778
#> beta  -0.2004733     1.000851

summary(fit)

#> Call: reReg(formula = reSurv(t.stop, id, event, death) ~ sex + chemo, 
#>     data = subset(readmission, id < 50), B = 20, method = "sc.XCYH", 
#>     se = "resampling")
#> 
#> Method: Generalized Scale-Change Model 
#> 
#> Scale-change effect:
#>              Estimate StdErr z.value p.value
#> sexFemale      -0.343 11.924  -0.029   0.977
#> chemoTreated    2.027 24.285   0.083   0.933
#> 
#> Multiplicative coefficients:
#>              Estimate StdErr z.value p.value
#> sexFemale      -0.200  3.592  -0.056   0.955
#> chemoTreated    1.001 15.047   0.067   0.947
#> 
#> Hypothesis tests:
#> 
#> H0 Cox-type model:
#>      X-squared = 0.5529, df = 2, p-value = 0.7585
#> H0 Accelerated rate model:
#>      X-squared = 0.0046, df = 2, p-value = 0.9977
#> H0 Accelerated mean model:
#>      X-squared = 1.3497, df = 2, p-value = 0.5092