Function to generate simulated recurrent event data

The function simGSC() generates simulated recurrent event data from either a Cox-type model, an accelerated mean model, an accelerated rate model, or a generalized scale-change model.

simGSC(
  n,
  summary = FALSE,
  para,
  xmat,
  censoring,
  frailty,
  tau,
  origin,
  Lam0,
  Haz0
)

Arguments

n
number of observation.

summary
a logical value indicating whether a brief data summary will be printed.

para
a list of numerical vectors for the regression coefficients in the joint scale-change model. The names of the list elements are alpha, beta, eta, and theta, correspond to \(\alpha\), \(\beta\), \(\eta\), and \(\theta\) in the joint scale-change model, respectively. See Details for reReg.

xmat
an optional matrix specifying the design matrix.

censoring
a numeric variable specifying the censoring times for each of the \(n\) observation.

frailty
a numeric variable specifying the frailty variable.

tau
a numeric value specifying the maximum observation time.

origin
a numeric value specifying the time origin.

Lam0
is an optional function that specifies the baseline cumulative rate function. When left-unspecified, the recurrent events are generated using the baseline rate function of $$\lambda_0(t) = \frac{2}{1 + t},$$ or equivalently, the cumulative rate function of $$\Lambda_0(t) = 2\log(1 + t).$$

Haz0
is an optional function that specifies the baseline hazard function. When left-unspecified, the recurrent events are generated using the baseline hazard function $$h_0(t) = \frac{1}{5(1 + t)},$$ or equivalently, the cumulative hazard function of $$H_0(t) = \log(1 + t) / 5.$$

Details

The function simGSC() generates simulated recurrent event data over the interval \((0, \tau)\) based on the specification of the recurrent process and the terminal events. Specifically, the rate function, \(\lambda(t)\), of the recurrent process can be specified as one of the following model: $$\lambda(t) = Z \lambda_0(te^{X^\top\alpha}) e^{X^\top\beta}, h(t) = Z h_0(te^{X^\top\eta})e^{X^\top\theta}, $$ where \(\lambda_0(t)\) is the baseline rate function, \(h_0(t)\) is the baseline hazard function, \(X\) is a \(n\) by \(p\) covariate matrix and \(\alpha\), \(Z\) is an unobserved shared frailty variable, and \((\alpha, \eta)\) and \((\beta, \theta)\) correspond to the shape and size parameters of the rate function and the hazard function, respectively.

Under the default settings, the simGSC() function assumes \(p = 2\) and the regression parameters to be \(\alpha = \eta = (0, 0)^\top\), and \(\beta = \theta = (1, 1)^\top\). When the xmat argument is not specified, the simGSC() function assumes \(X_i\) is a two-dimensional vector \(X_i = (X_{i1}, X_{i2}), i = 1, \ldots, n\), where \(X_{i1}\) is a Bernoulli variable with rate 0.5 and \(X_{i2}\) is a standard normal variable. With the default xmat, the censoring time $C$ is generated from an independent uniform distribution in \([0, 2\tau X_{i1} + 2Z^2\tau(1 - X_{i1})]\). Thus, the censoring distribution is covariate dependent and is informative when \(Z\) is not a constant. When the frailty argument is not specified, the frailty variable \(Z\) is generated from a gamma distribution with a unit mean and a variance of 0.25. The default values for tau and origin are 60 and 0, respectively. When arguments Lam0 and Haz0 are left unspecified, the simGSC() function uses \(\Lambda_0(t) = 2\log(1 + t)\) and \(H_0(t) = \log(1 + t) / 5\), respectively. This is equivalent to setting Lam0 = function(x) 2 * log(1 + x) and Haz0 = function(x) log(1 + x) / 5. Overall, the default specifications generate the recurrent events and the terminal events from the model: $$\lambda(t) = \displaystyle \frac{2Z}{1 + te^{-X_{i1} - X_{i2}}}, h(t) = \displaystyle \frac{Z}{5(1 + te^{X_{i1} + X_{i2}})}, t\in[0, 60].$$

See also

Examples

set.seed(123)
simGSC(100, summary = TRUE)

#> Call: 
#> simGSC(n = 100, summary = TRUE)
#> 
#> Summary:
#> Sample size:                                    100 
#> Number of recurrent event observed:             369 
#> Average number of recurrent event per subject:  3.69 
#> Proportion of subjects with a terminal event:   0.52 
#> Median time-to-terminal event:                  13.874 
#> 
#> 

#>      id      t.start       t.stop event status x1           x2
#> 1     1 0.000000e+00 7.673334e-01     0      1  1  0.157076450
#> 2     2 0.000000e+00 4.792243e-01     1      0  0  0.922419336
#> 3     2 4.792243e-01 5.278326e-01     1      0  0  0.922419336
#> 4     2 5.278326e-01 5.352025e-01     1      0  0  0.922419336
#> 5     2 5.352025e-01 5.390913e-01     0      1  0  0.922419336
#> 6     3 0.000000e+00 5.938063e+00     1      0  1 -0.490958575
#> 7     3 5.938063e+00 2.459264e+01     0      0  1 -0.490958575
#> 8     4 0.000000e+00 6.064374e-01     1      0  1  1.098001882
#> 9     4 6.064374e-01 3.449541e+00     0      1  1  1.098001882
#> 10    5 0.000000e+00 1.039255e-02     1      0  0 -0.102333833
#> 11    5 1.039255e-02 8.348449e-01     1      0  0 -0.102333833
#> 12    5 8.348449e-01 1.241384e+00     1      0  0 -0.102333833
#> 13    5 1.241384e+00 2.128515e+00     1      0  0 -0.102333833
#> 14    5 2.128515e+00 2.263810e+00     1      0  0 -0.102333833
#> 15    5 2.263810e+00 2.787350e+00     1      0  0 -0.102333833
#> 16    5 2.787350e+00 5.155177e+00     1      0  0 -0.102333833
#> 17    5 5.155177e+00 6.851851e+00     1      0  0 -0.102333833
#> 18    5 6.851851e+00 7.713927e+00     1      0  0 -0.102333833
#> 19    5 7.713927e+00 9.232733e+00     1      0  0 -0.102333833
#> 20    5 9.232733e+00 1.022518e+01     1      0  0 -0.102333833
#> 21    5 1.022518e+01 1.034746e+01     1      0  0 -0.102333833
#> 22    5 1.034746e+01 1.635524e+01     1      0  0 -0.102333833
#> 23    5 1.635524e+01 1.967031e+01     1      0  0 -0.102333833
#> 24    5 1.967031e+01 2.106709e+01     1      0  0 -0.102333833
#> 25    5 2.106709e+01 2.149465e+01     1      0  0 -0.102333833
#> 26    5 2.149465e+01 3.142128e+01     1      0  0 -0.102333833
#> 27    5 3.142128e+01 3.863199e+01     1      0  0 -0.102333833
#> 28    5 3.863199e+01 4.026081e+01     1      0  0 -0.102333833
#> 29    5 4.026081e+01 6.000000e+01     0      0  0 -0.102333833
#> 30    6 0.000000e+00 9.999753e-01     1      0  0  0.487571468
#> 31    6 9.999753e-01 2.435904e+00     1      0  0  0.487571468
#> 32    6 2.435904e+00 6.000000e+01     0      0  0  0.487571468
#> 33    7 0.000000e+00 8.164655e-02     1      0  0 -0.433783058
#> 34    7 8.164655e-02 1.979892e+00     1      0  0 -0.433783058
#> 35    7 1.979892e+00 1.035070e+01     0      0  0 -0.433783058
#> 36    8 0.000000e+00 6.321277e-01     1      0  0 -0.250593796
#> 37    8 6.321277e-01 2.299605e+00     1      0  0 -0.250593796
#> 38    8 2.299605e+00 5.138238e+00     0      0  0 -0.250593796
#> 39    9 0.000000e+00 1.779847e-01     0      1  0  0.392795581
#> 40   10 0.000000e+00 4.900986e+00     1      0  1 -1.051123662
#> 41   10 4.900986e+00 7.380450e+00     0      1  1 -1.051123662
#> 42   11 0.000000e+00 1.237283e+00     1      0  0 -0.021102465
#> 43   11 1.237283e+00 2.230570e+00     0      0  0 -0.021102465
#> 44   12 0.000000e+00 9.890380e-04     1      0  1 -0.202404707
#> 45   12 9.890380e-04 3.381851e+00     0      1  1 -0.202404707
#> 46   13 0.000000e+00 4.833602e-01     1      0  1 -0.056382983
#> 47   13 4.833602e-01 9.760000e+00     1      0  1 -0.056382983
#> 48   13 9.760000e+00 1.104869e+01     1      0  1 -0.056382983
#> 49   13 1.104869e+01 3.886510e+01     1      0  1 -0.056382983
#> 50   13 3.886510e+01 6.000000e+01     0      0  1 -0.056382983
#> 51   14 0.000000e+00 2.023218e-01     0      1  1  0.898570220
#> 52   15 0.000000e+00 7.668470e-01     1      0  0 -0.405163129
#> 53   15 7.668470e-01 8.195277e-01     1      0  0 -0.405163129
#> 54   15 8.195277e-01 1.186631e+00     1      0  0 -0.405163129
#> 55   15 1.186631e+00 2.662091e+00     1      0  0 -0.405163129
#> 56   15 2.662091e+00 2.687997e+00     1      0  0 -0.405163129
#> 57   15 2.687997e+00 4.726670e+00     1      0  0 -0.405163129
#> 58   15 4.726670e+00 4.933471e+00     1      0  0 -0.405163129
#> 59   15 4.933471e+00 7.796493e+00     1      0  0 -0.405163129
#> 60   15 7.796493e+00 1.208258e+01     1      0  0 -0.405163129
#> 61   15 1.208258e+01 1.744109e+01     1      0  0 -0.405163129
#> 62   15 1.744109e+01 1.925079e+01     1      0  0 -0.405163129
#> 63   15 1.925079e+01 2.090413e+01     1      0  0 -0.405163129
#> 64   15 2.090413e+01 2.369241e+01     1      0  0 -0.405163129
#> 65   15 2.369241e+01 2.789842e+01     1      0  0 -0.405163129
#> 66   15 2.789842e+01 4.045521e+01     1      0  0 -0.405163129
#> 67   15 4.045521e+01 4.101196e+01     0      0  0 -0.405163129
#> 68   16 0.000000e+00 1.347988e+00     0      1  1  0.950450037
#> 69   17 0.000000e+00 9.322742e-01     1      0  0  0.354477220
#> 70   17 9.322742e-01 3.145312e+00     0      1  0  0.354477220
#> 71   18 0.000000e+00 1.848603e+00     0      1  1  0.368097387
#> 72   19 0.000000e+00 2.070892e+01     0      0  0  0.682888277
#> 73   20 0.000000e+00 1.837757e-01     1      0  0 -0.288131946
#> 74   20 1.837757e-01 1.567255e+00     1      0  0 -0.288131946
#> 75   20 1.567255e+00 3.649105e+00     1      0  0 -0.288131946
#> 76   20 3.649105e+00 1.488528e+01     1      0  0 -0.288131946
#> 77   20 1.488528e+01 2.190166e+01     1      0  0 -0.288131946
#> 78   20 2.190166e+01 3.717587e+01     1      0  0 -0.288131946
#> 79   20 3.717587e+01 4.582297e+01     0      0  0 -0.288131946
#> 80   21 0.000000e+00 7.398890e+00     0      0  0 -0.402366160
#> 81   22 0.000000e+00 3.573538e-01     1      0  0 -0.267532906
#> 82   22 3.573538e-01 8.493560e-01     1      0  0 -0.267532906
#> 83   22 8.493560e-01 1.142279e+01     0      0  0 -0.267532906
#> 84   23 0.000000e+00 6.460289e+00     1      0  0  0.395773955
#> 85   23 6.460289e+00 7.835392e+00     1      0  0  0.395773955
#> 86   23 7.835392e+00 1.951356e+01     0      0  0  0.395773955
#> 87   24 0.000000e+00 9.008201e-02     1      0  0 -0.353803575
#> 88   24 9.008201e-02 1.565315e-01     1      0  0 -0.353803575
#> 89   24 1.565315e-01 3.860905e-01     1      0  0 -0.353803575
#> 90   24 3.860905e-01 1.173988e+00     1      0  0 -0.353803575
#> 91   24 1.173988e+00 1.405757e+00     1      0  0 -0.353803575
#> 92   24 1.405757e+00 1.467210e+00     1      0  0 -0.353803575
#> 93   24 1.467210e+00 9.443959e+00     1      0  0 -0.353803575
#> 94   24 9.443959e+00 3.774314e+01     0      0  0 -0.353803575
#> 95   25 0.000000e+00 2.057282e+00     1      0  1 -0.637771686
#> 96   25 2.057282e+00 2.514794e+00     1      0  1 -0.637771686
#> 97   25 2.514794e+00 4.213940e+00     1      0  1 -0.637771686
#> 98   25 4.213940e+00 6.067740e+00     1      0  1 -0.637771686
#> 99   25 6.067740e+00 1.180050e+01     0      1  1 -0.637771686
#> 100  26 0.000000e+00 1.348991e+00     1      0  0  1.187732231
#> 101  26 1.348991e+00 6.000000e+01     0      0  0  1.187732231
#> 102  27 0.000000e+00 1.565897e-02     0      1  1 -0.546794874
#> 103  28 0.000000e+00 2.253134e-01     1      0  0  0.096220791
#> 104  28 2.253134e-01 4.033071e-01     1      0  0  0.096220791
#> 105  28 4.033071e-01 4.364948e-01     1      0  0  0.096220791
#> 106  28 4.364948e-01 4.555481e-01     1      0  0  0.096220791
#> 107  28 4.555481e-01 6.614891e-01     1      0  0  0.096220791
#> 108  28 6.614891e-01 1.467013e+00     1      0  0  0.096220791
#> 109  28 1.467013e+00 4.160644e+00     1      0  0  0.096220791
#> 110  28 4.160644e+00 1.436581e+01     1      0  0  0.096220791
#> 111  28 1.436581e+01 2.927817e+01     1      0  0  0.096220791
#> 112  28 2.927817e+01 3.924106e+01     0      1  0  0.096220791
#> 113  29 0.000000e+00 2.255300e-02     1      0  0 -0.063080191
#> 114  29 2.255300e-02 2.632242e-02     1      0  0 -0.063080191
#> 115  29 2.632242e-02 2.011347e-01     1      0  0 -0.063080191
#> 116  29 2.011347e-01 2.380527e-01     1      0  0 -0.063080191
#> 117  29 2.380527e-01 2.910620e-01     1      0  0 -0.063080191
#> 118  29 2.910620e-01 5.764885e-01     1      0  0 -0.063080191
#> 119  29 5.764885e-01 7.146380e-01     1      0  0 -0.063080191
#> 120  29 7.146380e-01 1.308733e+00     1      0  0 -0.063080191
#> 121  29 1.308733e+00 2.930865e+00     0      0  0 -0.063080191
#> 122  30 0.000000e+00 2.783760e-02     1      0  0 -0.694216070
#> 123  30 2.783760e-02 2.352572e-01     1      0  0 -0.694216070
#> 124  30 2.352572e-01 4.394143e-01     1      0  0 -0.694216070
#> 125  30 4.394143e-01 3.353034e+00     1      0  0 -0.694216070
#> 126  30 3.353034e+00 5.233463e+00     1      0  0 -0.694216070
#> 127  30 5.233463e+00 1.600442e+01     1      0  0 -0.694216070
#> 128  30 1.600442e+01 1.927304e+01     1      0  0 -0.694216070
#> 129  30 1.927304e+01 2.165265e+01     1      0  0 -0.694216070
#> 130  30 2.165265e+01 2.279783e+01     1      0  0 -0.694216070
#> 131  30 2.279783e+01 2.528339e+01     1      0  0 -0.694216070
#> 132  30 2.528339e+01 2.545949e+01     1      0  0 -0.694216070
#> 133  30 2.545949e+01 2.989786e+01     1      0  0 -0.694216070
#> 134  30 2.989786e+01 3.035262e+01     1      0  0 -0.694216070
#> 135  30 3.035262e+01 4.870705e+01     0      0  0 -0.694216070
#> 136  31 0.000000e+00 1.151866e+00     1      0  0  0.234951096
#> 137  31 1.151866e+00 2.943520e+00     1      0  0  0.234951096
#> 138  31 2.943520e+00 3.202859e+00     0      1  0  0.234951096
#> 139  32 0.000000e+00 6.000000e+01     0      0  1  0.480191179
#> 140  33 0.000000e+00 2.068138e+00     0      1  1 -0.002570644
#> 141  34 0.000000e+00 8.226492e-01     1      0  0 -0.435232221
#> 142  34 8.226492e-01 1.295081e+00     1      0  0 -0.435232221
#> 143  34 1.295081e+00 1.430225e+00     1      0  0 -0.435232221
#> 144  34 1.430225e+00 1.627231e+00     1      0  0 -0.435232221
#> 145  34 1.627231e+00 1.746745e+00     1      0  0 -0.435232221
#> 146  34 1.746745e+00 1.765047e+00     1      0  0 -0.435232221
#> 147  34 1.765047e+00 1.875574e+00     0      1  0 -0.435232221
#> 148  35 0.000000e+00 5.456630e-01     1      0  0 -0.231439520
#> 149  35 5.456630e-01 1.172969e+00     1      0  0 -0.231439520
#> 150  35 1.172969e+00 1.860006e+00     1      0  0 -0.231439520
#> 151  35 1.860006e+00 3.061952e+00     1      0  0 -0.231439520
#> 152  35 3.061952e+00 4.141062e+00     1      0  0 -0.231439520
#> 153  35 4.141062e+00 4.304177e+00     1      0  0 -0.231439520
#> 154  35 4.304177e+00 7.098237e+00     1      0  0 -0.231439520
#> 155  35 7.098237e+00 7.162430e+00     1      0  0 -0.231439520
#> 156  35 7.162430e+00 7.252725e+00     1      0  0 -0.231439520
#> 157  35 7.252725e+00 1.065667e+01     1      0  0 -0.231439520
#> 158  35 1.065667e+01 1.412273e+01     1      0  0 -0.231439520
#> 159  35 1.412273e+01 1.512011e+01     1      0  0 -0.231439520
#> 160  35 1.512011e+01 2.281397e+01     1      0  0 -0.231439520
#> 161  35 2.281397e+01 2.527802e+01     1      0  0 -0.231439520
#> 162  35 2.527802e+01 4.807647e+01     1      0  0 -0.231439520
#> 163  35 4.807647e+01 5.579706e+01     1      0  0 -0.231439520
#> 164  35 5.579706e+01 6.000000e+01     0      0  0 -0.231439520
#> 165  36 0.000000e+00 4.132026e-01     1      0  0 -0.955575628
#> 166  36 4.132026e-01 1.045911e+00     1      0  0 -0.955575628
#> 167  36 1.045911e+00 1.326520e+00     1      0  0 -0.955575628
#> 168  36 1.326520e+00 1.331590e+00     1      0  0 -0.955575628
#> 169  36 1.331590e+00 3.310202e+00     1      0  0 -0.955575628
#> 170  36 3.310202e+00 6.340162e+00     1      0  0 -0.955575628
#> 171  36 6.340162e+00 7.227096e+00     1      0  0 -0.955575628
#> 172  36 7.227096e+00 7.587522e+00     1      0  0 -0.955575628
#> 173  36 7.587522e+00 1.670387e+01     1      0  0 -0.955575628
#> 174  36 1.670387e+01 1.926063e+01     0      0  0 -0.955575628
#> 175  37 0.000000e+00 2.457078e-01     1      0  1  0.184908022
#> 176  37 2.457078e-01 1.923592e+00     0      1  1  0.184908022
#> 177  38 0.000000e+00 8.140886e-01     0      1  1 -0.231144029
#> 178  39 0.000000e+00 7.025056e-01     0      1  1  0.160977946
#> 179  40 0.000000e+00 3.212027e+00     1      0  0 -0.142000695
#> 180  40 3.212027e+00 1.111957e+01     1      0  0 -0.142000695
#> 181  40 1.111957e+01 1.395034e+01     1      0  0 -0.142000695
#> 182  40 1.395034e+01 2.014947e+01     1      0  0 -0.142000695
#> 183  40 2.014947e+01 3.307767e+01     0      1  0 -0.142000695
#> 184  41 0.000000e+00 1.181103e+01     1      0  1  0.545928222
#> 185  41 1.181103e+01 1.320714e+01     0      1  1  0.545928222
#> 186  42 0.000000e+00 6.611685e-02     0      1  0  0.080108181
#> 187  43 0.000000e+00 6.652349e-01     1      0  0  1.354084142
#> 188  43 6.652349e-01 1.497815e+00     0      0  0  1.354084142
#> 189  44 0.000000e+00 2.076827e+00     1      0  0  0.143144965
#> 190  44 2.076827e+00 3.273389e+00     1      0  0  0.143144965
#> 191  44 3.273389e+00 1.050416e+01     0      0  0  0.143144965
#> 192  45 0.000000e+00 1.906165e+00     1      0  0 -0.016255125
#> 193  45 1.906165e+00 2.478052e+00     1      0  0 -0.016255125
#> 194  45 2.478052e+00 2.871168e+00     0      0  0 -0.016255125
#> 195  46 0.000000e+00 5.602840e-01     1      0  0 -0.494839010
#> 196  46 5.602840e-01 6.831704e-01     1      0  0 -0.494839010
#> 197  46 6.831704e-01 7.342415e-01     1      0  0 -0.494839010
#> 198  46 7.342415e-01 1.836757e+00     1      0  0 -0.494839010
#> 199  46 1.836757e+00 1.943934e+00     1      0  0 -0.494839010
#> 200  46 1.943934e+00 3.023834e+00     1      0  0 -0.494839010
#> 201  46 3.023834e+00 3.613537e+00     1      0  0 -0.494839010
#> 202  46 3.613537e+00 5.032584e+00     1      0  0 -0.494839010
#> 203  46 5.032584e+00 1.016109e+01     0      0  0 -0.494839010
#> 204  47 0.000000e+00 6.108031e+00     0      1  1  0.105904964
#> 205  48 0.000000e+00 1.300019e+00     1      0  1 -0.486008767
#> 206  48 1.300019e+00 2.704357e+00     1      0  1 -0.486008767
#> 207  48 2.704357e+00 8.397697e+00     0      1  1 -0.486008767
#> 208  49 0.000000e+00 2.276894e-01     0      1  1  0.274612007
#> 209  50 0.000000e+00 1.117742e+00     1      0  0 -0.524510765
#> 210  50 1.117742e+00 4.520894e+00     1      0  0 -0.524510765
#> 211  50 4.520894e+00 1.407878e+01     0      0  0 -0.524510765
#> 212  51 0.000000e+00 6.431133e-02     1      0  0  0.224566967
#> 213  51 6.431133e-02 1.621943e+00     1      0  0  0.224566967
#> 214  51 1.621943e+00 2.327757e+00     1      0  0  0.224566967
#> 215  51 2.327757e+00 3.365998e+00     1      0  0  0.224566967
#> 216  51 3.365998e+00 4.012916e+00     0      0  0  0.224566967
#> 217  52 0.000000e+00 3.871640e+00     0      1  1 -0.192927826
#> 218  53 0.000000e+00 2.365260e+00     1      0  1 -0.785067761
#> 219  53 2.365260e+00 5.571977e+00     1      0  1 -0.785067761
#> 220  53 5.571977e+00 1.562437e+01     1      0  1 -0.785067761
#> 221  53 1.562437e+01 2.276292e+01     1      0  1 -0.785067761
#> 222  53 2.276292e+01 3.968548e+01     1      0  1 -0.785067761
#> 223  53 3.968548e+01 6.000000e+01     0      0  1 -0.785067761
#> 224  54 0.000000e+00 3.363785e+00     1      0  0  0.613471301
#> 225  54 3.363785e+00 3.434397e+00     0      1  0  0.613471301
#> 226  55 0.000000e+00 9.101713e-02     0      1  1  0.333347198
#> 227  56 0.000000e+00 1.247283e+00     1      0  0 -0.575801582
#> 228  56 1.247283e+00 1.304254e+00     1      0  0 -0.575801582
#> 229  56 1.304254e+00 1.506118e+00     1      0  0 -0.575801582
#> 230  56 1.506118e+00 2.045403e+00     1      0  0 -0.575801582
#> 231  56 2.045403e+00 6.889056e+00     1      0  0 -0.575801582
#> 232  56 6.889056e+00 9.105803e+00     0      0  0 -0.575801582
#> 233  57 0.000000e+00 3.202656e+00     1      0  0 -0.080394242
#> 234  57 3.202656e+00 3.764321e+00     1      0  0 -0.080394242
#> 235  57 3.764321e+00 2.446862e+01     0      0  0 -0.080394242
#> 236  58 0.000000e+00 8.882910e-01     1      0  1 -0.486456888
#> 237  58 8.882910e-01 6.484381e+00     1      0  1 -0.486456888
#> 238  58 6.484381e+00 1.755385e+01     0      0  1 -0.486456888
#> 239  59 0.000000e+00 1.139834e+00     1      0  0 -0.305230830
#> 240  59 1.139834e+00 1.608917e+00     1      0  0 -0.305230830
#> 241  59 1.608917e+00 1.958803e+00     1      0  0 -0.305230830
#> 242  59 1.958803e+00 2.085734e+00     1      0  0 -0.305230830
#> 243  59 2.085734e+00 2.539565e+00     1      0  0 -0.305230830
#> 244  59 2.539565e+00 3.139280e+00     1      0  0 -0.305230830
#> 245  59 3.139280e+00 5.039785e+00     1      0  0 -0.305230830
#> 246  59 5.039785e+00 6.848964e+00     1      0  0 -0.305230830
#> 247  59 6.848964e+00 7.004673e+00     1      0  0 -0.305230830
#> 248  59 7.004673e+00 9.042904e+00     1      0  0 -0.305230830
#> 249  59 9.042904e+00 1.133672e+01     1      0  0 -0.305230830
#> 250  59 1.133672e+01 1.754341e+01     1      0  0 -0.305230830
#> 251  59 1.754341e+01 3.615164e+01     1      0  0 -0.305230830
#> 252  59 3.615164e+01 3.792654e+01     1      0  0 -0.305230830
#> 253  59 3.792654e+01 4.108082e+01     1      0  0 -0.305230830
#> 254  59 4.108082e+01 4.456794e+01     1      0  0 -0.305230830
#> 255  59 4.456794e+01 6.000000e+01     0      0  0 -0.305230830
#> 256  60 0.000000e+00 3.574408e-01     1      0  0  0.249638438
#> 257  60 3.574408e-01 6.410603e-01     1      0  0  0.249638438
#> 258  60 6.410603e-01 6.881801e+00     1      0  0  0.249638438
#> 259  60 6.881801e+00 7.772954e+00     1      0  0  0.249638438
#> 260  60 7.772954e+00 2.016170e+01     1      0  0  0.249638438
#> 261  60 2.016170e+01 3.020942e+01     0      0  0  0.249638438
#> 262  61 0.000000e+00 7.083687e+00     0      0  0  0.439682070
#> 263  62 0.000000e+00 1.246346e+00     0      1  1  0.461564620
#> 264  63 0.000000e+00 2.515486e-01     1      0  0  0.346566986
#> 265  63 2.515486e-01 4.125132e-01     1      0  0  0.346566986
#> 266  63 4.125132e-01 1.013612e+00     1      0  0  0.346566986
#> 267  63 1.013612e+00 4.656791e+00     1      0  0  0.346566986
#> 268  63 4.656791e+00 8.933061e+00     0      0  0  0.346566986
#> 269  64 0.000000e+00 6.485166e-01     0      1  1 -0.069750162
#> 270  65 0.000000e+00 4.932926e-02     1      0  0 -1.655339363
#> 271  65 4.932926e-02 6.261737e-02     1      0  0 -1.655339363
#> 272  65 6.261737e-02 2.559807e-01     1      0  0 -1.655339363
#> 273  65 2.559807e-01 3.573117e-01     1      0  0 -1.655339363
#> 274  65 3.573117e-01 4.295981e-01     1      0  0 -1.655339363
#> 275  65 4.295981e-01 4.663109e-01     1      0  0 -1.655339363
#> 276  65 4.663109e-01 1.022415e+00     1      0  0 -1.655339363
#> 277  65 1.022415e+00 2.527235e+00     1      0  0 -1.655339363
#> 278  65 2.527235e+00 3.583408e+00     1      0  0 -1.655339363
#> 279  65 3.583408e+00 3.630561e+00     1      0  0 -1.655339363
#> 280  65 3.630561e+00 4.466413e+00     1      0  0 -1.655339363
#> 281  65 4.466413e+00 4.721404e+00     1      0  0 -1.655339363
#> 282  65 4.721404e+00 5.067057e+00     1      0  0 -1.655339363
#> 283  65 5.067057e+00 5.700125e+00     1      0  0 -1.655339363
#> 284  65 5.700125e+00 6.234603e+00     1      0  0 -1.655339363
#> 285  65 6.234603e+00 6.317163e+00     1      0  0 -1.655339363
#> 286  65 6.317163e+00 7.784869e+00     1      0  0 -1.655339363
#> 287  65 7.784869e+00 7.849777e+00     1      0  0 -1.655339363
#> 288  65 7.849777e+00 9.430274e+00     1      0  0 -1.655339363
#> 289  65 9.430274e+00 1.031342e+01     1      0  0 -1.655339363
#> 290  65 1.031342e+01 1.374463e+01     1      0  0 -1.655339363
#> 291  65 1.374463e+01 1.422748e+01     1      0  0 -1.655339363
#> 292  65 1.422748e+01 2.087489e+01     1      0  0 -1.655339363
#> 293  65 2.087489e+01 3.203278e+01     1      0  0 -1.655339363
#> 294  65 3.203278e+01 3.603228e+01     1      0  0 -1.655339363
#> 295  65 3.603228e+01 6.000000e+01     0      0  0 -1.655339363
#> 296  66 0.000000e+00 6.716704e+00     0      1  1  0.269232900
#> 297  67 0.000000e+00 2.231122e-01     0      1  1  0.446242933
#> 298  68 0.000000e+00 1.516876e+00     0      1  1  0.244731018
#> 299  69 0.000000e+00 1.867339e-01     1      0  1 -0.598175873
#> 300  69 1.867339e-01 1.179108e+00     1      0  1 -0.598175873
#> 301  69 1.179108e+00 1.400892e+00     1      0  1 -0.598175873
#> 302  69 1.400892e+00 1.923110e+01     0      1  1 -0.598175873
#> 303  70 0.000000e+00 4.806616e-01     1      0  0  0.231368927
#> 304  70 4.806616e-01 1.604211e+00     1      0  0  0.231368927
#> 305  70 1.604211e+00 1.652917e+00     1      0  0  0.231368927
#> 306  70 1.652917e+00 1.982361e+00     1      0  0  0.231368927
#> 307  70 1.982361e+00 6.000000e+01     0      0  0  0.231368927
#> 308  71 0.000000e+00 4.890369e-02     1      0  0  0.485851788
#> 309  71 4.890369e-02 2.111682e-01     1      0  0  0.485851788
#> 310  71 2.111682e-01 5.027649e+00     0      0  0  0.485851788
#> 311  72 0.000000e+00 1.584647e-01     1      0  0 -0.736079482
#> 312  72 1.584647e-01 1.504156e+00     1      0  0 -0.736079482
#> 313  72 1.504156e+00 1.735140e+00     1      0  0 -0.736079482
#> 314  72 1.735140e+00 1.796161e+00     1      0  0 -0.736079482
#> 315  72 1.796161e+00 1.148065e+01     1      0  0 -0.736079482
#> 316  72 1.148065e+01 1.369229e+01     1      0  0 -0.736079482
#> 317  72 1.369229e+01 2.076923e+01     1      0  0 -0.736079482
#> 318  72 2.076923e+01 2.164684e+01     1      0  0 -0.736079482
#> 319  72 2.164684e+01 2.590099e+01     1      0  0 -0.736079482
#> 320  72 2.590099e+01 2.841050e+01     1      0  0 -0.736079482
#> 321  72 2.841050e+01 4.029755e+01     1      0  0 -0.736079482
#> 322  72 4.029755e+01 4.740414e+01     0      0  0 -0.736079482
#> 323  73 0.000000e+00 1.121501e+00     1      0  1  0.264482470
#> 324  73 1.121501e+00 1.653021e+00     1      0  1  0.264482470
#> 325  73 1.653021e+00 1.002727e+01     1      0  1  0.264482470
#> 326  73 1.002727e+01 1.639212e+01     0      1  1  0.264482470
#> 327  74 0.000000e+00 2.691809e-01     0      1  0  1.227200176
#> 328  75 0.000000e+00 9.133437e-02     1      0  0 -0.824672740
#> 329  75 9.133437e-02 9.929944e-02     1      0  0 -0.824672740
#> 330  75 9.929944e-02 2.405427e-01     0      1  0 -0.824672740
#> 331  76 0.000000e+00 4.722508e-01     0      1  1  0.397902578
#> 332  77 0.000000e+00 1.786242e-01     1      0  0 -0.170646583
#> 333  77 1.786242e-01 2.187578e-01     1      0  0 -0.170646583
#> 334  77 2.187578e-01 2.331823e+00     1      0  0 -0.170646583
#> 335  77 2.331823e+00 3.317999e+00     1      0  0 -0.170646583
#> 336  77 3.317999e+00 3.434961e+00     1      0  0 -0.170646583
#> 337  77 3.434961e+00 3.916352e+00     0      1  0 -0.170646583
#> 338  78 0.000000e+00 1.263900e-01     1      0  0 -0.702626900
#> 339  78 1.263900e-01 1.666562e+00     1      0  0 -0.702626900
#> 340  78 1.666562e+00 4.108156e+00     1      0  0 -0.702626900
#> 341  78 4.108156e+00 6.417344e+00     1      0  0 -0.702626900
#> 342  78 6.417344e+00 8.272144e+00     1      0  0 -0.702626900
#> 343  78 8.272144e+00 1.384900e+01     1      0  0 -0.702626900
#> 344  78 1.384900e+01 1.677600e+01     1      0  0 -0.702626900
#> 345  78 1.677600e+01 1.938759e+01     0      0  0 -0.702626900
#> 346  79 0.000000e+00 3.379712e-01     0      1  1 -0.461480408
#> 347  80 0.000000e+00 3.171416e-01     1      0  0  0.004949065
#> 348  80 3.171416e-01 4.060231e-01     1      0  0  0.004949065
#> 349  80 4.060231e-01 9.453294e-01     1      0  0  0.004949065
#> 350  80 9.453294e-01 1.340897e+00     1      0  0  0.004949065
#> 351  80 1.340897e+00 1.356481e+00     1      0  0  0.004949065
#> 352  80 1.356481e+00 2.047346e+00     1      0  0  0.004949065
#> 353  80 2.047346e+00 2.181133e+00     1      0  0  0.004949065
#> 354  80 2.181133e+00 2.384316e+00     1      0  0  0.004949065
#> 355  80 2.384316e+00 3.191958e+00     1      0  0  0.004949065
#> 356  80 3.191958e+00 3.777452e+00     1      0  0  0.004949065
#> 357  80 3.777452e+00 3.949139e+00     0      1  0  0.004949065
#> 358  81 0.000000e+00 1.921153e-01     0      1  1 -0.204429887
#> 359  82 0.000000e+00 1.490047e-01     1      0  1 -0.694122644
#> 360  82 1.490047e-01 2.579894e-01     1      0  1 -0.694122644
#> 361  82 2.579894e-01 2.997549e-01     1      0  1 -0.694122644
#> 362  82 2.997549e-01 4.186610e-01     1      0  1 -0.694122644
#> 363  82 4.186610e-01 7.240851e+00     0      1  1 -0.694122644
#> 364  83 0.000000e+00 1.391789e+01     1      0  1 -0.132297619
#> 365  83 1.391789e+01 6.000000e+01     0      0  1 -0.132297619
#> 366  84 0.000000e+00 3.530912e-01     1      0  0 -0.473649139
#> 367  84 3.530912e-01 1.334723e+00     1      0  0 -0.473649139
#> 368  84 1.334723e+00 1.984983e+00     1      0  0 -0.473649139
#> 369  84 1.984983e+00 3.076728e+00     1      0  0 -0.473649139
#> 370  84 3.076728e+00 6.471379e+00     1      0  0 -0.473649139
#> 371  84 6.471379e+00 6.952363e+00     1      0  0 -0.473649139
#> 372  84 6.952363e+00 1.312876e+01     1      0  0 -0.473649139
#> 373  84 1.312876e+01 2.641999e+01     1      0  0 -0.473649139
#> 374  84 2.641999e+01 3.920847e+01     1      0  0 -0.473649139
#> 375  84 3.920847e+01 4.008262e+01     0      0  0 -0.473649139
#> 376  85 0.000000e+00 1.612917e-01     1      0  0  0.369760664
#> 377  85 1.612917e-01 4.829317e-01     1      0  0  0.369760664
#> 378  85 4.829317e-01 1.270574e+00     1      0  0  0.369760664
#> 379  85 1.270574e+00 3.967454e+00     1      0  0  0.369760664
#> 380  85 3.967454e+00 4.010330e+00     1      0  0  0.369760664
#> 381  85 4.010330e+00 8.297516e+00     1      0  0  0.369760664
#> 382  85 8.297516e+00 4.105444e+01     1      0  0  0.369760664
#> 383  85 4.105444e+01 4.482648e+01     1      0  0  0.369760664
#> 384  85 4.482648e+01 4.611664e+01     1      0  0  0.369760664
#> 385  85 4.611664e+01 5.496999e+01     1      0  0  0.369760664
#> 386  85 5.496999e+01 6.000000e+01     0      0  0  0.369760664
#> 387  86 0.000000e+00 1.671088e+00     1      0  0  0.448389354
#> 388  86 1.671088e+00 3.832459e+00     1      0  0  0.448389354
#> 389  86 3.832459e+00 4.035030e+00     1      0  0  0.448389354
#> 390  86 4.035030e+00 5.836090e+00     1      0  0  0.448389354
#> 391  86 5.836090e+00 6.381097e+00     1      0  0  0.448389354
#> 392  86 6.381097e+00 6.746338e+00     0      1  0  0.448389354
#> 393  87 0.000000e+00 1.026519e-01     1      0  0 -0.173000444
#> 394  87 1.026519e-01 8.448658e-01     1      0  0 -0.173000444
#> 395  87 8.448658e-01 1.484730e+00     1      0  0 -0.173000444
#> 396  87 1.484730e+00 3.379099e+00     1      0  0 -0.173000444
#> 397  87 3.379099e+00 4.124874e+00     1      0  0 -0.173000444
#> 398  87 4.124874e+00 4.321140e+00     1      0  0 -0.173000444
#> 399  87 4.321140e+00 1.002313e+01     1      0  0 -0.173000444
#> 400  87 1.002313e+01 1.567652e+01     0      0  0 -0.173000444
#> 401  88 0.000000e+00 7.071221e-01     1      0  1 -0.891028534
#> 402  88 7.071221e-01 4.149804e+00     1      0  1 -0.891028534
#> 403  88 4.149804e+00 1.424293e+01     1      0  1 -0.891028534
#> 404  88 1.424293e+01 1.874865e+01     1      0  1 -0.891028534
#> 405  88 1.874865e+01 2.615463e+01     1      0  1 -0.891028534
#> 406  88 2.615463e+01 6.000000e+01     0      0  1 -0.891028534
#> 407  89 0.000000e+00 1.226071e+00     0      1  0  0.232470877
#> 408  90 0.000000e+00 3.354035e-04     1      0  0 -0.975515305
#> 409  90 3.354035e-04 2.135813e-01     1      0  0 -0.975515305
#> 410  90 2.135813e-01 2.749449e-01     1      0  0 -0.975515305
#> 411  90 2.749449e-01 4.342548e-01     1      0  0 -0.975515305
#> 412  90 4.342548e-01 5.956394e-01     1      0  0 -0.975515305
#> 413  90 5.956394e-01 1.962013e+00     1      0  0 -0.975515305
#> 414  90 1.962013e+00 3.746200e+00     1      0  0 -0.975515305
#> 415  90 3.746200e+00 8.077611e+00     1      0  0 -0.975515305
#> 416  90 8.077611e+00 9.601795e+00     1      0  0 -0.975515305
#> 417  90 9.601795e+00 1.605971e+01     1      0  0 -0.975515305
#> 418  90 1.605971e+01 4.680406e+01     1      0  0 -0.975515305
#> 419  90 4.680406e+01 4.795610e+01     1      0  0 -0.975515305
#> 420  90 4.795610e+01 5.603764e+01     0      0  0 -0.975515305
#> 421  91 0.000000e+00 1.719284e-01     1      0  0 -0.258056670
#> 422  91 1.719284e-01 8.099332e-01     1      0  0 -0.258056670
#> 423  91 8.099332e-01 1.560305e+00     1      0  0 -0.258056670
#> 424  91 1.560305e+00 1.628930e+00     1      0  0 -0.258056670
#> 425  91 1.628930e+00 2.255163e+00     1      0  0 -0.258056670
#> 426  91 2.255163e+00 2.292483e+00     1      0  0 -0.258056670
#> 427  91 2.292483e+00 2.481852e+00     1      0  0 -0.258056670
#> 428  91 2.481852e+00 3.361098e+00     1      0  0 -0.258056670
#> 429  91 3.361098e+00 1.701309e+01     1      0  0 -0.258056670
#> 430  91 1.701309e+01 2.058453e+01     1      0  0 -0.258056670
#> 431  91 2.058453e+01 2.884336e+01     1      0  0 -0.258056670
#> 432  91 2.884336e+01 3.590606e+01     1      0  0 -0.258056670
#> 433  91 3.590606e+01 3.769629e+01     1      0  0 -0.258056670
#> 434  91 3.769629e+01 4.209508e+01     1      0  0 -0.258056670
#> 435  91 4.209508e+01 6.000000e+01     0      0  0 -0.258056670
#> 436  92 0.000000e+00 8.550434e-01     1      0  0  0.244540159
#> 437  92 8.550434e-01 9.143162e-01     0      1  0  0.244540159
#> 438  93 0.000000e+00 7.368084e+00     0      1  1  0.451012363
#> 439  94 0.000000e+00 1.387448e+01     0      1  1  0.320181497
#> 440  95 0.000000e+00 2.699010e-01     1      0  0  0.475604439
#> 441  95 2.699010e-01 1.600107e+01     1      0  0  0.475604439
#> 442  95 1.600107e+01 2.125056e+01     1      0  0  0.475604439
#> 443  95 2.125056e+01 6.000000e+01     0      0  0  0.475604439
#> 444  96 0.000000e+00 7.175747e-01     0      1  1 -0.299561605
#> 445  97 0.000000e+00 1.790488e-01     1      0  0 -0.665347496
#> 446  97 1.790488e-01 5.636635e-01     1      0  0 -0.665347496
#> 447  97 5.636635e-01 1.351069e+00     1      0  0 -0.665347496
#> 448  97 1.351069e+00 4.147202e+00     1      0  0 -0.665347496
#> 449  97 4.147202e+00 4.861684e+00     1      0  0 -0.665347496
#> 450  97 4.861684e+00 6.465001e+00     1      0  0 -0.665347496
#> 451  97 6.465001e+00 6.654952e+00     1      0  0 -0.665347496
#> 452  97 6.654952e+00 7.324026e+00     1      0  0 -0.665347496
#> 453  97 7.324026e+00 9.072599e+00     1      0  0 -0.665347496
#> 454  97 9.072599e+00 1.195089e+01     0      1  0 -0.665347496
#> 455  98 0.000000e+00 2.214560e-02     1      0  0 -0.296104873
#> 456  98 2.214560e-02 5.998626e-01     1      0  0 -0.296104873
#> 457  98 5.998626e-01 1.259891e+00     1      0  0 -0.296104873
#> 458  98 1.259891e+00 1.574676e+00     1      0  0 -0.296104873
#> 459  98 1.574676e+00 2.465017e+00     1      0  0 -0.296104873
#> 460  98 2.465017e+00 4.421609e+00     1      0  0 -0.296104873
#> 461  98 4.421609e+00 4.857524e+00     1      0  0 -0.296104873
#> 462  98 4.857524e+00 9.015672e+00     1      0  0 -0.296104873
#> 463  98 9.015672e+00 2.054160e+01     1      0  0 -0.296104873
#> 464  98 2.054160e+01 2.068759e+01     1      0  0 -0.296104873
#> 465  98 2.068759e+01 2.562922e+01     1      0  0 -0.296104873
#> 466  98 2.562922e+01 3.883632e+01     1      0  0 -0.296104873
#> 467  98 3.883632e+01 6.000000e+01     0      0  0 -0.296104873
#> 468  99 0.000000e+00 1.340162e+00     0      1  0  0.900525428
#> 469 100 0.000000e+00 5.127804e-01     0      1  1  0.527669296