Function to generate simulated data

The function simSC generates simulated recurrent event data from either a Cox-type model, an accelerated mean model, or a scale-change model. The censoring time could be either independent (given covariates) or informative. The simulated data is used for illustration.

simSC(n, a, b, indCen = TRUE, type = c("cox", "am", "sc"), tau = 60,
  summary = FALSE)

Arguments

n
number of observation.

a
a numeric vector of parameter of length 2.

b
a numeric vector of parameter of length 2.

indCen
a logical value indicating whether the censoring assumption is imposed. When indCen = TRUE, we set \(Z = 1\). Otherwise, \(Z\) is generated from a gamma distribution with mean 1 and variance 0.25 (e.g., rgamma(1, 4, 4)). See Details.

type
a character string specifying the underlying model. See Details

tau
a numeric value specifying the maximum observation time.

summary
a logical value indicating whether a brief data summary will be printed.

Details

The function simSC generates simulated recurrent event data under different scenarios based on the following assumptions. See Details in reReg for a more complete model assumptions.

type = "cox"

generates recurrent event data from a Cox-type model with $$\lambda(t) = Z \lambda_0(t) e^{X^\top a}, h(t) = Zh_0(t)e^{X^\top b}.$$

type = "am"

generates recurrent event data from an accelerated mean model with $$\lambda(t) = Z \lambda_0(te^{X^\top a}) e^{X^\top a}, h(t) = Zh_0(te^{X^\top b})e^{X^\top b}.$$

type = "sc"

generates recurrent event data from a generalized scale-change model with $$\lambda(t) = Z \lambda_0(te^{X^\top a}) e^{X^\top b}, h(t) = Zh_0(te^{X^\top a})e^{X^\top b}.$$

Let \(D\) be the informative failure time with the above hazard function. An non-informative failure time, \(C\), is generated separately from an exponential distribution with mean 80. The observed follow-up time is then taken to be \(min(D, C, \tau)\). We further assume $$\lambda_0(t) = \frac{2}{1 + t}, h_0(t) = \frac{1}{8(1 + t)}.$$ Two covariates are considered; x1 follows a Bernoulli distribution with probability 0.5 and x2 follows a standard normal distribution.

See also

Examples

set.seed(123)
simSC(200, c(-1, 1), c(-1, 1), summary = TRUE)

#> 
#> Summary results for number of recurrent event per subject:
#>    Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
#>    0.00    1.00    2.00    3.87    5.00   33.00 
#> 
#> Number of failures: 49 (24.5%); Number of censored events: 151 (75.5%)
#> 
#> Number of x1 == 1: 96 (48%); Number of x1 == 0: 104 (52%)
#> Summary results for x2:
#>     Min.  1st Qu.   Median     Mean  3rd Qu.     Max. 
#> -2.16226 -0.68832 -0.08922 -0.04727  0.61537  2.43023 
#> 
#> 

#> # A tibble: 974 x 6
#>       id    Time event status    x1     x2
#>    <int>   <dbl> <dbl>  <dbl> <dbl>  <dbl>
#>  1     1  0.0737     1      0     0  0.801
#>  2     1  0.152      1      0     0  0.801
#>  3     1  0.190      1      0     0  0.801
#>  4     1  0.224      0      1     0  0.801
#>  5     2  1.14       1      0     1 -0.687
#>  6     2  4.92       1      0     1 -0.687
#>  7     2  8.84       1      0     1 -0.687
#>  8     2 60          0      0     1 -0.687
#>  9     3  0.403      1      0     0  1.79 
#> 10     3  0.505      1      0     0  1.79 
#> # ... with 964 more rows