Function to generate simulated data

The function simSC generates simulated recurrent event data from either a Cox-type model, an accelerated mean model, or a scale-change model. The censoring time could be either independent (given covariates) or informative. The simulated data is used for illustration.

simSC(n, a, b, indCen = TRUE, type = c("cox", "am", "sc"), tau = 60,
  zVar = 0.25, summary = FALSE)

Arguments

n
number of observation.

a
a numeric vector of parameter of length 2.

b
a numeric vector of parameter of length 2.

indCen
a logical value indicating whether the censoring assumption is imposed. When indCen = TRUE, we set \(Z = 1\). Otherwise, \(Z\) is generated from a unit-mean gamma distribution See Details.

type
a character string specifying the underlying model. See Details

tau
a numeric value specifying the maximum observation time.

zVar
a numeric variable specifying the variance of \(Z\). This is only needed when \(indCen\) is TRUE. The default value is 0.25.

summary
a logical value indicating whether a brief data summary will be printed.

Details

The function simSC generates simulated recurrent event data under different scenarios based on the following assumptions. See Details in reReg for a more complete model assumptions.

type = "cox"

generates recurrent event data from a Cox-type model with $$\lambda(t) = Z \lambda_0(t) e^{X^\top a}, h(t) = Zh_0(t)e^{X^\top b}.$$

type = "am"

generates recurrent event data from an accelerated mean model with $$\lambda(t) = Z \lambda_0(te^{X^\top a}) e^{X^\top a}, h(t) = Zh_0(te^{X^\top b})e^{X^\top b}.$$

type = "sc"

generates recurrent event data from a generalized scale-change model with $$\lambda(t) = Z \lambda_0(te^{X^\top a}) e^{X^\top b}, h(t) = Zh_0(te^{X^\top a})e^{X^\top b}.$$

Let \(D\) be the informative failure time with the above hazard function. An non-informative failure time, \(C\), is generated separately from an exponential distribution with mean 80. The observed follow-up time is then taken to be \(min(D, C, \tau)\). We further assume $$\lambda_0(t) = \frac{2}{1 + t}, h_0(t) = \frac{1}{8(1 + t)}.$$ Two covariates are considered; x1 follows a Bernoulli distribution with probability 0.5 and x2 follows a standard normal distribution.

See also

Examples

set.seed(123)
simSC(200, c(-1, 1), c(-1, 1), summary = TRUE)

#> 
#> Summary results for number of recurrent event per subject:
#>    Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
#>   0.000   1.000   2.000   3.925   4.250  33.000 
#> 
#> Number of failures: 63 (31.5%); Number of censored events: 137 (68.5%)
#> 
#> Number of x1 == 1: 96 (48%); Number of x1 == 0: 104 (52%)
#> Summary results for x2:
#>     Min.  1st Qu.   Median     Mean  3rd Qu.     Max. 
#> -2.31592 -0.57302 -0.01835  0.06422  0.59445  3.24104 
#> 
#> 

#>      id         Time event status x1            x2
#> 1     1 7.365335e-02     1      0  0  0.8005542820
#> 2     1 1.521431e-01     1      0  0  0.8005542820
#> 3     1 1.903436e-01     1      0  0  0.8005542820
#> 4     1 2.237817e-01     0      1  0  0.8005542820
#> 5     2 3.221384e-01     1      0  0 -0.6868528519
#> 6     2 9.232221e-01     1      0  0 -0.6868528519
#> 7     2 1.318927e+00     1      0  0 -0.6868528519
#> 8     2 4.395465e+00     1      0  0 -0.6868528519
#> 9     2 2.450702e+01     1      0  0 -0.6868528519
#> 10    2 3.014490e+01     0      0  0 -0.6868528519
#> 11    3 3.101514e+01     0      0  1 -1.9666171566
#> 12    4 1.322024e+01     1      0  1 -0.2179749147
#> 13    4 1.400463e+01     1      0  1 -0.2179749147
#> 14    4 4.021594e+01     1      0  1 -0.2179749147
#> 15    4 6.000000e+01     0      0  1 -0.2179749147
#> 16    5 1.917115e-01     1      0  0  1.2538149211
#> 17    5 3.340591e-01     1      0  0  1.2538149211
#> 18    5 4.594051e-01     1      0  0  1.2538149211
#> 19    5 7.455591e-01     1      0  0  1.2538149211
#> 20    5 8.985383e-01     1      0  0  1.2538149211
#> 21    5 1.230536e+00     1      0  0  1.2538149211
#> 22    5 1.368438e+00     1      0  0  1.2538149211
#> 23    5 5.627924e+00     1      0  0  1.2538149211
#> 24    5 6.478125e+00     1      0  0  1.2538149211
#> 25    5 6.722734e+00     1      0  0  1.2538149211
#> 26    5 8.035800e+00     1      0  0  1.2538149211
#> 27    5 9.906912e+00     0      0  0  1.2538149211
#> 28    6 7.120155e-02     1      0  0  0.8001768658
#> 29    6 3.613862e-01     1      0  0  0.8001768658
#> 30    6 3.848405e-01     0      1  0  0.8001768658
#> 31    7 2.832726e+01     1      0  1 -0.8864367164
#> 32    7 6.000000e+01     0      0  1 -0.8864367164
#> 33    8 3.812927e-02     1      0  0  0.2760234783
#> 34    8 7.768721e-02     1      0  0  0.2760234783
#> 35    8 1.986486e-01     1      0  0  0.2760234783
#> 36    8 2.910956e-01     0      1  0  0.2760234783
#> 37    9 1.347929e+00     1      0  1 -0.5023234531
#> 38    9 3.011829e+00     1      0  1 -0.5023234531
#> 39    9 8.658776e+00     0      1  1 -0.5023234531
#> 40   10 2.146130e+01     0      1  1  0.4120222748
#> 41   11 5.040348e-02     1      0  0  1.0057385245
#> 42   11 3.727188e-01     1      0  0  1.0057385245
#> 43   11 8.184109e-01     1      0  0  1.0057385245
#> 44   11 8.199410e-01     1      0  0  1.0057385245
#> 45   11 1.229067e+00     1      0  0  1.0057385245
#> 46   11 1.354772e+00     1      0  0  1.0057385245
#> 47   11 1.928370e+00     1      0  0  1.0057385245
#> 48   11 2.590718e+00     1      0  0  1.0057385245
#> 49   11 2.635237e+00     1      0  0  1.0057385245
#> 50   11 2.890532e+00     0      1  0  1.0057385245
#> 51   12 3.347501e-01     1      0  0 -0.2204865618
#> 52   12 5.463154e-01     1      0  0 -0.2204865618
#> 53   12 2.230790e+00     1      0  0 -0.2204865618
#> 54   12 2.353224e+00     1      0  0 -0.2204865618
#> 55   12 3.311464e+00     1      0  0 -0.2204865618
#> 56   12 6.760535e+00     1      0  0 -0.2204865618
#> 57   12 7.798383e+00     0      0  0 -0.2204865618
#> 58   13 7.771103e-02     1      0  0  0.2387317351
#> 59   13 5.847286e-01     1      0  0  0.2387317351
#> 60   13 7.999571e-01     1      0  0  0.2387317351
#> 61   13 1.256850e+00     0      1  0  0.2387317351
#> 62   14 5.199450e+00     0      1  1 -1.0264209003
#> 63   15 4.063197e+00     1      0  1 -0.3475425994
#> 64   15 8.904519e+00     1      0  1 -0.3475425994
#> 65   15 6.000000e+01     0      0  1 -0.3475425994
#> 66   16 1.479485e+00     1      0  1 -0.3802265203
#> 67   16 4.201991e+00     1      0  1 -0.3802265203
#> 68   16 3.851416e+01     0      0  1 -0.3802265203
#> 69   17 1.285365e-01     1      0  1 -0.0555619655
#> 70   17 2.379139e+01     0      0  1 -0.0555619655
#> 71   18 1.836897e-01     1      0  1  0.9618103517
#> 72   18 1.177950e+00     0      1  1  0.9618103517
#> 73   19 3.174970e+01     1      0  1 -0.4905574437
#> 74   19 3.659357e+01     1      0  1 -0.4905574437
#> 75   19 6.000000e+01     0      0  1 -0.4905574437
#> 76   20 2.560680e-01     1      0  0 -0.2419765098
#> 77   20 1.433597e+00     1      0  0 -0.2419765098
#> 78   20 3.848943e+00     1      0  0 -0.2419765098
#> 79   20 1.834021e+01     0      0  0 -0.2419765098
#> 80   21 9.112601e-02     1      0  0  1.1313372134
#> 81   21 1.771321e-01     1      0  0  1.1313372134
#> 82   21 1.869626e-01     1      0  0  1.1313372134
#> 83   21 2.902296e-01     1      0  0  1.1313372134
#> 84   21 6.389443e-01     1      0  0  1.1313372134
#> 85   21 1.014628e+00     1      0  0  1.1313372134
#> 86   21 1.535840e+00     1      0  0  1.1313372134
#> 87   21 2.568721e+00     1      0  0  1.1313372134
#> 88   21 3.490327e+00     1      0  0  1.1313372134
#> 89   21 4.178595e+00     1      0  0  1.1313372134
#> 90   21 6.993208e+00     1      0  0  1.1313372134
#> 91   21 1.045330e+01     1      0  0  1.1313372134
#> 92   21 1.146104e+01     1      0  0  1.1313372134
#> 93   21 1.159979e+01     1      0  0  1.1313372134
#> 94   21 1.281126e+01     1      0  0  1.1313372134
#> 95   21 2.693022e+01     1      0  0  1.1313372134
#> 96   21 2.956136e+01     1      0  0  1.1313372134
#> 97   21 3.190874e+01     1      0  0  1.1313372134
#> 98   21 3.331968e+01     1      0  0  1.1313372134
#> 99   21 3.741480e+01     1      0  0  1.1313372134
#> 100  21 4.387163e+01     1      0  0  1.1313372134
#> 101  21 4.949095e+01     1      0  0  1.1313372134
#> 102  21 5.941308e+01     1      0  0  1.1313372134
#> 103  21 6.000000e+01     0      0  0  1.1313372134
#> 104  22 2.101505e+00     1      0  1 -0.7135608059
#> 105  22 1.230595e+01     0      0  1 -0.7135608059
#> 106  23 9.346804e-03     1      0  0  3.2410399349
#> 107  23 2.262481e-02     1      0  0  3.2410399349
#> 108  23 4.181114e-02     1      0  0  3.2410399349
#> 109  23 6.407745e-02     0      1  0  3.2410399349
#> 110  24 1.035638e-01     1      0  0 -0.2153805076
#> 111  24 1.561780e+00     0      0  0 -0.2153805076
#> 112  25 3.810208e-01     1      0  0  0.0377883992
#> 113  25 1.894299e+00     1      0  0  0.0377883992
#> 114  25 4.621029e+00     0      1  0  0.0377883992
#> 115  26 1.703679e+01     0      0  0 -0.5011875918
#> 116  27 2.278467e-01     1      0  1 -0.0422049297
#> 117  27 3.578328e+00     1      0  1 -0.0422049297
#> 118  27 5.603462e+00     1      0  1 -0.0422049297
#> 119  27 1.316655e+01     1      0  1 -0.0422049297
#> 120  27 1.476379e+01     1      0  1 -0.0422049297
#> 121  27 1.791217e+01     1      0  1 -0.0422049297
#> 122  27 6.000000e+01     0      0  1 -0.0422049297
#> 123  28 2.064709e-01     1      0  0  1.9972133847
#> 124  28 3.317872e-01     1      0  0  1.9972133847
#> 125  28 6.064133e-01     1      0  0  1.9972133847
#> 126  28 7.764621e-01     1      0  0  1.9972133847
#> 127  28 8.098131e-01     1      0  0  1.9972133847
#> 128  28 8.595687e-01     0      1  0  1.9972133847
#> 129  29 4.120449e+00     1      0  1  0.4699021916
#> 130  29 5.188510e+00     1      0  1  0.4699021916
#> 131  29 3.978823e+01     0      0  1  0.4699021916
#> 132  30 1.688845e+00     1      0  0 -0.4622880573
#> 133  30 7.575417e+00     0      0  0 -0.4622880573
#> 134  31 1.188285e-01     0      1  0  1.6756969324
#> 135  32 5.464588e-01     1      0  0 -0.5739734793
#> 136  32 5.192514e+00     1      0  0 -0.5739734793
#> 137  32 1.390246e+01     0      0  0 -0.5739734793
#> 138  33 7.014546e+00     1      0  1  0.0597499374
#> 139  33 6.000000e+01     0      0  1  0.0597499374
#> 140  34 7.709748e-01     1      0  0 -0.6104616593
#> 141  34 1.147247e+01     0      0  0 -0.6104616593
#> 142  35 1.513857e+00     1      0  0 -0.1829253884
#> 143  35 2.664773e+00     1      0  0 -0.1829253884
#> 144  35 5.382902e+00     1      0  0 -0.1829253884
#> 145  35 6.996870e+00     1      0  0 -0.1829253884
#> 146  35 7.391595e+00     1      0  0 -0.1829253884
#> 147  35 9.393997e+00     1      0  0 -0.1829253884
#> 148  35 9.743130e+00     0      0  0 -0.1829253884
#> 149  36 1.284272e+00     0      1  1 -0.1009748853
#> 150  37 3.086413e+01     1      0  1 -0.3412931670
#> 151  37 3.557445e+01     1      0  1 -0.3412931670
#> 152  37 3.613090e+01     1      0  1 -0.3412931670
#> 153  37 5.736988e+01     1      0  1 -0.3412931670
#> 154  37 6.000000e+01     0      0  1 -0.3412931670
#> 155  38 2.052348e+00     1      0  1 -0.2645952385
#> 156  38 1.396352e+01     1      0  1 -0.2645952385
#> 157  38 6.000000e+01     0      0  1 -0.2645952385
#> 158  39 1.318916e+00     0      1  0 -1.7820570685
#> 159  40 4.894749e-01     1      0  1  0.4890803175
#> 160  40 2.688414e+00     1      0  1  0.4890803175
#> 161  40 5.385307e+00     1      0  1  0.4890803175
#> 162  40 5.775009e+00     1      0  1  0.4890803175
#> 163  40 1.210172e+01     1      0  1  0.4890803175
#> 164  40 2.063626e+01     1      0  1  0.4890803175
#> 165  40 3.797738e+01     0      0  1  0.4890803175
#> 166  41 3.388099e+00     0      0  1 -0.5922097459
#> 167  42 2.035214e-01     1      0  1 -0.8241093762
#> 168  42 2.792204e-01     1      0  1 -0.8241093762
#> 169  42 2.203019e+01     0      0  1 -0.8241093762
#> 170  43 3.595170e-02     1      0  0  1.4709722365
#> 171  43 1.992056e-01     1      0  0  1.4709722365
#> 172  43 2.883939e-01     1      0  0  1.4709722365
#> 173  43 3.023858e-01     1      0  0  1.4709722365
#> 174  43 3.610363e-01     1      0  0  1.4709722365
#> 175  43 3.886561e-01     1      0  0  1.4709722365
#> 176  43 5.910444e-01     1      0  0  1.4709722365
#> 177  43 7.371503e-01     0      1  0  1.4709722365
#> 178  44 1.280800e-01     1      0  1  0.1680653839
#> 179  44 6.546102e-01     1      0  1  0.1680653839
#> 180  44 7.366691e-01     1      0  1  0.1680653839
#> 181  44 8.632764e-01     1      0  1  0.1680653839
#> 182  44 8.994609e-01     1      0  1  0.1680653839
#> 183  44 5.296525e+00     1      0  1  0.1680653839
#> 184  44 5.458895e+00     1      0  1  0.1680653839
#> 185  44 8.914203e+00     1      0  1  0.1680653839
#> 186  44 3.380559e+01     0      0  1  0.1680653839
#> 187  45 4.501981e-01     1      0  0 -1.0525132793
#> 188  45 3.051415e+01     0      0  0 -1.0525132793
#> 189  46 4.477785e-01     1      0  1  2.4595680103
#> 190  46 5.410554e-01     1      0  1  2.4595680103
#> 191  46 7.716608e-01     1      0  1  2.4595680103
#> 192  46 1.545604e+00     1      0  1  2.4595680103
#> 193  46 1.644625e+00     0      1  1  2.4595680103
#> 194  47 1.298373e-01     1      0  0  1.0384404773
#> 195  47 2.937335e-01     1      0  0  1.0384404773
#> 196  47 8.133530e-01     1      0  0  1.0384404773
#> 197  47 1.030110e+00     1      0  0  1.0384404773
#> 198  47 1.584591e+00     1      0  0  1.0384404773
#> 199  47 1.671892e+00     1      0  0  1.0384404773
#> 200  47 1.970031e+00     1      0  0  1.0384404773
#> 201  47 2.003126e+00     1      0  0  1.0384404773
#> 202  47 2.139827e+00     1      0  0  1.0384404773
#> 203  47 2.425079e+00     1      0  0  1.0384404773
#> 204  47 3.339542e+00     0      1  0  1.0384404773
#> 205  48 1.471070e-02     1      0  0  0.2912296583
#> 206  48 1.691857e-02     1      0  0  0.2912296583
#> 207  48 5.312277e-01     1      0  0  0.2912296583
#> 208  48 9.824216e-01     1      0  0  0.2912296583
#> 209  48 1.492613e+00     1      0  0  0.2912296583
#> 210  48 4.762075e+00     1      0  0  0.2912296583
#> 211  48 9.892954e+00     0      0  0  0.2912296583
#> 212  49 6.002807e-02     1      0  0 -1.0061661467
#> 213  49 2.964007e+01     1      0  0 -1.0061661467
#> 214  49 5.323309e+01     0      0  0 -1.0061661467
#> 215  50 5.971561e-01     1      0  1  0.2205488301
#> 216  50 2.288330e+00     1      0  1  0.2205488301
#> 217  50 3.985033e+00     1      0  1  0.2205488301
#> 218  50 4.233803e+00     1      0  1  0.2205488301
#> 219  50 3.433694e+01     1      0  1  0.2205488301
#> 220  50 6.000000e+01     0      0  1  0.2205488301
#> 221  51 1.165329e+00     1      0  0  0.0993275941
#> 222  51 2.329678e+00     0      1  0  0.0993275941
#> 223  52 5.781521e-01     1      0  1 -0.3698112355
#> 224  52 8.944408e-01     1      0  1 -0.3698112355
#> 225  52 1.903687e+01     1      0  1 -0.3698112355
#> 226  52 4.402062e+01     0      0  1 -0.3698112355
#> 227  53 9.682294e-01     1      0  1  0.3339029425
#> 228  53 1.116831e+00     1      0  1  0.3339029425
#> 229  53 2.189989e+00     1      0  1  0.3339029425
#> 230  53 3.823156e+00     1      0  1  0.3339029425
#> 231  53 1.915456e+01     0      0  1  0.3339029425
#> 232  54 1.308044e-04     1      0  0  1.1497782392
#> 233  54 2.830484e-01     1      0  0  1.1497782392
#> 234  54 1.181644e+00     1      0  0  1.1497782392
#> 235  54 1.249957e+00     1      0  0  1.1497782392
#> 236  54 1.627030e+00     1      0  0  1.1497782392
#> 237  54 1.650370e+00     1      0  0  1.1497782392
#> 238  54 1.910316e+00     1      0  0  1.1497782392
#> 239  54 3.111740e+00     1      0  0  1.1497782392
#> 240  54 3.373706e+00     1      0  0  1.1497782392
#> 241  54 3.495092e+00     1      0  0  1.1497782392
#> 242  54 3.581320e+00     0      0  0  1.1497782392
#> 243  55 2.102608e-01     0      1  0  0.1444757047
#> 244  56 2.340154e+00     1      0  1 -1.0618481450
#> 245  56 2.005843e+01     1      0  1 -1.0618481450
#> 246  56 6.000000e+01     0      0  1 -1.0618481450
#> 247  57 8.300327e-01     1      0  0 -0.6720418220
#> 248  57 2.733980e+00     1      0  0 -0.6720418220
#> 249  57 5.980115e+00     1      0  0 -0.6720418220
#> 250  57 3.966673e+01     0      0  0 -0.6720418220
#> 251  58 2.256326e+01     0      0  1 -2.0798811914
#> 252  59 1.771419e+01     1      0  0 -0.5811633734
#> 253  59 2.190380e+01     0      0  0 -0.5811633734
#> 254  60 4.618214e+00     1      0  0 -0.0288415529
#> 255  60 5.867692e+00     1      0  0 -0.0288415529
#> 256  60 1.138758e+01     1      0  0 -0.0288415529
#> 257  60 1.201823e+01     1      0  0 -0.0288415529
#> 258  60 1.481138e+01     0      1  0 -0.0288415529
#> 259  61 8.775919e-02     1      0  0  0.2272919217
#> 260  61 1.107826e-01     1      0  0  0.2272919217
#> 261  61 2.681293e+00     1      0  0  0.2272919217
#> 262  61 1.132327e+01     0      0  0  0.2272919217
#> 263  62 8.367194e-01     1      0  1 -0.0929410185
#> 264  62 1.935791e+00     1      0  1 -0.0929410185
#> 265  62 9.439486e+00     1      0  1 -0.0929410185
#> 266  62 1.998069e+01     0      0  1 -0.0929410185
#> 267  63 7.631447e-02     0      1  1  0.6320840773
#> 268  64 3.967656e-02     0      1  0  0.6019578201
#> 269  65 1.386625e-01     1      0  0  1.2213407736
#> 270  65 4.510757e-01     1      0  0  1.2213407736
#> 271  65 1.004785e+00     0      1  0  1.2213407736
#> 272  66 6.667340e-01     0      1  1 -0.9550782863
#> 273  67 2.217521e-01     1      0  1  0.3367575683
#> 274  67 1.483347e+00     1      0  1  0.3367575683
#> 275  67 6.179839e+00     0      0  1  0.3367575683
#> 276  68 3.355013e+00     1      0  1  0.1168809296
#> 277  68 3.110002e+01     0      0  1  0.1168809296
#> 278  69 1.249253e-01     0      1  1  1.4304023412
#> 279  70 2.783390e-01     1      0  0  0.9171749176
#> 280  70 5.521823e-01     1      0  0  0.9171749176
#> 281  70 7.557697e-01     1      0  0  0.9171749176
#> 282  70 2.246774e+00     0      1  0  0.9171749176
#> 283  71 1.902617e-02     1      0  0  1.0651543971
#> 284  71 3.160541e-01     1      0  0  1.0651543971
#> 285  71 4.518616e-01     1      0  0  1.0651543971
#> 286  71 4.876935e-01     1      0  0  1.0651543971
#> 287  71 5.379353e-01     1      0  0  1.0651543971
#> 288  71 5.673667e-01     1      0  0  1.0651543971
#> 289  71 5.718082e-01     1      0  0  1.0651543971
#> 290  71 6.466041e-01     1      0  0  1.0651543971
#> 291  71 1.562482e+00     1      0  0  1.0651543971
#> 292  71 2.423405e+00     1      0  0  1.0651543971
#> 293  71 3.294035e+00     1      0  0  1.0651543971
#> 294  71 4.155037e+00     1      0  0  1.0651543971
#> 295  71 5.510065e+00     1      0  0  1.0651543971
#> 296  71 6.614586e+00     1      0  0  1.0651543971
#> 297  71 6.774451e+00     1      0  0  1.0651543971
#> 298  71 9.301806e+00     1      0  0  1.0651543971
#> 299  71 9.356054e+00     1      0  0  1.0651543971
#> 300  71 9.432119e+00     1      0  0  1.0651543971
#> 301  71 1.106919e+01     0      0  0  1.0651543971
#> 302  72 2.046306e-01     1      0  0  0.0204507085
#> 303  72 2.924664e+00     1      0  0  0.0204507085
#> 304  72 4.173461e+00     1      0  0  0.0204507085
#> 305  72 7.395683e+00     0      0  0  0.0204507085
#> 306  73 2.153635e-01     0      1  1 -0.6146030671
#> 307  74 2.283606e+00     0      1  1  0.5583362028
#> 308  75 1.886979e+00     1      0  1 -0.1682707588
#> 309  75 3.559939e+00     1      0  1 -0.1682707588
#> 310  75 2.805041e+01     0      1  1 -0.1682707588
#> 311  76 3.686094e+00     1      0  1 -1.2328060390
#> 312  76 1.126403e+01     0      0  1 -1.2328060390
#> 313  77 2.961239e+00     0      1  1 -1.5091665373
#> 314  78 7.641721e-02     1      0  0  0.1501201312
#> 315  78 8.609022e-01     1      0  0  0.1501201312
#> 316  78 1.432486e+00     1      0  0  0.1501201312
#> 317  78 2.688382e+00     1      0  0  0.1501201312
#> 318  78 2.956192e+00     1      0  0  0.1501201312
#> 319  78 4.776043e+00     1      0  0  0.1501201312
#> 320  78 1.098598e+01     1      0  0  0.1501201312
#> 321  78 2.116396e+01     1      0  0  0.1501201312
#> 322  78 2.361179e+01     1      0  0  0.1501201312
#> 323  78 2.464964e+01     1      0  0  0.1501201312
#> 324  78 2.660071e+01     0      0  0  0.1501201312
#> 325  79 6.782495e-01     1      0  0 -0.1843088688
#> 326  79 5.661984e+00     1      0  0 -0.1843088688
#> 327  79 1.319537e+01     1      0  0 -0.1843088688
#> 328  79 1.999769e+01     0      0  0 -0.1843088688
#> 329  80 3.550196e-02     1      0  0  2.4380030474
#> 330  80 4.999968e-02     1      0  0  2.4380030474
#> 331  80 1.253017e-01     1      0  0  2.4380030474
#> 332  80 2.157615e-01     1      0  0  2.4380030474
#> 333  80 2.655092e-01     1      0  0  2.4380030474
#> 334  80 2.985268e-01     0      1  0  2.4380030474
#> 335  81 4.099906e-01     1      0  1  0.5626112714
#> 336  81 7.215896e-01     1      0  1  0.5626112714
#> 337  81 2.146474e+00     1      0  1  0.5626112714
#> 338  81 5.473858e+01     1      0  1  0.5626112714
#> 339  81 6.000000e+01     0      0  1  0.5626112714
#> 340  82 4.352866e+00     1      0  0 -0.4686997798
#> 341  82 8.729673e+00     1      0  0 -0.4686997798
#> 342  82 1.556014e+01     0      0  0 -0.4686997798
#> 343  83 7.465105e-02     1      0  1  2.4302266523
#> 344  83 1.348373e-01     1      0  1  2.4302266523
#> 345  83 4.248119e-01     1      0  1  2.4302266523
#> 346  83 4.348714e-01     1      0  1  2.4302266523
#> 347  83 4.691368e-01     1      0  1  2.4302266523
#> 348  83 5.518471e-01     1      0  1  2.4302266523
#> 349  83 1.027768e+00     1      0  1  2.4302266523
#> 350  83 1.192434e+00     1      0  1  2.4302266523
#> 351  83 1.441798e+00     1      0  1  2.4302266523
#> 352  83 1.946165e+00     1      0  1  2.4302266523
#> 353  83 1.994897e+00     1      0  1  2.4302266523
#> 354  83 2.806140e+00     1      0  1  2.4302266523
#> 355  83 3.793906e+00     1      0  1  2.4302266523
#> 356  83 4.051319e+00     1      0  1  2.4302266523
#> 357  83 4.657390e+00     1      0  1  2.4302266523
#> 358  83 5.764010e+00     1      0  1  2.4302266523
#> 359  83 8.879082e+00     1      0  1  2.4302266523
#> 360  83 1.224971e+01     1      0  1  2.4302266523
#> 361  83 1.330288e+01     1      0  1  2.4302266523
#> 362  83 1.401462e+01     1      0  1  2.4302266523
#> 363  83 1.425998e+01     1      0  1  2.4302266523
#> 364  83 1.508976e+01     1      0  1  2.4302266523
#> 365  83 1.609150e+01     1      0  1  2.4302266523
#> 366  83 1.877439e+01     1      0  1  2.4302266523
#> 367  83 2.087830e+01     1      0  1  2.4302266523
#> 368  83 2.104478e+01     1      0  1  2.4302266523
#> 369  83 2.306145e+01     1      0  1  2.4302266523
#> 370  83 2.504936e+01     1      0  1  2.4302266523
#> 371  83 2.948564e+01     1      0  1  2.4302266523
#> 372  83 3.886133e+01     1      0  1  2.4302266523
#> 373  83 3.958569e+01     1      0  1  2.4302266523
#> 374  83 4.082450e+01     1      0  1  2.4302266523
#> 375  83 4.212935e+01     1      0  1  2.4302266523
#> 376  83 4.372314e+01     0      1  1  2.4302266523
#> 377  84 1.088507e+00     1      0  1  1.3700039940
#> 378  84 1.196182e+00     1      0  1  1.3700039940
#> 379  84 2.319977e+00     1      0  1  1.3700039940
#> 380  84 3.156876e+00     0      1  1  1.3700039940
#> 381  85 1.335003e+00     1      0  0 -0.7014784203
#> 382  85 1.050566e+01     1      0  0 -0.7014784203
#> 383  85 1.423637e+01     1      0  0 -0.7014784203
#> 384  85 1.672756e+01     1      0  0 -0.7014784203
#> 385  85 2.663004e+01     1      0  0 -0.7014784203
#> 386  85 6.000000e+01     0      0  0 -0.7014784203
#> 387  86 3.369515e-02     1      0  0  0.3448499357
#> 388  86 6.054668e-01     1      0  0  0.3448499357
#> 389  86 9.694354e-01     1      0  0  0.3448499357
#> 390  86 1.257804e+00     1      0  0  0.3448499357
#> 391  86 1.413719e+00     1      0  0  0.3448499357
#> 392  86 4.629134e+00     1      0  0  0.3448499357
#> 393  86 6.605761e+00     0      1  0  0.3448499357
#> 394  87 3.835246e-01     1      0  0  1.1811808914
#> 395  87 7.815277e-01     1      0  0  1.1811808914
#> 396  87 9.445864e-01     1      0  0  1.1811808914
#> 397  87 9.699292e-01     1      0  0  1.1811808914
#> 398  87 1.426861e+00     1      0  0  1.1811808914
#> 399  87 1.447333e+00     1      0  0  1.1811808914
#> 400  87 1.952061e+00     1      0  0  1.1811808914
#> 401  87 2.238947e+00     1      0  0  1.1811808914
#> 402  87 3.382030e+00     1      0  0  1.1811808914
#> 403  87 3.546151e+00     1      0  0  1.1811808914
#> 404  87 5.873190e+00     1      0  0  1.1811808914
#> 405  87 8.009980e+00     0      0  0  1.1811808914
#> 406  88 3.712288e+00     0      0  1 -0.7518730444
#> 407  89 4.545933e-01     1      0  0  0.2242182846
#> 408  89 2.698230e+00     1      0  0  0.2242182846
#> 409  89 4.471116e+00     1      0  0  0.2242182846
#> 410  89 5.045412e+00     1      0  0  0.2242182846
#> 411  89 7.615919e+00     1      0  0  0.2242182846
#> 412  89 8.116822e+00     1      0  0  0.2242182846
#> 413  89 3.072340e+01     1      0  0  0.2242182846
#> 414  89 6.000000e+01     0      0  0  0.2242182846
#> 415  90 3.495267e-01     1      0  0  0.9411658067
#> 416  90 3.646775e-01     1      0  0  0.9411658067
#> 417  90 4.113572e-01     1      0  0  0.9411658067
#> 418  90 1.431755e+00     1      0  0  0.9411658067
#> 419  90 2.295295e+00     1      0  0  0.9411658067
#> 420  90 2.567016e+00     1      0  0  0.9411658067
#> 421  90 4.063570e+00     0      0  0  0.9411658067
#> 422  91 4.237787e-02     1      0  0  1.5327183192
#> 423  91 7.046592e-02     1      0  0  1.5327183192
#> 424  91 1.844680e-01     1      0  0  1.5327183192
#> 425  91 2.569287e-01     0      1  0  1.5327183192
#> 426  92 1.498413e-01     1      0  0  1.0484328670
#> 427  92 2.488011e-01     1      0  0  1.0484328670
#> 428  92 2.705576e-01     1      0  0  1.0484328670
#> 429  92 4.198920e-01     0      1  0  1.0484328670
#> 430  93 1.163759e-01     1      0  0  0.8993540532
#> 431  93 2.276013e-01     1      0  0  0.8993540532
#> 432  93 4.090842e-01     1      0  0  0.8993540532
#> 433  93 4.492284e-01     1      0  0  0.8993540532
#> 434  93 9.823614e-01     0      1  0  0.8993540532
#> 435  94 1.208222e+00     0      1  0  0.1071904070
#> 436  95 7.895186e-01     1      0  1  0.4806255999
#> 437  95 5.109233e+00     1      0  1  0.4806255999
#> 438  95 8.766021e+00     1      0  1  0.4806255999
#> 439  95 1.679035e+01     1      0  1  0.4806255999
#> 440  95 1.828670e+01     1      0  1  0.4806255999
#> 441  95 6.000000e+01     0      0  1  0.4806255999
#> 442  96 6.000000e+01     0      0  1 -0.4186039231
#> 443  97 1.910628e+00     1      0  1  0.0315260046
#> 444  97 4.590223e+00     1      0  1  0.0315260046
#> 445  97 9.771979e+00     0      0  1  0.0315260046
#> 446  98 1.804677e-01     1      0  1 -0.5179824302
#> 447  98 8.415075e-01     1      0  1 -0.5179824302
#> 448  98 6.000000e+01     0      0  1 -0.5179824302
#> 449  99 6.000000e+01     0      0  1 -1.1198997214
#> 450 100 1.055845e-01     1      0  0 -0.1596039606
#> 451 100 1.787203e+00     1      0  0 -0.1596039606
#> 452 100 7.099397e+00     1      0  0 -0.1596039606
#> 453 100 1.399412e+01     1      0  0 -0.1596039606
#> 454 100 2.665713e+01     0      0  0 -0.1596039606
#> 455 101 1.338255e-02     1      0  1  0.2886087909
#> 456 101 1.636915e+00     1      0  1  0.2886087909
#> 457 101 2.622406e+01     1      0  1  0.2886087909
#> 458 101 3.009257e+01     1      0  1  0.2886087909
#> 459 101 3.145696e+01     1      0  1  0.2886087909
#> 460 101 3.882307e+01     0      0  1  0.2886087909
#> 461 102 2.027554e-01     1      0  0  0.8482544971
#> 462 102 2.182339e-01     1      0  0  0.8482544971
#> 463 102 3.399400e-01     1      0  0  0.8482544971
#> 464 102 3.723308e-01     1      0  0  0.8482544971
#> 465 102 1.039127e+00     1      0  0  0.8482544971
#> 466 102 1.250331e+00     1      0  0  0.8482544971
#> 467 102 1.338264e+00     1      0  0  0.8482544971
#> 468 102 1.861378e+00     1      0  0  0.8482544971
#> 469 102 2.215888e+00     1      0  0  0.8482544971
#> 470 102 3.016425e+00     1      0  0  0.8482544971
#> 471 102 3.271995e+00     1      0  0  0.8482544971
#> 472 102 3.335439e+00     1      0  0  0.8482544971
#> 473 102 4.769084e+00     1      0  0  0.8482544971
#> 474 102 5.996935e+00     1      0  0  0.8482544971
#> 475 102 6.949052e+00     1      0  0  0.8482544971
#> 476 102 8.040548e+00     1      0  0  0.8482544971
#> 477 102 1.079130e+01     1      0  0  0.8482544971
#> 478 102 1.430891e+01     1      0  0  0.8482544971
#> 479 102 1.529306e+01     1      0  0  0.8482544971
#> 480 102 1.659045e+01     1      0  0  0.8482544971
#> 481 102 2.190191e+01     0      0  0  0.8482544971
#> 482 103 1.437049e+01     0      0  0 -1.7283045153
#> 483 104 2.376551e-01     1      0  0  0.0016836885
#> 484 104 3.263867e+00     1      0  0  0.0016836885
#> 485 104 5.928129e+00     0      0  0  0.0016836885
#> 486 105 4.372712e-01     1      0  0  1.7386337669
#> 487 105 4.493712e-01     1      0  0  1.7386337669
#> 488 105 5.844613e-01     1      0  0  1.7386337669
#> 489 105 9.103945e-01     1      0  0  1.7386337669
#> 490 105 9.902673e-01     1      0  0  1.7386337669
#> 491 105 1.226611e+00     1      0  0  1.7386337669
#> 492 105 1.435189e+00     1      0  0  1.7386337669
#> 493 105 1.451876e+00     1      0  0  1.7386337669
#> 494 105 1.590927e+00     1      0  0  1.7386337669
#> 495 105 1.598563e+00     1      0  0  1.7386337669
#> 496 105 1.636353e+00     1      0  0  1.7386337669
#> 497 105 1.794146e+00     1      0  0  1.7386337669
#> 498 105 3.029779e+00     1      0  0  1.7386337669
#> 499 105 3.222424e+00     1      0  0  1.7386337669
#> 500 105 3.590796e+00     1      0  0  1.7386337669
#> 501 105 3.849583e+00     1      0  0  1.7386337669
#> 502 105 3.909253e+00     1      0  0  1.7386337669
#> 503 105 4.047626e+00     1      0  0  1.7386337669
#> 504 105 4.898231e+00     1      0  0  1.7386337669
#> 505 105 4.986705e+00     1      0  0  1.7386337669
#> 506 105 5.530053e+00     1      0  0  1.7386337669
#> 507 105 6.306516e+00     1      0  0  1.7386337669
#> 508 105 7.467007e+00     1      0  0  1.7386337669
#> 509 105 1.082009e+01     1      0  0  1.7386337669
#> 510 105 1.215195e+01     1      0  0  1.7386337669
#> 511 105 1.357755e+01     0      1  0  1.7386337669
#> 512 106 1.184513e+01     1      0  1  0.5397906055
#> 513 106 2.774437e+01     0      0  1  0.5397906055
#> 514 107 1.118586e+00     1      0  0 -0.7728898441
#> 515 107 6.672145e+00     1      0  0 -0.7728898441
#> 516 107 1.168604e+01     0      0  0 -0.7728898441
#> 517 108 6.735287e-01     1      0  1  0.7505261239
#> 518 108 6.345176e+00     0      0  1  0.7505261239
#> 519 109 4.413963e-01     1      0  0  0.2796278624
#> 520 109 6.547517e-01     1      0  0  0.2796278624
#> 521 109 1.266021e+00     1      0  0  0.2796278624
#> 522 109 1.839985e+00     0      0  0  0.2796278624
#> 523 110 2.640627e+00     1      0  1  0.8134003741
#> 524 110 2.682352e+00     1      0  1  0.8134003741
#> 525 110 4.206418e+00     1      0  1  0.8134003741
#> 526 110 9.272456e+00     1      0  1  0.8134003741
#> 527 110 3.805413e+01     0      0  1  0.8134003741
#> 528 111 4.317927e+01     0      0  1 -1.8863251587
#> 529 112 4.331024e-01     1      0  0  0.3333699702
#> 530 112 2.255705e+00     1      0  0  0.3333699702
#> 531 112 2.476021e+00     1      0  0  0.3333699702
#> 532 112 5.655183e+00     1      0  0  0.3333699702
#> 533 112 6.911556e+00     1      0  0  0.3333699702
#> 534 112 9.543750e+00     0      1  0  0.3333699702
#> 535 113 3.846896e+01     0      0  1 -1.1512125827
#> 536 114 6.181427e+00     1      0  0 -1.1358527745
#> 537 114 9.078751e+00     1      0  0 -1.1358527745
#> 538 114 3.238296e+01     0      0  0 -1.1358527745
#> 539 115 1.456670e-01     1      0  0 -0.7417085823
#> 540 115 4.404603e-01     1      0  0 -0.7417085823
#> 541 115 1.078736e+00     1      0  0 -0.7417085823
#> 542 115 7.899102e+00     1      0  0 -0.7417085823
#> 543 115 1.130588e+01     1      0  0 -0.7417085823
#> 544 115 1.301153e+01     1      0  0 -0.7417085823
#> 545 115 2.893480e+01     1      0  0 -0.7417085823
#> 546 115 6.000000e+01     0      0  0 -0.7417085823
#> 547 116 2.100286e-01     1      0  0  0.7422970242
#> 548 116 3.010440e-01     0      1  0  0.7422970242
#> 549 117 1.751614e-02     1      0  0 -0.2983660992
#> 550 117 2.322303e-01     1      0  0 -0.2983660992
#> 551 117 8.094322e-01     1      0  0 -0.2983660992
#> 552 117 1.286717e+00     0      0  0 -0.2983660992
#> 553 118 2.350354e-01     1      0  1 -0.2779718871
#> 554 118 5.033004e+01     0      0  1 -0.2779718871
#> 555 119 3.054819e-01     1      0  0 -0.8186697775
#> 556 119 1.344880e+00     0      0  0 -0.8186697775
#> 557 120 3.477751e+01     1      0  1  0.0979634986
#> 558 120 3.710871e+01     1      0  1  0.0979634986
#> 559 120 4.737750e+01     1      0  1  0.0979634986
#> 560 120 6.000000e+01     0      0  1  0.0979634986
#> 561 121 6.032077e-01     1      0  1 -0.0260186348
#> 562 121 2.133581e+01     1      0  1 -0.0260186348
#> 563 121 3.946915e+01     1      0  1 -0.0260186348
#> 564 121 4.397155e+01     0      0  1 -0.0260186348
#> 565 122 2.626123e-01     1      0  1  1.5212967765
#> 566 122 1.354315e+00     1      0  1  1.5212967765
#> 567 122 3.572909e+00     1      0  1  1.5212967765
#> 568 122 3.927397e+00     1      0  1  1.5212967765
#> 569 122 4.101006e+00     1      0  1  1.5212967765
#> 570 122 4.139498e+00     1      0  1  1.5212967765
#> 571 122 4.256945e+00     1      0  1  1.5212967765
#> 572 122 6.512340e+00     1      0  1  1.5212967765
#> 573 122 6.769281e+00     1      0  1  1.5212967765
#> 574 122 9.296487e+00     1      0  1  1.5212967765
#> 575 122 1.997625e+01     1      0  1  1.5212967765
#> 576 122 2.376927e+01     1      0  1  1.5212967765
#> 577 122 2.594241e+01     0      0  1  1.5212967765
#> 578 123 6.534991e-01     0      0  0 -0.1632849322
#> 579 124 1.540023e-02     0      1  1  1.6659909000
#> 580 125 6.582040e-02     1      0  0  0.7795210873
#> 581 125 3.927864e-01     1      0  0  0.7795210873
#> 582 125 2.200457e+00     1      0  0  0.7795210873
#> 583 125 3.737620e+00     1      0  0  0.7795210873
#> 584 125 3.874151e+00     0      1  0  0.7795210873
#> 585 126 2.498341e+00     1      0  1  0.5412841435
#> 586 126 3.105278e+00     1      0  1  0.5412841435
#> 587 126 6.633556e+00     1      0  1  0.5412841435
#> 588 126 1.392007e+01     1      0  1  0.5412841435
#> 589 126 1.802679e+01     1      0  1  0.5412841435
#> 590 126 2.115515e+01     0      0  1  0.5412841435
#> 591 127 2.961042e-01     1      0  0  0.2880705915
#> 592 127 1.199334e+00     1      0  0  0.2880705915
#> 593 127 1.829783e+00     1      0  0  0.2880705915
#> 594 127 3.342012e+00     1      0  0  0.2880705915
#> 595 127 3.622788e+00     0      1  0  0.2880705915
#> 596 128 1.804132e-01     1      0  1  1.7348367335
#> 597 128 8.238773e-01     1      0  1  1.7348367335
#> 598 128 1.233602e+00     0      0  1  1.7348367335
#> 599 129 2.397373e-01     1      0  1  0.2084265523
#> 600 129 1.380228e+00     1      0  1  0.2084265523
#> 601 129 2.216070e+00     0      0  1  0.2084265523
#> 602 130 1.427262e-02     1      0  0 -0.0855064444
#> 603 130 2.922311e-01     1      0  0 -0.0855064444
#> 604 130 7.337143e-01     1      0  0 -0.0855064444
#> 605 130 1.038981e+00     0      1  0 -0.0855064444
#> 606 131 1.240149e+00     1      0  0 -1.9251452518
#> 607 131 9.606735e+00     0      0  0 -1.9251452518
#> 608 132 4.773581e+01     0      0  0 -2.1622590028
#> 609 133 1.770162e-01     1      0  0  0.2008174430
#> 610 133 2.293438e+00     1      0  0  0.2008174430
#> 611 133 2.508643e+00     1      0  0  0.2008174430
#> 612 133 5.076847e+00     1      0  0  0.2008174430
#> 613 133 5.993603e+00     1      0  0  0.2008174430
#> 614 133 7.875250e+00     0      1  0  0.2008174430
#> 615 134 2.705448e-01     1      0  0 -0.6868156524
#> 616 134 4.796078e+00     1      0  0 -0.6868156524
#> 617 134 1.683072e+01     1      0  0 -0.6868156524
#> 618 134 5.022011e+01     0      0  0 -0.6868156524
#> 619 135 9.608403e-02     1      0  0 -0.4577605332
#> 620 135 1.976128e+00     1      0  0 -0.4577605332
#> 621 135 2.476301e+00     0      0  0 -0.4577605332
#> 622 136 6.067891e-01     1      0  1  0.0898024702
#> 623 136 1.021315e+00     1      0  1  0.0898024702
#> 624 136 2.560214e+00     1      0  1  0.0898024702
#> 625 136 4.028010e+00     1      0  1  0.0898024702
#> 626 136 5.681961e+00     1      0  1  0.0898024702
#> 627 136 4.733856e+01     0      0  1  0.0898024702
#> 628 137 1.536920e+00     1      0  0  0.2207112910
#> 629 137 1.947600e+00     1      0  0  0.2207112910
#> 630 137 2.841556e+00     1      0  0  0.2207112910
#> 631 137 6.162221e+00     1      0  0  0.2207112910
#> 632 137 7.313439e+00     0      0  0  0.2207112910
#> 633 138 5.094094e-03     1      0  1  2.2996193612
#> 634 138 3.692285e-01     1      0  1  2.2996193612
#> 635 138 3.836658e-01     1      0  1  2.2996193612
#> 636 138 4.420170e-01     1      0  1  2.2996193612
#> 637 138 1.113684e+00     1      0  1  2.2996193612
#> 638 138 2.172978e+00     1      0  1  2.2996193612
#> 639 138 3.967154e+00     1      0  1  2.2996193612
#> 640 138 5.240364e+00     1      0  1  2.2996193612
#> 641 138 6.043040e+00     1      0  1  2.2996193612
#> 642 138 6.549664e+00     1      0  1  2.2996193612
#> 643 138 6.594723e+00     1      0  1  2.2996193612
#> 644 138 7.471264e+00     0      0  1  2.2996193612
#> 645 139 1.061196e-01     1      0  0  0.4353602679
#> 646 139 3.764681e-01     1      0  0  0.4353602679
#> 647 139 7.396117e-01     1      0  0  0.4353602679
#> 648 139 1.802767e+00     1      0  0  0.4353602679
#> 649 139 2.323746e+00     1      0  0  0.4353602679
#> 650 139 4.527143e+00     1      0  0  0.4353602679
#> 651 139 6.520688e+00     1      0  0  0.4353602679
#> 652 139 6.809317e+00     1      0  0  0.4353602679
#> 653 139 1.222438e+01     1      0  0  0.4353602679
#> 654 139 1.261672e+01     1      0  0  0.4353602679
#> 655 139 1.916153e+01     1      0  0  0.4353602679
#> 656 139 2.018319e+01     0      0  0  0.4353602679
#> 657 140 8.799912e-01     1      0  1 -1.7017649951
#> 658 140 8.246044e+00     0      0  1 -1.7017649951
#> 659 141 6.573989e-01     1      0  0 -1.0585905359
#> 660 141 3.913478e+00     0      0  0 -1.0585905359
#> 661 142 1.139778e+01     1      0  1 -0.1225113379
#> 662 142 4.187066e+01     0      0  1 -0.1225113379
#> 663 143 3.637194e+01     1      0  1 -0.5226166972
#> 664 143 3.892848e+01     1      0  1 -0.5226166972
#> 665 143 6.000000e+01     0      0  1 -0.5226166972
#> 666 144 8.870903e-01     1      0  1 -0.7660132673
#> 667 144 1.780134e+00     1      0  1 -0.7660132673
#> 668 144 4.575192e+01     0      0  1 -0.7660132673
#> 669 145 1.191568e-01     1      0  1  2.7973911470
#> 670 145 1.624827e-01     1      0  1  2.7973911470
#> 671 145 1.768642e-01     1      0  1  2.7973911470
#> 672 145 1.921042e-01     1      0  1  2.7973911470
#> 673 145 2.100185e-01     1      0  1  2.7973911470
#> 674 145 2.273465e-01     0      1  1  2.7973911470
#> 675 146 5.134183e-02     1      0  0  1.6567240533
#> 676 146 1.095822e-01     1      0  0  1.6567240533
#> 677 146 3.407789e-01     1      0  0  1.6567240533
#> 678 146 8.474204e-01     1      0  0  1.6567240533
#> 679 146 9.863356e-01     1      0  0  1.6567240533
#> 680 146 1.007823e+00     1      0  0  1.6567240533
#> 681 146 1.296521e+00     1      0  0  1.6567240533
#> 682 146 1.368934e+00     1      0  0  1.6567240533
#> 683 146 1.500835e+00     1      0  0  1.6567240533
#> 684 146 1.630313e+00     1      0  0  1.6567240533
#> 685 146 1.718344e+00     1      0  0  1.6567240533
#> 686 146 2.452915e+00     1      0  0  1.6567240533
#> 687 146 2.464934e+00     0      1  0  1.6567240533
#> 688 147 5.664918e-01     1      0  0 -2.3159235182
#> 689 147 3.550825e+00     1      0  0 -2.3159235182
#> 690 147 2.315634e+01     0      0  0 -2.3159235182
#> 691 148 2.938262e-01     1      0  1  0.7309635162
#> 692 148 4.268762e-01     0      1  1  0.7309635162
#> 693 149 6.336934e-01     1      0  0 -0.0889305675
#> 694 149 1.539535e+00     0      1  0 -0.0889305675
#> 695 150 7.837320e-01     1      0  1 -0.4944808755
#> 696 150 1.729275e+01     0      0  1 -0.4944808755
#> 697 151 1.819744e-01     1      0  0  1.7962018886
#> 698 151 2.319851e-01     1      0  0  1.7962018886
#> 699 151 3.184299e-01     1      0  0  1.7962018886
#> 700 151 5.649650e-01     1      0  0  1.7962018886
#> 701 151 6.786861e-01     1      0  0  1.7962018886
#> 702 151 1.346138e+00     1      0  0  1.7962018886
#> 703 151 1.429929e+00     1      0  0  1.7962018886
#> 704 151 1.571333e+00     1      0  0  1.7962018886
#> 705 151 1.611561e+00     1      0  0  1.7962018886
#> 706 151 1.934020e+00     1      0  0  1.7962018886
#> 707 151 2.008421e+00     1      0  0  1.7962018886
#> 708 151 2.037024e+00     1      0  0  1.7962018886
#> 709 151 2.436826e+00     1      0  0  1.7962018886
#> 710 151 2.919518e+00     1      0  0  1.7962018886
#> 711 151 3.035968e+00     1      0  0  1.7962018886
#> 712 151 3.077781e+00     1      0  0  1.7962018886
#> 713 151 3.525384e+00     1      0  0  1.7962018886
#> 714 151 3.536970e+00     1      0  0  1.7962018886
#> 715 151 3.692956e+00     1      0  0  1.7962018886
#> 716 151 4.463901e+00     1      0  0  1.7962018886
#> 717 151 4.992467e+00     1      0  0  1.7962018886
#> 718 151 6.925206e+00     1      0  0  1.7962018886
#> 719 151 8.055285e+00     1      0  0  1.7962018886
#> 720 151 9.224220e+00     0      1  0  1.7962018886
#> 721 152 2.238315e-01     1      0  1 -0.2859709111
#> 722 152 9.150560e+00     1      0  1 -0.2859709111
#> 723 152 2.484763e+01     0      0  1 -0.2859709111
#> 724 153 1.029782e-01     1      0  1  0.9279480567
#> 725 153 1.217614e-01     1      0  1  0.9279480567
#> 726 153 1.099042e+00     0      1  1  0.9279480567
#> 727 154 2.363313e+00     1      0  0 -0.8915215744
#> 728 154 6.752486e+00     1      0  0 -0.8915215744
#> 729 154 1.924679e+01     0      0  0 -0.8915215744
#> 730 155 6.061902e-02     1      0  0  1.7699041099
#> 731 155 1.060879e-01     1      0  0  1.7699041099
#> 732 155 1.789783e-01     1      0  0  1.7699041099
#> 733 155 2.294163e-01     1      0  0  1.7699041099
#> 734 155 2.658812e-01     1      0  0  1.7699041099
#> 735 155 4.709482e-01     1      0  0  1.7699041099
#> 736 155 5.140956e-01     1      0  0  1.7699041099
#> 737 155 5.855506e-01     1      0  0  1.7699041099
#> 738 155 6.365492e-01     1      0  0  1.7699041099
#> 739 155 6.622208e-01     1      0  0  1.7699041099
#> 740 155 8.578373e-01     1      0  0  1.7699041099
#> 741 155 1.095457e+00     1      0  0  1.7699041099
#> 742 155 1.212917e+00     1      0  0  1.7699041099
#> 743 155 1.296140e+00     1      0  0  1.7699041099
#> 744 155 1.411733e+00     1      0  0  1.7699041099
#> 745 155 1.641214e+00     0      1  0  1.7699041099
#> 746 156 1.393447e+00     1      0  1  0.2498516592
#> 747 156 7.644374e+00     0      1  1  0.2498516592
#> 748 157 4.013308e+00     0      0  1 -0.9855391126
#> 749 158 3.251263e-01     1      0  0  0.1612263513
#> 750 158 4.669904e+00     1      0  0  0.1612263513
#> 751 158 4.706964e+00     1      0  0  0.1612263513
#> 752 158 5.025905e+00     1      0  0  0.1612263513
#> 753 158 5.075969e+00     0      1  0  0.1612263513
#> 754 159 2.842673e+01     1      0  1 -0.0849829798
#> 755 159 4.447516e+01     1      0  1 -0.0849829798
#> 756 159 6.000000e+01     0      0  1 -0.0849829798
#> 757 160 6.833293e+00     0      0  0 -0.5549840322
#> 758 161 1.153776e+01     1      0  1 -0.0961632011
#> 759 161 1.369631e+01     1      0  1 -0.0961632011
#> 760 161 2.163545e+01     0      0  1 -0.0961632011
#> 761 162 6.303535e-02     1      0  1  1.7437363669
#> 762 162 7.707309e-02     1      0  1  1.7437363669
#> 763 162 6.319862e-01     1      0  1  1.7437363669
#> 764 162 6.622051e-01     1      0  1  1.7437363669
#> 765 162 1.980643e+00     1      0  1  1.7437363669
#> 766 162 3.281574e+00     1      0  1  1.7437363669
#> 767 162 4.944079e+00     1      0  1  1.7437363669
#> 768 162 5.181966e+00     1      0  1  1.7437363669
#> 769 162 5.845704e+00     1      0  1  1.7437363669
#> 770 162 7.299639e+00     1      0  1  1.7437363669
#> 771 162 7.791284e+00     1      0  1  1.7437363669
#> 772 162 8.773226e+00     1      0  1  1.7437363669
#> 773 162 1.438082e+01     1      0  1  1.7437363669
#> 774 162 1.489150e+01     1      0  1  1.7437363669
#> 775 162 1.624150e+01     1      0  1  1.7437363669
#> 776 162 1.719144e+01     1      0  1  1.7437363669
#> 777 162 4.367518e+01     0      0  1  1.7437363669
#> 778 163 1.806116e-01     1      0  1  0.3521081511
#> 779 163 7.141378e-01     1      0  1  0.3521081511
#> 780 163 8.093781e-01     1      0  1  0.3521081511
#> 781 163 4.542210e+00     1      0  1  0.3521081511
#> 782 163 7.099276e+00     1      0  1  0.3521081511
#> 783 163 1.730469e+01     1      0  1  0.3521081511
#> 784 163 3.819506e+01     0      0  1  0.3521081511
#> 785 164 6.733214e-01     1      0  1  0.1359775708
#> 786 164 2.227533e+00     1      0  1  0.1359775708
#> 787 164 4.561455e+01     1      0  1  0.1359775708
#> 788 164 6.000000e+01     0      0  1  0.1359775708
#> 789 165 1.340900e-02     1      0  0  1.4824027185
#> 790 165 3.088836e-01     1      0  0  1.4824027185
#> 791 165 3.287180e-01     1      0  0  1.4824027185
#> 792 165 5.148759e-01     1      0  0  1.4824027185
#> 793 165 6.999758e-01     1      0  0  1.4824027185
#> 794 165 8.296338e-01     1      0  0  1.4824027185
#> 795 165 1.008754e+00     1      0  0  1.4824027185
#> 796 165 1.116452e+00     1      0  0  1.4824027185
#> 797 165 1.296550e+00     1      0  0  1.4824027185
#> 798 165 1.778911e+00     1      0  0  1.4824027185
#> 799 165 1.917121e+00     1      0  0  1.4824027185
#> 800 165 2.546599e+00     1      0  0  1.4824027185
#> 801 165 2.859462e+00     1      0  0  1.4824027185
#> 802 165 3.310215e+00     1      0  0  1.4824027185
#> 803 165 3.474073e+00     1      0  0  1.4824027185
#> 804 165 3.711884e+00     1      0  0  1.4824027185
#> 805 165 3.788946e+00     1      0  0  1.4824027185
#> 806 165 3.866990e+00     1      0  0  1.4824027185
#> 807 165 4.369119e+00     1      0  0  1.4824027185
#> 808 165 4.375014e+00     1      0  0  1.4824027185
#> 809 165 4.378417e+00     1      0  0  1.4824027185
#> 810 165 4.437843e+00     1      0  0  1.4824027185
#> 811 165 4.515220e+00     1      0  0  1.4824027185
#> 812 165 4.896496e+00     1      0  0  1.4824027185
#> 813 165 6.347453e+00     1      0  0  1.4824027185
#> 814 165 6.707559e+00     1      0  0  1.4824027185
#> 815 165 6.823431e+00     1      0  0  1.4824027185
#> 816 165 8.996319e+00     0      0  0  1.4824027185
#> 817 166 3.781556e-01     1      0  1 -0.0728028900
#> 818 166 8.732918e+00     0      0  1 -0.0728028900
#> 819 167 4.885724e+00     1      0  0 -0.7479299709
#> 820 167 1.414387e+01     0      0  0 -0.7479299709
#> 821 168 2.263374e+00     1      0  0 -1.2775005910
#> 822 168 1.976591e+01     1      0  0 -1.2775005910
#> 823 168 2.784838e+01     0      0  0 -1.2775005910
#> 824 169 8.864930e-01     1      0  0 -0.0106887373
#> 825 169 1.480715e+01     1      0  0 -0.0106887373
#> 826 169 1.532711e+01     1      0  0 -0.0106887373
#> 827 169 2.075570e+01     0      0  0 -0.0106887373
#> 828 170 1.933924e+01     1      0  0 -0.6927239746
#> 829 170 3.636487e+01     1      0  0 -0.6927239746
#> 830 170 6.000000e+01     0      0  0 -0.6927239746
#> 831 171 9.260389e-02     1      0  1  0.7507981317
#> 832 171 5.846941e+00     1      0  1  0.7507981317
#> 833 171 1.056470e+01     0      0  1  0.7507981317
#> 834 172 5.996133e-01     1      0  0 -1.6712278879
#> 835 172 2.434210e+01     1      0  0 -1.6712278879
#> 836 172 5.323172e+01     0      0  0 -1.6712278879
#> 837 173 3.349436e-01     1      0  1 -0.7261049649
#> 838 173 1.711127e+00     1      0  1 -0.7261049649
#> 839 173 1.957583e+01     1      0  1 -0.7261049649
#> 840 173 3.478920e+01     0      0  1 -0.7261049649
#> 841 174 3.204319e+00     1      0  0 -0.6360976232
#> 842 174 5.502285e+00     1      0  0 -0.6360976232
#> 843 174 1.473118e+01     0      0  0 -0.6360976232
#> 844 175 2.152671e+00     1      0  1 -0.0906465712
#> 845 175 1.259035e+01     0      1  1 -0.0906465712
#> 846 176 1.615849e-01     1      0  0  0.5919460034
#> 847 176 2.516567e-01     1      0  0  0.5919460034
#> 848 176 4.568253e-01     1      0  0  0.5919460034
#> 849 176 1.408276e+00     1      0  0  0.5919460034
#> 850 176 2.430297e+00     1      0  0  0.5919460034
#> 851 176 2.481689e+00     1      0  0  0.5919460034
#> 852 176 2.538656e+00     1      0  0  0.5919460034
#> 853 176 2.931798e+00     1      0  0  0.5919460034
#> 854 176 3.307498e+00     1      0  0  0.5919460034
#> 855 176 3.308567e+00     1      0  0  0.5919460034
#> 856 176 3.385184e+00     1      0  0  0.5919460034
#> 857 176 4.126238e+00     1      0  0  0.5919460034
#> 858 176 4.989141e+00     1      0  0  0.5919460034
#> 859 176 8.365287e+00     1      0  0  0.5919460034
#> 860 176 9.634717e+00     1      0  0  0.5919460034
#> 861 176 1.127658e+01     1      0  0  0.5919460034
#> 862 176 1.241894e+01     1      0  0  0.5919460034
#> 863 176 1.346027e+01     1      0  0  0.5919460034
#> 864 176 1.718392e+01     1      0  0  0.5919460034
#> 865 176 2.753963e+01     1      0  0  0.5919460034
#> 866 176 2.784913e+01     1      0  0  0.5919460034
#> 867 176 3.180499e+01     0      0  0  0.5919460034
#> 868 177 1.368977e-01     1      0  0  0.1588231071
#> 869 177 5.327127e-01     1      0  0  0.1588231071
#> 870 177 1.278711e+00     1      0  0  0.1588231071
#> 871 177 5.501809e+00     1      0  0  0.1588231071
#> 872 177 7.564734e+00     1      0  0  0.1588231071
#> 873 177 1.193894e+01     1      0  0  0.1588231071
#> 874 177 1.655103e+01     1      0  0  0.1588231071
#> 875 177 5.775250e+01     0      0  0  0.1588231071
#> 876 178 7.329994e-02     1      0  0 -0.3543692199
#> 877 178 1.773361e-01     1      0  0 -0.3543692199
#> 878 178 5.320070e-01     1      0  0 -0.3543692199
#> 879 178 2.683472e+00     1      0  0 -0.3543692199
#> 880 178 2.932998e+00     1      0  0 -0.3543692199
#> 881 178 5.003381e+00     1      0  0 -0.3543692199
#> 882 178 6.198142e+00     1      0  0 -0.3543692199
#> 883 178 1.459690e+01     0      0  0 -0.3543692199
#> 884 179 1.752920e-01     1      0  0  1.0031694144
#> 885 179 5.891855e-01     1      0  0  1.0031694144
#> 886 179 6.147432e-01     1      0  0  1.0031694144
#> 887 179 6.988512e-01     1      0  0  1.0031694144
#> 888 179 1.396022e+00     1      0  0  1.0031694144
#> 889 179 1.545018e+00     1      0  0  1.0031694144
#> 890 179 1.556267e+00     1      0  0  1.0031694144
#> 891 179 1.793104e+00     1      0  0  1.0031694144
#> 892 179 2.379402e+00     1      0  0  1.0031694144
#> 893 179 4.207547e+00     1      0  0  1.0031694144
#> 894 179 4.687190e+00     1      0  0  1.0031694144
#> 895 179 5.468205e+00     0      1  0  1.0031694144
#> 896 180 1.045758e+00     1      0  1 -0.0003846487
#> 897 180 1.101407e+00     1      0  1 -0.0003846487
#> 898 180 6.359207e+00     0      0  1 -0.0003846487
#> 899 181 1.175924e+00     1      0  1  0.4387217506
#> 900 181 6.000000e+01     0      0  1  0.4387217506
#> 901 182 9.417778e-03     0      1  0  1.9022097714
#> 902 183 3.135609e+00     1      0  1  0.5510743694
#> 903 183 9.886934e+00     1      0  1  0.5510743694
#> 904 183 1.353237e+01     1      0  1  0.5510743694
#> 905 183 1.886052e+01     0      0  1  0.5510743694
#> 906 184 3.683474e+00     1      0  1 -0.1125595515
#> 907 184 1.847678e+01     1      0  1 -0.1125595515
#> 908 184 2.517254e+01     1      0  1 -0.1125595515
#> 909 184 5.157826e+01     1      0  1 -0.1125595515
#> 910 184 6.000000e+01     0      0  1 -0.1125595515
#> 911 185 5.859502e+00     1      0  1 -0.8056980816
#> 912 185 1.318815e+01     1      0  1 -0.8056980816
#> 913 185 1.564889e+01     1      0  1 -0.8056980816
#> 914 185 3.579556e+01     1      0  1 -0.8056980816
#> 915 185 6.000000e+01     0      0  1 -0.8056980816
#> 916 186 3.154891e-02     1      0  0  0.9030760860
#> 917 186 1.000131e-01     1      0  0  0.9030760860
#> 918 186 1.301208e-01     1      0  0  0.9030760860
#> 919 186 1.469023e-01     1      0  0  0.9030760860
#> 920 186 2.794915e-01     1      0  0  0.9030760860
#> 921 186 8.144476e-01     1      0  0  0.9030760860
#> 922 186 1.811783e+00     1      0  0  0.9030760860
#> 923 186 1.985377e+00     1      0  0  0.9030760860
#> 924 186 2.779249e+00     1      0  0  0.9030760860
#> 925 186 3.070822e+00     1      0  0  0.9030760860
#> 926 186 3.176689e+00     1      0  0  0.9030760860
#> 927 186 6.039410e+00     1      0  0  0.9030760860
#> 928 186 7.142305e+00     0      0  0  0.9030760860
#> 929 187 1.720681e+00     0      0  0 -1.2716818640
#> 930 188 2.809487e+00     1      0  1 -0.1498524302
#> 931 188 5.382822e+00     1      0  1 -0.1498524302
#> 932 188 1.919451e+01     0      0  1 -0.1498524302
#> 933 189 2.890565e-01     1      0  1 -0.2205419122
#> 934 189 4.904706e+01     0      0  1 -0.2205419122
#> 935 190 2.140436e+00     0      1  1 -0.1488163614
#> 936 191 1.299334e+00     1      0  0 -1.0736429908
#> 937 191 1.572703e+00     1      0  0 -1.0736429908
#> 938 191 3.284927e+00     1      0  0 -1.0736429908
#> 939 191 1.214235e+01     0      0  0 -1.0736429908
#> 940 192 1.138347e+00     1      0  0  0.3953447148
#> 941 192 1.552092e+00     1      0  0  0.3953447148
#> 942 192 1.741615e+00     1      0  0  0.3953447148
#> 943 192 2.871842e+00     1      0  0  0.3953447148
#> 944 192 2.884463e+00     0      0  0  0.3953447148
#> 945 193 6.306994e-01     1      0  1 -0.2872301871
#> 946 193 1.440694e+00     0      1  1 -0.2872301871
#> 947 194 9.959967e-01     1      0  1  1.1758906762
#> 948 194 1.177982e+00     1      0  1  1.1758906762
#> 949 194 1.299085e+00     1      0  1  1.1758906762
#> 950 194 1.706327e+00     1      0  1  1.1758906762
#> 951 194 2.522137e+00     1      0  1  1.1758906762
#> 952 194 4.245035e+00     1      0  1  1.1758906762
#> 953 194 4.300865e+00     1      0  1  1.1758906762
#> 954 194 5.491457e+00     1      0  1  1.1758906762
#> 955 194 6.304380e+00     1      0  1  1.1758906762
#> 956 194 7.734315e+00     1      0  1  1.1758906762
#> 957 194 9.329484e+00     1      0  1  1.1758906762
#> 958 194 1.171964e+01     1      0  1  1.1758906762
#> 959 194 1.635179e+01     1      0  1  1.1758906762
#> 960 194 2.324748e+01     1      0  1  1.1758906762
#> 961 194 4.794506e+01     0      1  1  1.1758906762
#> 962 195 2.348195e+00     1      0  0 -0.6093516696
#> 963 195 7.120367e+00     1      0  0 -0.6093516696
#> 964 195 1.176223e+01     1      0  0 -0.6093516696
#> 965 195 2.050131e+01     1      0  0 -0.6093516696
#> 966 195 6.000000e+01     0      0  0 -0.6093516696
#> 967 196 1.885757e-03     1      0  0 -0.4357368460
#> 968 196 2.511524e-01     1      0  0 -0.4357368460
#> 969 196 8.815269e+00     1      0  0 -0.4357368460
#> 970 196 1.606646e+01     0      0  0 -0.4357368460
#> 971 197 6.716962e-02     1      0  1 -0.4539137334
#> 972 197 8.896661e-01     1      0  1 -0.4539137334
#> 973 197 1.254780e+00     1      0  1 -0.4539137334
#> 974 197 6.000000e+01     0      0  1 -0.4539137334
#> 975 198 1.255663e+01     0      0  0 -1.9925857712
#> 976 199 2.054697e+00     1      0  1 -0.5726975274
#> 977 199 6.818948e+00     1      0  1 -0.5726975274
#> 978 199 1.550632e+01     0      0  1 -0.5726975274
#> 979 200 1.569684e-01     1      0  0  0.3314015164
#> 980 200 1.310937e+00     1      0  0  0.3314015164
#> 981 200 1.479520e+00     1      0  0  0.3314015164
#> 982 200 2.003779e+00     1      0  0  0.3314015164
#> 983 200 4.055331e+00     1      0  0  0.3314015164
#> 984 200 4.993162e+00     1      0  0  0.3314015164
#> 985 200 5.516588e+00     0      1  0  0.3314015164