Function to generate simulated recurrent event data

The function simSC() generates simulated recurrent event data from either a Cox-type model, an accelerated mean model, an accelerated rate model, or a scale-change model.

simSC(
  n,
  a1 = a2,
  b1 = b2,
  a2 = a1,
  b2 = b1,
  type = "cox",
  zVar = 0.25,
  tau = 60,
  summary = FALSE
)

Arguments

n
number of observation.

a1, a2, b1, b2
are numeric vectors of length 2. These correspond to the \(\alpha\), \(\beta\), \(\eta\), and \(\theta\) in the joint model. See Details

type
is a character string specifying the underlying model. The rate function type and the hazard function type are separated by a vertical bar "|", with the rate function on the left. For example, type = "cox|am" generates the recurrent process from a Cox model and the terminal event from an accelerated mean model. Setting type = "cox" gives type = "cox|cox".

zVar
a numeric variable specifying the variance of the fraility variable,\(Z\), when zVar > 0. When zVar = 0, \(Z\) is set to a fixed constant 1. The default value is 0.25.

tau
a numeric value specifying the maximum observation time.

summary
a logical value indicating whether a brief data summary will be printed.

Details

The function simSC() generates simulated recurrent event data over the interval \((0, \tau)\) based on the specification of the recurrent process and the terminal events. Specifically, the rate function, \(\lambda(t)\), of the recurrent process can be specified as one of the following model: $$\lambda(t) = Z \lambda_0(te^{X^\top\alpha}) e^{X^\top\beta}, h(t) = Z h_0(te^{X^\top\eta})e^{X^\top\theta}, $$ where \(\lambda_0(t)\) is the baseline rate function, \(h_0(t)\) is the baseline hazard function, \(X\) is a \(n\) by \(p\) covariate matrix and \(\alpha\), \(Z\) is an unobserved shared frailty variable, and \((\alpha, \eta)\) and \((\beta, \theta)\) correspond to the shape and size parameters of the rate function and the hazard function, respectively.

For all scenarios, two covariates are considered; \(X = (X_1, X_2)\), where \(X_1\) follows a Bernoulli distribution with probability 0.5 and \(X_2\) follows a standard normal distribution. The censoring time could be either independent (given covariates) or informative. The simulated data is used for illustration. An informative censoring time, \(C\), is generated separately from an exponential distribution with a rate parameter of 1 / 60 if \(X_1\) is 1, or \(Z^2 / 30\) if \(X_1\) is 0. The observed recurrent events is then observed up to the minimum of \(C\), terminal event, and \(\tau\). Lastly, we assume the baseline functions $$\lambda_0(t) = \frac{2}{1 + t}, h_0(t) = \frac{1}{8(1 + t)}.$$

See also

Examples

set.seed(123)
simSC(200, c(-1, 1), c(-1, 1), summary = TRUE)

#> 
#> Summary results for number of recurrent event per subject:
#>    Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
#>   0.000   1.000   2.000   3.685   4.000  26.000 
#> 
#> Number of failures: 45 (22.5%); Number of censored events: 155 (77.5%)
#> 
#> Number of x1 == 1: 108 (54%); Number of x1 == 0: 92 (46%)
#> Summary results for x2:
#>     Min.  1st Qu.   Median     Mean  3rd Qu.     Max. 
#> -3.03778 -0.82739 -0.05204 -0.13558  0.56892  2.43800 
#> 
#> 

#>      id         Time event status x1          x2
#> 1     1  0.857778670     1      0  0  0.61088135
#> 2     1  3.774670324     1      0  0  0.61088135
#> 3     1  4.327448943     1      0  0  0.61088135
#> 4     1  6.169270476     1      0  0  0.61088135
#> 5     1 12.133817448     1      0  0  0.61088135
#> 6     1 16.527029561     1      0  0  0.61088135
#> 7     1 27.614026787     1      0  0  0.61088135
#> 8     1 43.364851214     1      0  0  0.61088135
#> 9     1 49.233148662     1      0  0  0.61088135
#> 10    1 57.467185223     1      0  0  0.61088135
#> 11    1 58.860254847     0      1  0  0.61088135
#> 12    2  1.292431730     1      0  1 -0.29379674
#> 13    2 19.203900537     1      0  1 -0.29379674
#> 14    2 53.378296995     0      0  1 -0.29379674
#> 15    3 60.000000000     0      0  1 -2.67971620
#> 16    4  0.203013495     1      0  1 -1.23784442
#> 17    4  8.228250481     0      0  1 -1.23784442
#> 18    5  0.029834196     1      0  0  0.57656861
#> 19    5  0.125954039     1      0  0  0.57656861
#> 20    5  0.192170017     1      0  0  0.57656861
#> 21    5  0.389791304     1      0  0  0.57656861
#> 22    5  0.543827322     1      0  0  0.57656861
#> 23    5  0.562997955     1      0  0  0.57656861
#> 24    5  0.569149089     1      0  0  0.57656861
#> 25    5  0.678095820     1      0  0  0.57656861
#> 26    5  0.820505206     1      0  0  0.57656861
#> 27    5  0.918427968     1      0  0  0.57656861
#> 28    5  1.019574052     0      1  0  0.57656861
#> 29    6  3.292726537     1      0  1  1.13589059
#> 30    6  5.254577281     1      0  1  1.13589059
#> 31    6  5.864360838     1      0  1  1.13589059
#> 32    6  8.623298052     0      1  1  1.13589059
#> 33    7  2.477542393     1      0  1  0.69414179
#> 34    7  7.341796193     1      0  1  0.69414179
#> 35    7 26.233409420     0      0  1  0.69414179
#> 36    8  2.627249321     1      0  1  0.28657117
#> 37    8 56.833125760     0      0  1  0.28657117
#> 38    9  1.751201165     1      0  1 -1.99806731
#> 39    9  6.745961242     1      0  1 -1.99806731
#> 40    9 34.114559831     1      0  1 -1.99806731
#> 41    9 60.000000000     0      0  1 -1.99806731
#> 42   10  0.059449295     1      0  0  1.14977824
#> 43   10  0.547488403     1      0  0  1.14977824
#> 44   10  0.583980252     1      0  0  1.14977824
#> 45   10  0.939034798     1      0  0  1.14977824
#> 46   10  1.041999502     1      0  0  1.14977824
#> 47   10  1.311531946     1      0  0  1.14977824
#> 48   10  1.543186889     1      0  0  1.14977824
#> 49   10  2.211708713     1      0  0  1.14977824
#> 50   10  2.937924041     1      0  0  1.14977824
#> 51   10  3.925624140     1      0  0  1.14977824
#> 52   10  4.107674399     1      0  0  1.14977824
#> 53   10  4.904792228     0      1  0  1.14977824
#> 54   11  0.006501609     0      1  1  1.43586055
#> 55   12  2.437431206     1      0  0 -0.47944007
#> 56   12  9.301189913     1      0  0 -0.47944007
#> 57   12 22.295212540     0      0  0 -0.47944007
#> 58   13  0.045346455     1      0  1  2.28910634
#> 59   13  0.064254103     1      0  1  2.28910634
#> 60   13  0.105698514     1      0  1  2.28910634
#> 61   13  0.133929350     1      0  1  2.28910634
#> 62   13  0.492101404     1      0  1  2.28910634
#> 63   13  0.718084494     1      0  1  2.28910634
#> 64   13  0.986977755     0      1  1  2.28910634
#> 65   14  1.617342909     0      0  0 -3.03778215
#> 66   15  4.997897336     1      0  0 -1.04352595
#> 67   15  7.812530888     0      0  0 -1.04352595
#> 68   16  0.088727958     1      0  0  0.24650280
#> 69   16  0.342698710     1      0  0  0.24650280
#> 70   16  0.525838146     1      0  0  0.24650280
#> 71   16  0.695334508     1      0  0  0.24650280
#> 72   16  7.293572274     1      0  0  0.24650280
#> 73   16  7.806145817     1      0  0  0.24650280
#> 74   16  8.543268443     1      0  0  0.24650280
#> 75   16  9.742039665     1      0  0  0.24650280
#> 76   16 11.017322711     1      0  0  0.24650280
#> 77   16 21.006263222     0      0  0  0.24650280
#> 78   17  4.056153045     1      0  0  0.07006490
#> 79   17  5.565292424     1      0  0  0.07006490
#> 80   17  9.411522341     1      0  0  0.07006490
#> 81   17 29.794500405     1      0  0  0.07006490
#> 82   17 44.748298888     1      0  0  0.07006490
#> 83   17 60.000000000     0      0  0  0.07006490
#> 84   18 50.711298340     0      0  1 -1.48378791
#> 85   19  0.021730666     1      0  1  1.63172709
#> 86   19  2.779347705     1      0  1  1.63172709
#> 87   19  2.950843503     1      0  1  1.63172709
#> 88   19  3.705315318     1      0  1  1.63172709
#> 89   19  3.801999260     0      0  1  1.63172709
#> 90   20 43.903887959     0      0  1 -0.67635833
#> 91   21 60.000000000     0      0  0 -1.46375788
#> 92   22  5.602562533     0      1  0 -0.02547748
#> 93   23  1.322154197     1      0  1  1.21791351
#> 94   23  2.766497844     1      0  1  1.21791351
#> 95   23  2.925827028     1      0  1  1.21791351
#> 96   23  7.787536542     1      0  1  1.21791351
#> 97   23 60.000000000     0      0  1  1.21791351
#> 98   24  3.186522780     1      0  1 -0.82832296
#> 99   24  9.032419609     0      0  1 -0.82832296
#> 100  25 11.497533129     1      0  1 -0.71812572
#> 101  25 15.080197854     0      0  1 -0.71812572
#> 102  26  0.721519139     1      0  1 -0.04448748
#> 103  26  4.589523623     1      0  1 -0.04448748
#> 104  26 15.408576457     0      0  1 -0.04448748
#> 105  27  4.144047657     1      0  0 -1.06184814
#> 106  27  8.237677137     1      0  0 -1.06184814
#> 107  27 17.820114993     0      0  0 -1.06184814
#> 108  28  0.623072190     1      0  0 -0.22167586
#> 109  28  0.722096191     1      0  0 -0.22167586
#> 110  28  0.882344625     0      1  0 -0.22167586
#> 111  29  0.026758183     1      0  0  0.39459155
#> 112  29  0.463588893     1      0  0  0.39459155
#> 113  29  1.037152584     1      0  0  0.39459155
#> 114  29  1.652867072     1      0  0  0.39459155
#> 115  29  5.744603197     1      0  0  0.39459155
#> 116  29  6.053072484     1      0  0  0.39459155
#> 117  29  6.753206535     1      0  0  0.39459155
#> 118  29  8.155072682     1      0  0  0.39459155
#> 119  29  8.981397901     1      0  0  0.39459155
#> 120  29  9.367254817     0      0  0  0.39459155
#> 121  30  0.424169855     0      1  1  0.88942683
#> 122  31  0.076093205     1      0  1  0.86871727
#> 123  31  0.678840805     1      0  1  0.86871727
#> 124  31  1.044594670     1      0  1  0.86871727
#> 125  31  3.383167083     0      1  1  0.86871727
#> 126  32  8.733573735     0      0  1 -0.67204182
#> 127  33  1.018768475     1      0  0 -0.23457505
#> 128  33  2.884046729     1      0  0 -0.23457505
#> 129  33 13.647418549     1      0  0 -0.23457505
#> 130  33 58.930902943     0      0  0 -0.23457505
#> 131  34  0.055916366     1      0  0  0.87279300
#> 132  34  0.482519592     1      0  0  0.87279300
#> 133  34  0.823737130     1      0  0  0.87279300
#> 134  34  1.754885582     1      0  0  0.87279300
#> 135  34  1.775730130     1      0  0  0.87279300
#> 136  34  2.360129851     1      0  0  0.87279300
#> 137  34  8.207246832     0      1  0  0.87279300
#> 138  35  0.750084150     1      0  1  0.91333624
#> 139  35  0.877402040     1      0  1  0.91333624
#> 140  35  1.028376601     1      0  1  0.91333624
#> 141  35  1.126395805     1      0  1  0.91333624
#> 142  35  2.039189773     1      0  1  0.91333624
#> 143  35 23.395831970     1      0  1  0.91333624
#> 144  35 24.076442068     1      0  1  0.91333624
#> 145  35 25.494863203     1      0  1  0.91333624
#> 146  35 41.980178315     1      0  1  0.91333624
#> 147  35 46.525022573     0      0  1  0.91333624
#> 148  36  0.372065079     1      0  1  0.39722919
#> 149  36  0.910421455     1      0  1  0.39722919
#> 150  36  1.803997605     1      0  1  0.39722919
#> 151  36 60.000000000     0      0  1  0.39722919
#> 152  37  0.122234543     1      0  0 -0.09698557
#> 153  37  0.462524151     1      0  0 -0.09698557
#> 154  37  1.220826910     1      0  0 -0.09698557
#> 155  37  1.907086479     1      0  0 -0.09698557
#> 156  37  2.133492344     0      1  0 -0.09698557
#> 157  38 50.458659453     0      0  1 -2.07988119
#> 158  39  0.662714538     1      0  0  1.37395794
#> 159  39  0.865363005     1      0  0  1.37395794
#> 160  39  0.897472585     1      0  0  1.37395794
#> 161  39  1.348142421     1      0  0  1.37395794
#> 162  39  1.466669216     1      0  0  1.37395794
#> 163  39  1.688056432     1      0  0  1.37395794
#> 164  39  1.912142428     1      0  0  1.37395794
#> 165  39  2.068251807     1      0  0  1.37395794
#> 166  39  3.484059615     1      0  0  1.37395794
#> 167  39  3.964562913     1      0  0  1.37395794
#> 168  39  5.465293236     1      0  0  1.37395794
#> 169  39 11.723955104     1      0  0  1.37395794
#> 170  39 13.773293167     1      0  0  1.37395794
#> 171  39 15.603346380     1      0  0  1.37395794
#> 172  39 17.320611345     1      0  0  1.37395794
#> 173  39 17.567790871     1      0  0  1.37395794
#> 174  39 18.168564545     1      0  0  1.37395794
#> 175  39 22.314936320     1      0  0  1.37395794
#> 176  39 23.514457149     1      0  0  1.37395794
#> 177  39 25.582078244     1      0  0  1.37395794
#> 178  39 25.758783329     1      0  0  1.37395794
#> 179  39 31.433398747     1      0  0  1.37395794
#> 180  39 31.600443794     1      0  0  1.37395794
#> 181  39 33.601202006     1      0  0  1.37395794
#> 182  39 36.070269649     1      0  0  1.37395794
#> 183  39 52.311367685     1      0  0  1.37395794
#> 184  39 60.000000000     0      0  0  1.37395794
#> 185  40  0.209464103     1      0  1  0.10681370
#> 186  40  5.005879697     1      0  1  0.10681370
#> 187  40 16.710968410     0      0  1  0.10681370
#> 188  41  9.010725044     0      0  1 -0.02650367
#> 189  42  0.633091349     1      0  1 -0.23364408
#> 190  42 26.406054944     0      0  1 -0.23364408
#> 191  43  0.669196899     1      0  0 -0.58116337
#> 192  43 13.820443594     1      0  0 -0.58116337
#> 193  43 60.000000000     0      0  0 -0.58116337
#> 194  44 42.314792516     1      0  1 -0.64567527
#> 195  44 60.000000000     0      0  1 -0.64567527
#> 196  45 16.187568605     0      0  1 -0.34287707
#> 197  46 39.509692247     1      0  1 -1.67472926
#> 198  46 60.000000000     0      0  1 -1.67472926
#> 199  47  0.508822618     1      0  1  1.86118774
#> 200  47  0.785414691     1      0  1  1.86118774
#> 201  47  2.986486967     1      0  1  1.86118774
#> 202  47  3.336696186     1      0  1  1.86118774
#> 203  47  3.967099545     1      0  1  1.86118774
#> 204  47  4.155605802     1      0  1  1.86118774
#> 205  47  5.817204124     1      0  1  1.86118774
#> 206  47  6.239395305     1      0  1  1.86118774
#> 207  47  6.405652186     1      0  1  1.86118774
#> 208  47  8.964143119     1      0  1  1.86118774
#> 209  47 12.658167167     1      0  1  1.86118774
#> 210  47 13.652186243     1      0  1  1.86118774
#> 211  47 14.019072557     1      0  1  1.86118774
#> 212  47 18.283106512     1      0  1  1.86118774
#> 213  47 18.401779653     1      0  1  1.86118774
#> 214  47 20.041072085     1      0  1  1.86118774
#> 215  47 29.308965407     1      0  1  1.86118774
#> 216  47 36.826698580     1      0  1  1.86118774
#> 217  47 60.000000000     0      0  1  1.86118774
#> 218  48  2.651058747     1      0  0 -0.05562058
#> 219  48 24.973344154     0      0  0 -0.05562058
#> 220  49  0.391673648     1      0  1  0.43340640
#> 221  49  1.305599638     1      0  1  0.43340640
#> 222  49  1.571065005     1      0  1  0.43340640
#> 223  49  2.042018997     1      0  1  0.43340640
#> 224  49  3.561040845     1      0  1  0.43340640
#> 225  49  3.772636113     1      0  1  0.43340640
#> 226  49  5.614342374     0      0  1  0.43340640
#> 227  50  6.627962890     1      0  1  0.51037393
#> 228  50 51.259687189     1      0  1  0.51037393
#> 229  50 60.000000000     0      0  1  0.51037393
#> 230  51  0.012832971     1      0  0  1.19590720
#> 231  51  0.040519626     1      0  0  1.19590720
#> 232  51  0.045365435     1      0  0  1.19590720
#> 233  51  0.326395260     1      0  0  1.19590720
#> 234  51  0.497872921     1      0  0  1.19590720
#> 235  51  0.646915424     1      0  0  1.19590720
#> 236  51  1.716521557     1      0  0  1.19590720
#> 237  51  2.146907974     1      0  0  1.19590720
#> 238  51  2.482238233     1      0  0  1.19590720
#> 239  51  4.473740345     1      0  0  1.19590720
#> 240  51  4.967166206     1      0  0  1.19590720
#> 241  51  4.972499710     1      0  0  1.19590720
#> 242  51  5.636198870     1      0  0  1.19590720
#> 243  51  6.350014167     1      0  0  1.19590720
#> 244  51  6.905588557     1      0  0  1.19590720
#> 245  51  7.831753215     1      0  0  1.19590720
#> 246  51  8.497920314     1      0  0  1.19590720
#> 247  51  8.992268446     1      0  0  1.19590720
#> 248  51 11.967402284     1      0  0  1.19590720
#> 249  51 12.634804429     1      0  0  1.19590720
#> 250  51 13.771126712     1      0  0  1.19590720
#> 251  51 14.605601953     1      0  0  1.19590720
#> 252  51 15.032994840     1      0  0  1.19590720
#> 253  51 16.547819697     0      0  0  1.19590720
#> 254  52 23.149639992     0      0  1 -1.08390419
#> 255  53 11.312756566     0      0  1 -1.14510520
#> 256  54  0.449216137     1      0  1  0.15521887
#> 257  54  1.374130959     1      0  1  0.15521887
#> 258  54 48.490525548     0      1  1  0.15521887
#> 259  55  5.620709259     1      0  1 -0.49225142
#> 260  55  7.161720621     1      0  1 -0.49225142
#> 261  55 60.000000000     0      0  1 -0.49225142
#> 262  56  0.083293849     1      0  1  2.36703244
#> 263  56  0.110240848     1      0  1  2.36703244
#> 264  56  0.301668502     1      0  1  2.36703244
#> 265  56  0.660366407     1      0  1  2.36703244
#> 266  56  0.774815788     1      0  1  2.36703244
#> 267  56  1.456125240     1      0  1  2.36703244
#> 268  56  2.719655799     1      0  1  2.36703244
#> 269  56  5.112455368     1      0  1  2.36703244
#> 270  56  5.451597119     1      0  1  2.36703244
#> 271  56  5.613034531     1      0  1  2.36703244
#> 272  56  6.890465800     1      0  1  2.36703244
#> 273  56  8.131767202     1      0  1  2.36703244
#> 274  56  8.610855414     1      0  1  2.36703244
#> 275  56  9.919361807     1      0  1  2.36703244
#> 276  56 20.108454679     1      0  1  2.36703244
#> 277  56 20.170903656     1      0  1  2.36703244
#> 278  56 20.699420630     1      0  1  2.36703244
#> 279  56 22.113506596     1      0  1  2.36703244
#> 280  56 26.996635586     1      0  1  2.36703244
#> 281  56 32.071189257     1      0  1  2.36703244
#> 282  56 38.189061456     0      1  1  2.36703244
#> 283  57  0.242580025     1      0  0 -1.43350260
#> 284  57  3.220353172     1      0  0 -1.43350260
#> 285  57  8.974133998     0      0  0 -1.43350260
#> 286  58 60.000000000     0      0  1 -1.61148587
#> 287  59  0.211234472     1      0  0  0.62481008
#> 288  59  0.256734885     1      0  0  0.62481008
#> 289  59  0.325251603     1      0  0  0.62481008
#> 290  59  0.680650699     1      0  0  0.62481008
#> 291  59  2.493782854     1      0  0  0.62481008
#> 292  59  2.563196136     1      0  0  0.62481008
#> 293  59  2.864625574     1      0  0  0.62481008
#> 294  59  6.179414896     0      1  0  0.62481008
#> 295  60  2.490837916     1      0  1  0.74515000
#> 296  60  2.776032094     1      0  1  0.74515000
#> 297  60  9.695689752     1      0  1  0.74515000
#> 298  60 16.344725095     1      0  1  0.74515000
#> 299  60 26.171017075     0      0  1  0.74515000
#> 300  61  1.212013263     1      0  0  0.58533975
#> 301  61  1.649091398     1      0  0  0.58533975
#> 302  61  1.753614074     1      0  0  0.58533975
#> 303  61  2.274701422     0      1  0  0.58533975
#> 304  62  0.060585423     1      0  0  0.18165619
#> 305  62  5.897016388     1      0  0  0.18165619
#> 306  62 21.026219933     1      0  0  0.18165619
#> 307  62 57.178126519     0      0  0  0.18165619
#> 308  63  0.185954443     1      0  1  0.40489066
#> 309  63  0.847426513     1      0  1  0.40489066
#> 310  63  3.266645943     1      0  1  0.40489066
#> 311  63  4.479127287     1      0  1  0.40489066
#> 312  63  5.062594777     0      0  1  0.40489066
#> 313  64 25.579035674     1      0  0 -1.29744269
#> 314  64 32.662459061     1      0  0 -1.29744269
#> 315  64 35.989559390     0      0  0 -1.29744269
#> 316  65  0.257991095     1      0  1  0.63208408
#> 317  65  3.284318945     0      0  1  0.63208408
#> 318  66 60.000000000     0      0  1 -1.17534240
#> 319  67  0.284517034     1      0  0 -0.63487099
#> 320  67  0.567994347     1      0  0 -0.63487099
#> 321  67  1.797944635     1      0  0 -0.63487099
#> 322  67 16.858378738     0      0  0 -0.63487099
#> 323  68  0.873167356     1      0  0 -2.35827900
#> 324  68  6.668874674     1      0  0 -2.35827900
#> 325  68  8.408135022     1      0  0 -2.35827900
#> 326  68 27.600437614     0      0  0 -2.35827900
#> 327  69  3.599070780     1      0  1 -1.12618034
#> 328  69 14.376114055     0      0  1 -1.12618034
#> 329  70  1.414016722     1      0  0  0.60195782
#> 330  70  2.168187152     1      0  0  0.60195782
#> 331  70  2.299536978     1      0  0  0.60195782
#> 332  70 13.053730518     1      0  0  0.60195782
#> 333  70 29.638220732     1      0  0  0.60195782
#> 334  70 32.943597580     1      0  0  0.60195782
#> 335  70 38.952059879     1      0  0  0.60195782
#> 336  70 48.187798572     1      0  0  0.60195782
#> 337  70 60.000000000     0      0  0  0.60195782
#> 338  71  1.166499446     1      0  0  0.56844485
#> 339  71  1.608533528     1      0  0  0.56844485
#> 340  71  1.846143308     1      0  0  0.56844485
#> 341  71  3.580836993     1      0  0  0.56844485
#> 342  71  3.739749407     1      0  0  0.56844485
#> 343  71  8.125214957     1      0  0  0.56844485
#> 344  71  8.483448197     1      0  0  0.56844485
#> 345  71 12.266580176     1      0  0  0.56844485
#> 346  71 16.821353419     1      0  0  0.56844485
#> 347  71 20.511378335     1      0  0  0.56844485
#> 348  71 26.322450843     1      0  0  0.56844485
#> 349  71 30.231328149     1      0  0  0.56844485
#> 350  71 37.494336247     1      0  0  0.56844485
#> 351  71 60.000000000     0      0  0  0.56844485
#> 352  72  1.806932359     1      0  1 -0.16099753
#> 353  72 50.990556011     0      0  1 -0.16099753
#> 354  73  2.374488559     1      0  0 -1.17408003
#> 355  73 15.533826678     1      0  0 -1.17408003
#> 356  73 27.281362966     1      0  0 -1.17408003
#> 357  73 58.577306242     1      0  0 -1.17408003
#> 358  73 60.000000000     0      0  0 -1.17408003
#> 359  74  0.035241974     1      0  0  1.22134077
#> 360  74  0.277631290     1      0  0  1.22134077
#> 361  74  0.280638456     1      0  0  1.22134077
#> 362  74  0.282376201     1      0  0  1.22134077
#> 363  74  0.312925428     1      0  0  1.22134077
#> 364  74  0.353278589     1      0  0  1.22134077
#> 365  74  0.561664451     1      0  0  1.22134077
#> 366  74  1.502029437     1      0  0  1.22134077
#> 367  74  1.772136343     1      0  0  1.22134077
#> 368  74  1.862194766     1      0  0  1.22134077
#> 369  74  7.457635019     1      0  0  1.22134077
#> 370  74  9.079061045     1      0  0  1.22134077
#> 371  74 12.948617387     1      0  0  1.22134077
#> 372  74 22.058334637     1      0  0  1.22134077
#> 373  74 22.570905665     1      0  0  1.22134077
#> 374  74 24.298738879     1      0  0  1.22134077
#> 375  74 27.038203217     1      0  0  1.22134077
#> 376  74 43.978652972     1      0  0  1.22134077
#> 377  74 55.573735335     1      0  0  1.22134077
#> 378  74 60.000000000     0      0  0  1.22134077
#> 379  75  3.287947111     0      1  1 -1.72392510
#> 380  76 10.644753193     0      0  1 -0.07874138
#> 381  77 11.084130240     1      0  0 -0.61034866
#> 382  77 11.539131801     1      0  0 -0.61034866
#> 383  77 14.055011299     0      0  0 -0.61034866
#> 384  78  0.414963773     1      0  0 -1.20420970
#> 385  78  4.917347520     1      0  0 -1.20420970
#> 386  78 47.671175592     0      0  0 -1.20420970
#> 387  79  3.607116689     1      0  1  0.12069145
#> 388  79  5.001520295     1      0  1  0.12069145
#> 389  79 14.702123039     1      0  1  0.12069145
#> 390  79 28.375934084     1      0  1  0.12069145
#> 391  79 60.000000000     0      0  1  0.12069145
#> 392  80  1.355725067     1      0  0 -0.95507829
#> 393  80  8.105133394     0      0  0 -0.95507829
#> 394  81  0.095297687     1      0  1  0.06872299
#> 395  81 14.988664426     0      0  1  0.06872299
#> 396  82  0.071016162     1      0  0  0.44975095
#> 397  82  0.438055234     1      0  0  0.44975095
#> 398  82  0.655236641     1      0  0  0.44975095
#> 399  82  0.935172399     1      0  0  0.44975095
#> 400  82  2.654678649     1      0  0  0.44975095
#> 401  82  3.155226451     1      0  0  0.44975095
#> 402  82  4.190179375     1      0  0  0.44975095
#> 403  82  6.069942234     1      0  0  0.44975095
#> 404  82  6.935114287     1      0  0  0.44975095
#> 405  82  7.239392342     0      1  0  0.44975095
#> 406  83  1.506113422     1      0  1  0.50493827
#> 407  83  2.615395214     1      0  1  0.50493827
#> 408  83  8.485583207     0      0  1  0.50493827
#> 409  84  0.059412511     1      0  0  0.33675757
#> 410  84  2.503239682     1      0  0  0.33675757
#> 411  84  4.735862393     0      1  0  0.33675757
#> 412  85  0.163939282     1      0  0  1.41227304
#> 413  85  0.675428570     1      0  0  1.41227304
#> 414  85  1.492815361     1      0  0  1.41227304
#> 415  85  1.542087961     1      0  0  1.41227304
#> 416  85  1.852216103     1      0  0  1.41227304
#> 417  85  2.120160098     1      0  0  1.41227304
#> 418  85  2.475878138     1      0  0  1.41227304
#> 419  85  3.150640666     1      0  0  1.41227304
#> 420  85  3.490949230     1      0  0  1.41227304
#> 421  85  3.542412027     1      0  0  1.41227304
#> 422  85  3.671130454     1      0  0  1.41227304
#> 423  85  4.059856922     1      0  0  1.41227304
#> 424  85  4.297577115     1      0  0  1.41227304
#> 425  85  4.589939731     1      0  0  1.41227304
#> 426  85  5.069331940     1      0  0  1.41227304
#> 427  85  5.167796672     1      0  0  1.41227304
#> 428  85  5.512678385     1      0  0  1.41227304
#> 429  85  5.661733707     1      0  0  1.41227304
#> 430  85  6.884105895     1      0  0  1.41227304
#> 431  85  7.297883220     1      0  0  1.41227304
#> 432  85  8.221132778     1      0  0  1.41227304
#> 433  85  8.281005880     0      1  0  1.41227304
#> 434  86 27.250376483     0      0  1 -1.04248381
#> 435  87  0.040741584     1      0  1  1.57697064
#> 436  87  0.378640382     1      0  1  1.57697064
#> 437  87  1.349134369     1      0  1  1.57697064
#> 438  87  3.561167456     1      0  1  1.57697064
#> 439  87  5.711905191     1      0  1  1.57697064
#> 440  87  8.710514032     1      0  1  1.57697064
#> 441  87  9.977268827     0      1  1  1.57697064
#> 442  88  1.286529042     0      0  1 -0.53096049
#> 443  89  3.184708514     1      0  1 -0.76249835
#> 444  89  8.127139891     0      0  1 -0.76249835
#> 445  90  2.151966845     1      0  1  0.69556748
#> 446  90  2.307815398     1      0  1  0.69556748
#> 447  90  2.486715669     1      0  1  0.69556748
#> 448  90  3.907999121     1      0  1  0.69556748
#> 449  90  5.599846684     1      0  1  0.69556748
#> 450  90  5.605159358     1      0  1  0.69556748
#> 451  90  5.994026961     1      0  1  0.69556748
#> 452  90 10.609225664     1      0  1  0.69556748
#> 453  90 12.810452902     0      1  1  0.69556748
#> 454  91  0.884257093     1      0  0  0.11688093
#> 455  91  1.256166988     1      0  0  0.11688093
#> 456  91  1.765221947     1      0  0  0.11688093
#> 457  91  2.136806975     1      0  0  0.11688093
#> 458  91  2.487261310     1      0  0  0.11688093
#> 459  91  3.825054809     1      0  0  0.11688093
#> 460  91  8.139356672     1      0  0  0.11688093
#> 461  91  8.280126397     1      0  0  0.11688093
#> 462  91 14.719939247     1      0  0  0.11688093
#> 463  91 19.317287247     1      0  0  0.11688093
#> 464  91 20.347616049     1      0  0  0.11688093
#> 465  91 28.515711132     1      0  0  0.11688093
#> 466  91 35.795360279     0      0  0  0.11688093
#> 467  92  0.272153152     1      0  0  0.70362490
#> 468  92  0.495108798     1      0  0  0.70362490
#> 469  92  1.645236354     0      0  0  0.70362490
#> 470  93  0.974576914     1      0  0 -0.30310821
#> 471  93  6.043230095     0      0  0 -0.30310821
#> 472  94  0.262868024     1      0  0  0.83415416
#> 473  94  0.791065315     1      0  0  0.83415416
#> 474  94  1.318705857     1      0  0  0.83415416
#> 475  94  3.998189071     0      1  0  0.83415416
#> 476  95  6.759619384     1      0  1  0.10698201
#> 477  95 20.225458888     1      0  1  0.10698201
#> 478  95 32.682648711     1      0  1  0.10698201
#> 479  95 50.766468272     1      0  1  0.10698201
#> 480  95 60.000000000     0      0  1  0.10698201
#> 481  96  0.030915864     1      0  0  0.21101666
#> 482  96  0.072796994     1      0  0  0.21101666
#> 483  96  0.201548801     1      0  0  0.21101666
#> 484  96  0.545687966     0      1  0  0.21101666
#> 485  97  0.118872139     1      0  1  0.02836530
#> 486  97  1.309532179     1      0  1  0.02836530
#> 487  97  2.152092406     1      0  1  0.02836530
#> 488  97 60.000000000     0      0  1  0.02836530
#> 489  98  0.285728785     1      0  0  1.30095042
#> 490  98  0.406051313     1      0  0  1.30095042
#> 491  98  0.651778350     1      0  0  1.30095042
#> 492  98  0.853405921     1      0  0  1.30095042
#> 493  98  1.298280926     1      0  0  1.30095042
#> 494  98  1.324576639     1      0  0  1.30095042
#> 495  98  1.410284893     1      0  0  1.30095042
#> 496  98  2.080029634     1      0  0  1.30095042
#> 497  98  2.215417259     1      0  0  1.30095042
#> 498  98  2.225544634     1      0  0  1.30095042
#> 499  98  2.435916774     1      0  0  1.30095042
#> 500  98  2.935963102     1      0  0  1.30095042
#> 501  98  4.357480345     1      0  0  1.30095042
#> 502  98  4.704871702     1      0  0  1.30095042
#> 503  98  6.187409827     1      0  0  1.30095042
#> 504  98  6.351807631     1      0  0  1.30095042
#> 505  98  6.741181246     0      0  0  1.30095042
#> 506  99  1.321167075     0      0  0 -1.46520978
#> 507 100  0.071558238     1      0  1 -1.62389961
#> 508 100 60.000000000     0      0  1 -1.62389961
#> 509 101  2.073361660     1      0  0  0.19229688
#> 510 101  2.460340338     1      0  0  0.19229688
#> 511 101  3.110774726     1      0  0  0.19229688
#> 512 101  5.747207641     1      0  0  0.19229688
#> 513 101  6.832136286     1      0  0  0.19229688
#> 514 101  9.680620485     1      0  0  0.19229688
#> 515 101  9.710535911     0      0  0  0.19229688
#> 516 102  1.169658626     1      0  1  1.74452014
#> 517 102  1.344649261     1      0  1  1.74452014
#> 518 102  1.516016716     1      0  1  1.74452014
#> 519 102  5.569064805     1      0  1  1.74452014
#> 520 102  7.515052690     1      0  1  1.74452014
#> 521 102  7.543793055     1      0  1  1.74452014
#> 522 102 16.254402914     1      0  1  1.74452014
#> 523 102 19.648460828     1      0  1  1.74452014
#> 524 102 60.000000000     0      0  1  1.74452014
#> 525 103  0.541924010     1      0  1  0.07881304
#> 526 103  6.547581553     0      0  1  0.07881304
#> 527 104 28.518264871     0      0  1 -0.46894848
#> 528 105  3.392447291     0      0  0 -1.72883299
#> 529 106  0.120599344     1      0  1  1.06515440
#> 530 106  0.702290358     1      0  1  1.06515440
#> 531 106  0.747201220     1      0  1  1.06515440
#> 532 106  1.178738526     1      0  1  1.06515440
#> 533 106  1.938080685     0      1  1  1.06515440
#> 534 107  0.319143128     1      0  1  0.48028177
#> 535 107  1.674601535     1      0  1  0.48028177
#> 536 107  6.658918847     1      0  1  0.48028177
#> 537 107 12.336170029     0      1  1  0.48028177
#> 538 108  2.464958031     1      0  1  0.13749904
#> 539 108  9.910660223     1      0  1  0.13749904
#> 540 108 12.839959672     1      0  1  0.13749904
#> 541 108 23.264338274     1      0  1  0.13749904
#> 542 108 60.000000000     0      0  1  0.13749904
#> 543 109  0.077178547     1      0  1  0.78895745
#> 544 109  0.249446212     1      0  1  0.78895745
#> 545 109  0.651435064     1      0  1  0.78895745
#> 546 109  0.859170290     1      0  1  0.78895745
#> 547 109  2.585783444     1      0  1  0.78895745
#> 548 109  3.511339431     1      0  1  0.78895745
#> 549 109 24.113315443     1      0  1  0.78895745
#> 550 109 25.237356487     0      0  1  0.78895745
#> 551 110  0.646327161     1      0  1 -0.82707992
#> 552 110  7.502017049     0      0  1 -0.82707992
#> 553 111  0.651894003     1      0  1  0.24394415
#> 554 111  3.666034888     1      0  1  0.24394415
#> 555 111  6.218465023     1      0  1  0.24394415
#> 556 111  8.159845805     1      0  1  0.24394415
#> 557 111 47.804593907     0      0  1  0.24394415
#> 558 112 28.267008439     0      0  1 -0.99986303
#> 559 113  0.511057023     1      0  0  1.14141501
#> 560 113  0.598820782     1      0  0  1.14141501
#> 561 113  0.996678433     1      0  0  1.14141501
#> 562 113  1.141537753     1      0  0  1.14141501
#> 563 113  1.193984685     1      0  0  1.14141501
#> 564 113  2.588235431     1      0  0  1.14141501
#> 565 113  4.229750258     0      0  0  1.14141501
#> 566 114 21.230643491     0      0  0 -1.98291175
#> 567 115  1.568335659     0      0  0 -0.02450675
#> 568 116  3.253721864     1      0  0 -0.20815247
#> 569 116  3.457166735     1      0  0 -0.20815247
#> 570 116  5.539496294     1      0  0 -0.20815247
#> 571 116 33.159543848     0      0  0 -0.20815247
#> 572 117 19.237129783     0      0  0 -0.98704941
#> 573 118  3.642133251     1      0  0 -0.40751621
#> 574 118 10.662049983     1      0  0 -0.40751621
#> 575 118 60.000000000     0      0  0 -0.40751621
#> 576 119  0.831266415     1      0  1 -0.51278041
#> 577 119 12.614939219     0      0  1 -0.51278041
#> 578 120  0.246535405     1      0  0 -1.01156035
#> 579 120  1.964749035     1      0  0 -1.01156035
#> 580 120  2.893244983     1      0  0 -1.01156035
#> 581 120  3.568926009     0      0  0 -1.01156035
#> 582 121 28.013826599     0      0  1 -1.70640206
#> 583 122  0.109521420     1      0  0  1.78187690
#> 584 122 17.201220292     1      0  0  1.78187690
#> 585 122 39.241485038     1      0  0  1.78187690
#> 586 122 60.000000000     0      0  0  1.78187690
#> 587 123  0.107823993     1      0  0 -1.54426057
#> 588 123  3.709650984     0      1  0 -1.54426057
#> 589 124  0.821737318     1      0  0 -0.18680915
#> 590 124 11.929129910     1      0  0 -0.18680915
#> 591 124 20.800199078     1      0  0 -0.18680915
#> 592 124 60.000000000     0      0  0 -0.18680915
#> 593 125  1.084465842     0      1  1 -0.12667223
#> 594 126  0.304036555     1      0  1 -1.08041850
#> 595 126  3.353547880     1      0  1 -1.08041850
#> 596 126  6.613656823     0      0  1 -1.08041850
#> 597 127  0.239226846     1      0  1  0.17495696
#> 598 127  1.031695073     1      0  1  0.17495696
#> 599 127  6.791173603     1      0  1  0.17495696
#> 600 127  8.094038397     1      0  1  0.17495696
#> 601 127 27.252576459     1      0  1  0.17495696
#> 602 127 60.000000000     0      0  1  0.17495696
#> 603 128 59.288576310     1      0  1 -1.19099574
#> 604 128 60.000000000     0      0  1 -1.19099574
#> 605 129  1.784023146     1      0  1 -0.05424634
#> 606 129  3.188672230     1      0  1 -0.05424634
#> 607 129  4.475529160     1      0  1 -0.05424634
#> 608 129  9.529705009     1      0  1 -0.05424634
#> 609 129  9.858531397     1      0  1 -0.05424634
#> 610 129 13.219223724     1      0  1 -0.05424634
#> 611 129 22.153043551     0      0  1 -0.05424634
#> 612 130 12.469123021     1      0  0 -1.68466565
#> 613 130 21.197664892     1      0  0 -1.68466565
#> 614 130 60.000000000     0      0  0 -1.68466565
#> 615 131 11.196787395     1      0  1 -0.61460307
#> 616 131 15.725597701     1      0  1 -0.61460307
#> 617 131 19.343594036     1      0  1 -0.61460307
#> 618 131 35.705992475     1      0  1 -0.61460307
#> 619 131 60.000000000     0      0  1 -0.61460307
#> 620 132  1.109134629     1      0  1  0.36708332
#> 621 132  1.151403650     1      0  1  0.36708332
#> 622 132  2.145073840     1      0  1  0.36708332
#> 623 132  2.923377138     1      0  1  0.36708332
#> 624 132  3.388836749     0      1  1  0.36708332
#> 625 133 32.770945916     1      0  0 -2.45734283
#> 626 133 47.203516156     0      0  0 -2.45734283
#> 627 134  0.539530764     1      0  0 -0.20433534
#> 628 134  1.450758095     1      0  0 -0.20433534
#> 629 134  6.777393941     1      0  0 -0.20433534
#> 630 134  7.677449560     1      0  0 -0.20433534
#> 631 134 39.549003186     0      0  0 -0.20433534
#> 632 135  0.254667280     1      0  1  0.55833620
#> 633 135  1.056227861     1      0  1  0.55833620
#> 634 135  1.953677756     1      0  1  0.55833620
#> 635 135  2.553296239     1      0  1  0.55833620
#> 636 135  3.397848589     1      0  1  0.55833620
#> 637 135  8.490284398     1      0  1  0.55833620
#> 638 135 28.340921776     1      0  1  0.55833620
#> 639 135 33.488264233     0      0  1  0.55833620
#> 640 136  0.309195064     1      0  0  0.32421061
#> 641 136  0.312006283     1      0  0  0.32421061
#> 642 136  0.689646595     1      0  0  0.32421061
#> 643 136 16.632908439     1      0  0  0.32421061
#> 644 136 60.000000000     0      0  0  0.32421061
#> 645 137  0.156661734     1      0  1  0.69773701
#> 646 137  0.759016701     0      1  1  0.69773701
#> 647 138  0.402978876     1      0  0 -1.02257306
#> 648 138  0.484545470     1      0  0 -1.02257306
#> 649 138  1.439513204     1      0  0 -1.02257306
#> 650 138  1.844411311     1      0  0 -1.02257306
#> 651 138 60.000000000     0      0  0 -1.02257306
#> 652 139  9.677674244     1      0  0 -0.16827076
#> 653 139 60.000000000     0      0  0 -0.16827076
#> 654 140 15.649355664     1      0  0 -0.75948914
#> 655 140 41.605392147     1      0  0 -0.75948914
#> 656 140 60.000000000     0      0  0 -0.75948914
#> 657 141  0.555926993     1      0  0 -0.25290924
#> 658 141  1.009890141     1      0  0 -0.25290924
#> 659 141  2.274064992     1      0  0 -0.25290924
#> 660 141  3.677544878     1      0  0 -0.25290924
#> 661 141  4.448212303     1      0  0 -0.25290924
#> 662 141  5.180454177     1      0  0 -0.25290924
#> 663 141  6.237856691     0      0  0 -0.25290924
#> 664 142  0.354541205     1      0  1  0.95267999
#> 665 142  1.739039142     1      0  1  0.95267999
#> 666 142  2.846975759     1      0  1  0.95267999
#> 667 142  4.393508140     1      0  1  0.95267999
#> 668 142  5.952649740     1      0  1  0.95267999
#> 669 142  8.536863799     1      0  1  0.95267999
#> 670 142 41.726223061     1      0  1  0.95267999
#> 671 142 48.771243556     1      0  1  0.95267999
#> 672 142 60.000000000     0      0  1  0.95267999
#> 673 143 42.098486325     0      0  1 -0.27541490
#> 674 144  1.508996598     0      0  1 -1.23280604
#> 675 145  0.726042783     1      0  1  0.23750358
#> 676 145  1.642825421     0      1  1  0.23750358
#> 677 146  0.193626020     1      0  0  1.05125929
#> 678 146  0.306108160     1      0  0  1.05125929
#> 679 146  0.444813197     0      1  0  1.05125929
#> 680 147  0.238552667     1      0  1 -0.05545795
#> 681 147  7.478349176     1      0  1 -0.05545795
#> 682 147 14.081328996     0      0  1 -0.05545795
#> 683 148  0.048955534     1      0  0 -0.53767687
#> 684 148  0.869989561     0      1  0 -0.53767687
#> 685 149  0.039633160     1      0  0 -0.14961768
#> 686 149  3.755629554     1      0  0 -0.14961768
#> 687 149  5.236415128     1      0  0 -0.14961768
#> 688 149 34.436668772     0      0  0 -0.14961768
#> 689 150  2.316456960     1      0  1 -0.40991441
#> 690 150  3.151506862     1      0  1 -0.40991441
#> 691 150  3.306826150     1      0  1 -0.40991441
#> 692 150 13.859182441     1      0  1 -0.40991441
#> 693 150 41.269764975     0      0  1 -0.40991441
#> 694 151  0.212833898     1      0  0  0.01754160
#> 695 151  0.674711255     1      0  0  0.01754160
#> 696 151  0.697098677     1      0  0  0.01754160
#> 697 151  1.268611185     1      0  0  0.01754160
#> 698 151  1.327944951     0      0  0  0.01754160
#> 699 152  0.196034127     1      0  0  0.75104110
#> 700 152  0.344280940     1      0  0  0.75104110
#> 701 152  0.880961503     1      0  0  0.75104110
#> 702 152  0.881127478     1      0  0  0.75104110
#> 703 152  1.109470331     1      0  0  0.75104110
#> 704 152  1.323462443     1      0  0  0.75104110
#> 705 152  2.925673643     1      0  0  0.75104110
#> 706 152  3.847252636     1      0  0  0.75104110
#> 707 152  5.391617822     1      0  0  0.75104110
#> 708 152  7.578930005     1      0  0  0.75104110
#> 709 152  8.819432922     1      0  0  0.75104110
#> 710 152 10.181205470     0      0  0  0.75104110
#> 711 153  0.195933684     1      0  0  0.36237112
#> 712 153  0.314299641     1      0  0  0.36237112
#> 713 153  1.381187753     1      0  0  0.36237112
#> 714 153  2.375506879     1      0  0  0.36237112
#> 715 153  4.341508681     1      0  0  0.36237112
#> 716 153  6.845123883     1      0  0  0.36237112
#> 717 153 16.880940182     1      0  0  0.36237112
#> 718 153 17.349275535     1      0  0  0.36237112
#> 719 153 21.416573408     1      0  0  0.36237112
#> 720 153 43.185629023     0      1  0  0.36237112
#> 721 154  0.121488402     1      0  0 -0.58514887
#> 722 154  1.517538906     1      0  0 -0.58514887
#> 723 154  1.892541796     1      0  0 -0.58514887
#> 724 154  2.189626535     0      0  0 -0.58514887
#> 725 155  0.268991600     1      0  0  0.81595948
#> 726 155  0.350979910     1      0  0  0.81595948
#> 727 155  0.425376443     1      0  0  0.81595948
#> 728 155  1.606587689     1      0  0  0.81595948
#> 729 155  2.732074216     1      0  0  0.81595948
#> 730 155  4.358104837     0      0  0  0.81595948
#> 731 156  2.846730132     1      0  1 -0.26693659
#> 732 156  4.363191705     0      0  1 -0.26693659
#> 733 157 11.694223993     0      0  1 -0.47172066
#> 734 158  0.358897918     1      0  1 -0.11180513
#> 735 158  9.752494726     1      0  1 -0.11180513
#> 736 158 16.190854769     1      0  1 -0.11180513
#> 737 158 19.651344659     1      0  1 -0.11180513
#> 738 158 24.263152443     1      0  1 -0.11180513
#> 739 158 59.132642584     1      0  1 -0.11180513
#> 740 158 60.000000000     0      0  1 -0.11180513
#> 741 159  0.147584366     1      0  0  0.13875268
#> 742 159  0.408255780     1      0  0  0.13875268
#> 743 159  0.928803956     1      0  0  0.13875268
#> 744 159  0.940072276     1      0  0  0.13875268
#> 745 159  1.725472580     1      0  0  0.13875268
#> 746 159  4.062785352     1      0  0  0.13875268
#> 747 159  6.016771040     1      0  0  0.13875268
#> 748 159  6.550504896     1      0  0  0.13875268
#> 749 159  8.422304963     1      0  0  0.13875268
#> 750 159  8.518601447     1      0  0  0.13875268
#> 751 159 10.467016779     1      0  0  0.13875268
#> 752 159 15.689560925     1      0  0  0.13875268
#> 753 159 17.790714616     1      0  0  0.13875268
#> 754 159 20.372793008     1      0  0  0.13875268
#> 755 159 20.953186875     1      0  0  0.13875268
#> 756 159 46.530437551     1      0  0  0.13875268
#> 757 159 59.929846049     0      0  0  0.13875268
#> 758 160  0.210242912     1      0  0  0.10858179
#> 759 160  1.543721965     1      0  0  0.10858179
#> 760 160  3.396463726     1      0  0  0.10858179
#> 761 160  4.595388886     1      0  0  0.10858179
#> 762 160  6.683968167     1      0  0  0.10858179
#> 763 160  8.661763109     1      0  0  0.10858179
#> 764 160  9.045786642     1      0  0  0.10858179
#> 765 160 24.718410833     1      0  0  0.10858179
#> 766 160 29.317959763     1      0  0  0.10858179
#> 767 160 60.000000000     0      0  0  0.10858179
#> 768 161  3.332279894     1      0  0 -1.12503012
#> 769 161 60.000000000     0      0  0 -1.12503012
#> 770 162  1.299944261     1      0  1  0.12059549
#> 771 162  2.480986384     1      0  1  0.12059549
#> 772 162 28.183615395     0      1  1  0.12059549
#> 773 163  0.721612509     0      1  1  0.14981793
#> 774 164  1.825022870     0      1  1 -0.13139834
#> 775 165  0.047067658     1      0  1  0.58161590
#> 776 165  0.883730378     1      0  1  0.58161590
#> 777 165  5.591573432     1      0  1  0.58161590
#> 778 165 10.673636080     1      0  1  0.58161590
#> 779 165 40.437349314     0      0  1  0.58161590
#> 780 166  0.172591704     1      0  0  0.05332687
#> 781 166  1.247368264     1      0  0  0.05332687
#> 782 166  1.255194452     0      1  0  0.05332687
#> 783 167  2.166231031     1      0  1 -1.43347667
#> 784 167 60.000000000     0      0  1 -1.43347667
#> 785 168  0.243840747     1      0  0 -2.22299065
#> 786 168  9.013183647     1      0  0 -2.22299065
#> 787 168 12.141140336     0      0  0 -2.22299065
#> 788 169  0.002590756     1      0  0  2.43800305
#> 789 169  0.021954919     1      0  0  2.43800305
#> 790 169  0.068789997     1      0  0  2.43800305
#> 791 169  0.111666928     1      0  0  2.43800305
#> 792 169  0.171465126     1      0  0  2.43800305
#> 793 169  0.177369541     1      0  0  2.43800305
#> 794 169  0.210686730     1      0  0  2.43800305
#> 795 169  0.265156038     1      0  0  2.43800305
#> 796 169  0.266638226     1      0  0  2.43800305
#> 797 169  0.289091911     1      0  0  2.43800305
#> 798 169  0.307209312     0      1  0  2.43800305
#> 799 170  0.031766573     1      0  0  0.94126511
#> 800 170  0.397948092     1      0  0  0.94126511
#> 801 170  1.371645566     1      0  0  0.94126511
#> 802 170  1.859811521     1      0  0  0.94126511
#> 803 170  2.074694334     1      0  0  0.94126511
#> 804 170  3.659091751     1      0  0  0.94126511
#> 805 170  5.048140986     1      0  0  0.94126511
#> 806 170  5.847760822     1      0  0  0.94126511
#> 807 170  8.198519908     1      0  0  0.94126511
#> 808 170  8.863099030     1      0  0  0.94126511
#> 809 170  8.908127946     1      0  0  0.94126511
#> 810 170 12.520655523     1      0  0  0.94126511
#> 811 170 24.071915876     1      0  0  0.94126511
#> 812 170 33.765383137     1      0  0  0.94126511
#> 813 170 60.000000000     0      0  0  0.94126511
#> 814 171  0.029217305     1      0  1  0.09894930
#> 815 171  0.414809681     1      0  1  0.09894930
#> 816 171  6.380029180     1      0  1  0.09894930
#> 817 171  7.250415657     1      0  1  0.09894930
#> 818 171  9.563514398     0      1  1  0.09894930
#> 819 172  0.282145015     1      0  0 -0.59345039
#> 820 172  3.509861006     1      0  0 -0.59345039
#> 821 172  9.414426324     1      0  0 -0.59345039
#> 822 172 11.079506020     1      0  0 -0.59345039
#> 823 172 18.186848089     1      0  0 -0.59345039
#> 824 172 20.042096445     1      0  0 -0.59345039
#> 825 172 28.422544502     1      0  0 -0.59345039
#> 826 172 54.194750897     0      0  0 -0.59345039
#> 827 173  4.377979287     1      0  1 -0.69529494
#> 828 173 60.000000000     0      0  1 -0.69529494
#> 829 174  0.056950448     1      0  1  0.57035334
#> 830 174  1.199341728     1      0  1  0.57035334
#> 831 174 13.627154203     0      0  1  0.57035334
#> 832 175 26.474660254     1      0  0 -0.94220816
#> 833 175 33.469410016     1      0  0 -0.94220816
#> 834 175 48.466635812     0      0  0 -0.94220816
#> 835 176  3.815174080     1      0  1 -0.33209894
#> 836 176  7.688142203     0      1  1 -0.33209894
#> 837 177  0.024047638     0      1  0  0.56261127
#> 838 178  0.783207475     0      0  1 -1.09908545
#> 839 179  0.172852543     1      0  0  0.15594842
#> 840 179  2.248278479     1      0  0  0.15594842
#> 841 179 33.779233401     1      0  0  0.15594842
#> 842 179 41.733352244     1      0  0  0.15594842
#> 843 179 51.000142741     1      0  0  0.15594842
#> 844 179 60.000000000     0      0  0  0.15594842
#> 845 180  0.097376926     1      0  0  1.03178718
#> 846 180  0.125433400     1      0  0  1.03178718
#> 847 180  0.187424841     1      0  0  1.03178718
#> 848 180  0.264430703     1      0  0  1.03178718
#> 849 180  0.424719035     1      0  0  1.03178718
#> 850 180  0.499129237     1      0  0  1.03178718
#> 851 180  0.706038787     1      0  0  1.03178718
#> 852 180  0.741974768     1      0  0  1.03178718
#> 853 180  1.710444559     1      0  0  1.03178718
#> 854 180  1.735763026     1      0  0  1.03178718
#> 855 180  2.001338969     1      0  0  1.03178718
#> 856 180  2.010446337     1      0  0  1.03178718
#> 857 180  2.179694316     1      0  0  1.03178718
#> 858 180  2.300451909     1      0  0  1.03178718
#> 859 180  2.744831507     1      0  0  1.03178718
#> 860 180  2.760729181     0      1  0  1.03178718
#> 861 181  0.787557315     1      0  1  0.84332403
#> 862 181  7.827926087     0      0  1  0.84332403
#> 863 182  0.785790956     1      0  1  0.32901483
#> 864 182  2.530183630     0      0  1  0.32901483
#> 865 183  2.881117036     1      0  1 -1.88167871
#> 866 183 15.691879346     0      0  1 -1.88167871
#> 867 184 60.000000000     0      0  1 -0.37068878
#> 868 185  4.975998502     1      0  1 -0.94653914
#> 869 185 43.803828529     0      0  1 -0.94653914
#> 870 186  0.560388911     1      0  1 -0.21840941
#> 871 186 58.778774131     1      0  1 -0.21840941
#> 872 186 60.000000000     0      0  1 -0.21840941
#> 873 187 16.437834627     1      0  1 -0.60710625
#> 874 187 51.741707849     0      0  1 -0.60710625
#> 875 188  3.583800731     0      0  1 -1.89063391
#> 876 189  0.281466611     1      0  0 -0.04983625
#> 877 189  1.657729008     1      0  0 -0.04983625
#> 878 189  4.193067713     1      0  0 -0.04983625
#> 879 189  7.407061813     0      0  0 -0.04983625
#> 880 190  0.481628982     1      0  0 -1.41179149
#> 881 190  1.535117529     1      0  0 -1.41179149
#> 882 190  2.673628438     0      0  0 -1.41179149
#> 883 191  0.992956406     1      0  0 -1.12915469
#> 884 191 60.000000000     0      0  0 -1.12915469
#> 885 192 19.402380178     1      0  1  0.69699980
#> 886 192 27.622027444     0      0  1  0.69699980
#> 887 193  1.107200385     1      0  0  0.31746073
#> 888 193 51.834529435     0      0  0  0.31746073
#> 889 194 20.103747220     1      0  0 -0.84193639
#> 890 194 60.000000000     0      0  0 -0.84193639
#> 891 195  0.034313703     1      0  0  0.60606049
#> 892 195  0.669132874     1      0  0  0.60606049
#> 893 195  1.111193331     1      0  0  0.60606049
#> 894 195 15.008750917     1      0  0  0.60606049
#> 895 195 15.288491077     1      0  0  0.60606049
#> 896 195 60.000000000     0      0  0  0.60606049
#> 897 196  0.027799092     1      0  1  0.65646811
#> 898 196  1.062470706     1      0  1  0.65646811
#> 899 196  2.844291847     1      0  1  0.65646811
#> 900 196 31.733281992     1      0  1  0.65646811
#> 901 196 60.000000000     0      0  1  0.65646811
#> 902 197  1.038823388     1      0  1  0.07377617
#> 903 197  5.506098574     1      0  1  0.07377617
#> 904 197 17.289419023     0      0  1  0.07377617
#> 905 198  0.051899192     1      0  0  1.53519748
#> 906 198  0.060919038     1      0  0  1.53519748
#> 907 198  0.570935115     1      0  0  1.53519748
#> 908 198  0.694539546     1      0  0  1.53519748
#> 909 198  0.840288618     1      0  0  1.53519748
#> 910 198  0.849575962     1      0  0  1.53519748
#> 911 198  1.080140568     1      0  0  1.53519748
#> 912 198  1.793407127     1      0  0  1.53519748
#> 913 198  1.822186069     1      0  0  1.53519748
#> 914 198  1.894589403     1      0  0  1.53519748
#> 915 198  1.966847554     1      0  0  1.53519748
#> 916 198  2.071096190     1      0  0  1.53519748
#> 917 198  2.400403800     1      0  0  1.53519748
#> 918 198  2.452260936     0      0  0  1.53519748
#> 919 199  0.987402745     1      0  1  0.41391630
#> 920 199  1.037235010     1      0  1  0.41391630
#> 921 199  2.214459537     1      0  1  0.41391630
#> 922 199  2.748081218     1      0  1  0.41391630
#> 923 199  3.572625458     1      0  1  0.41391630
#> 924 199  8.064460153     1      0  1  0.41391630
#> 925 199 11.224001423     1      0  1  0.41391630
#> 926 199 12.898509423     1      0  1  0.41391630
#> 927 199 13.953172481     1      0  1  0.41391630
#> 928 199 16.861074164     1      0  1  0.41391630
#> 929 199 21.616315809     1      0  1  0.41391630
#> 930 199 25.380277257     1      0  1  0.41391630
#> 931 199 30.037959735     0      1  1  0.41391630
#> 932 200  0.001120397     1      0  0  0.93976749
#> 933 200  0.803802670     1      0  0  0.93976749
#> 934 200  1.112366621     1      0  0  0.93976749
#> 935 200  1.166692342     1      0  0  0.93976749
#> 936 200  1.274243243     1      0  0  0.93976749
#> 937 200  1.338728831     0      1  0  0.93976749